coco dataset names updated
