+++ alonet/detr/data_modules/coco_detection2detr.py
@@ -74,11 +74,11 @@ class CocoDetection2Detr(Data2Detr):
     def setup(self, stage: Optional[str] = None):
         if stage == "fit" or stage is None:
             # Setup train/val loaders
-            self.train_dataset = alodataset.CocoDetectionDataset(
+            self.train_dataset = alodataset.CocoBaseDataset(
                 transform_fn=self.train_transform, sample=self.sample, **self.train_loader_kwargs
             )
             self.sample = self.train_dataset.sample or self.sample  # Update sample if user prompt is given
-            self.val_dataset = alodataset.CocoDetectionDataset(
+            self.val_dataset = alodataset.CocoBaseDataset(
                 transform_fn=self.val_transform, sample=self.sample, **self.val_loader_kwargs
             )
             self.sample = self.val_dataset.sample or self.sample  # Update sample if user prompt is given
