+++ docs/_sources/getting_started/alodataset.ipynb.txt
@@ -75,7 +75,7 @@
     "import alodataset\n",
     "\n",
     "# Coco Detection Dataset\n",
-    "dataset = alodataset.CocoDetectionDataset(sample=True)\n",
+    "dataset = alodataset.CocoBaseDataset(sample=True)\n",
     "frame = dataset.getitem(0)\n",
     "frame.get_view().render()\n",
     "\n",
@@ -223,7 +223,7 @@
     "import aloscene\n",
     "\n",
     "# Coco Detection Dataset\n",
-    "dataset = alodataset.CocoDetectionDataset(sample=True)\n",
+    "dataset = alodataset.CocoBaseDataset(sample=True)\n",
     "\n",
     "for frames in dataset.train_loader(batch_size=2):\n",
     "    frames = aloscene.Frame.batch_list(frames)\n",
@@ -307,7 +307,7 @@
     "    return frame\n",
     "\n",
     "# Coco Detection Dataset\n",
-    "dataset = alodataset.CocoDetectionDataset(sample=True, transform_fn=transform_fn)\n",
+    "dataset = alodataset.CocoBaseDataset(sample=True, transform_fn=transform_fn)\n",
     "\n",
     "for frames in dataset.train_loader(batch_size=2):\n",
     "    frames = aloscene.Frame.batch_list(frames)\n",
