+++ docs/_sources/getting_started/alonet.ipynb.txt
@@ -49,17 +49,17 @@
     "import alodataset\n",
     "\n",
     "# Using sample\n",
-    "train_loader = alodataset.CocoDetectionDataset(sample=True).train_loader(batch_size=2)\n",
+    "train_loader = alodataset.CocoBaseDataset(sample=True).train_loader(batch_size=2)\n",
     "\n",
     "# Using the full dataset\n",
     "# Training loader\n",
-    "train_loader = alodataset.CocoDetectionDataset(\n",
+    "train_loader = alodataset.CocoBaseDataset(\n",
     "    img_folder = \"train2017\",\n",
     "    ann_file = \"annotations/instances_train2017.json\"\n",
     ").train_loader(batch_size=2)\n",
     "\n",
     "# Validation loader\n",
-    "train_loader = alodataset.CocoDetectionDataset(\n",
+    "train_loader = alodataset.CocoBaseDataset(\n",
     "    img_folder = \"val2017\",\n",
     "    ann_file = \"annotations/instances_val2017.json\"\n",
     ").train_loader(batch_size=2)"
@@ -114,11 +114,11 @@
     "    def setup(self, stage: Optional[str] = None) -> None:\n",
     "        if stage == \"fit\" or stage is None:\n",
     "            # Setup train/val loaders\n",
-    "            self.coco_train = alodataset.CocoDetectionDataset(\n",
+    "            self.coco_train = alodataset.CocoBaseDataset(\n",
     "                img_folder = \"train2017\",\n",
     "                ann_file = \"annotations/instances_val2017.json\"\n",
     "            )\n",
-    "            self.coco_val = alodataset.CocoDetectionDataset(\n",
+    "            self.coco_val = alodataset.CocoBaseDataset(\n",
     "                img_folder = \"val2017\",\n",
     "                ann_file = \"annotations/instances_val2017.json\"\n",
     "            )\n",
