+++ docs/_sources/getting_started/aloscene.ipynb.txt
@@ -14,7 +14,7 @@
    "source": [
     "## Hello world\n",
     "\n",
-    "In the following, we simply use the CocoDetectionDataset to display one frame. \n",
+    "In the following, we simply use the CocoBaseDataset to display one frame. \n",
     "\n",
     "- The `get_view` method is specific to **Augmented Tensor**. It makes it easy for someone to display all the avaiable data within a **Frame**\n",
     "- `get_view` return a `View` object that can be used to `render()` the frame using matplotlib or opencv.\n",
@@ -26,9 +26,9 @@
    "cell_type": "code",
    "execution_count": 10,
    "source": [
-    "from alodataset import CocoDetectionDataset\n",
+    "from alodataset import CocoBaseDataset\n",
     "\n",
-    "coco_dataset = CocoDetectionDataset(sample=True)\n",
+    "coco_dataset = CocoBaseDataset(sample=True)\n",
     "frame = coco_dataset.getitem(0)\n",
     "frame.get_view().render()"
    ],
