+++ docs/_sources/getting_started/augmented_tensor.ipynb.txt
@@ -270,7 +270,7 @@
    ],
    "source": [
     "import alodataset\n",
-    "frame = alodataset.CocoDetectionDataset(sample=True).getitem(2)\n",
+    "frame = alodataset.CocoBaseDataset(sample=True).getitem(2)\n",
     "boxes = frame.boxes2d\n",
     "frame.get_view().render()"
    ]
