+++ docs/alonet/detr_connectors.html
@@ -4,71 +4,71 @@
 <html class="writer-html5" lang="en" >
 <head>
   <meta charset="utf-8" />
-  
+
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  
+
   <title>Detr: Data connectors &mdash; Aloception 0.0.1 documentation</title>
-  
 
-  
+
+
   <link rel="stylesheet" href="../_static/css/theme.css" type="text/css" />
   <link rel="stylesheet" href="../_static/pygments.css" type="text/css" />
   <link rel="stylesheet" href="../_static/pygments.css" type="text/css" />
   <link rel="stylesheet" href="../_static/css/theme.css" type="text/css" />
 
-  
-  
 
-  
-  
 
-  
 
-  
+
+
+
+
+
+
   <!--[if lt IE 9]>
     <script src="../_static/js/html5shiv.min.js"></script>
   <![endif]-->
-  
-    
+
+
       <script type="text/javascript" id="documentation_options" data-url_root="../" src="../_static/documentation_options.js"></script>
         <script data-url_root="../" id="documentation_options" src="../_static/documentation_options.js"></script>
         <script src="../_static/jquery.js"></script>
         <script src="../_static/underscore.js"></script>
         <script src="../_static/doctools.js"></script>
         <script crossorigin="anonymous" integrity="sha256-Ae2Vz/4ePdIu6ZyI/5ZGsYnb+m0JlOmKPjt6XZ9JJkA=" src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js"></script>
-    
+
     <script type="text/javascript" src="../_static/js/theme.js"></script>
 
-    
+
     <link rel="index" title="Index" href="../genindex.html" />
     <link rel="search" title="Search" href="../search.html" />
     <link rel="next" title="Deformable DETR" href="deformable.html" />
-    <link rel="prev" title="Training" href="detr_training.html" /> 
+    <link rel="prev" title="Training" href="detr_training.html" />
 </head>
 
 <body class="wy-body-for-nav">
 
-   
+
   <div class="wy-grid-for-nav">
-    
+
     <nav data-toggle="wy-nav-shift" class="wy-nav-side">
       <div class="wy-side-scroll">
         <div class="wy-side-nav-search" >
-          
 
-          
+
+
             <a href="../index.html" class="icon icon-home"> Aloception
-          
 
-          
+
+
           </a>
 
-          
-            
-            
-          
 
-          
+
+
+
+
+
 <div role="search">
   <form id="rtd-search-form" class="wy-form" action="../search.html" method="get">
     <input type="text" name="q" placeholder="Search docs" />
@@ -77,17 +77,17 @@
   </form>
 </div>
 
-          
+
         </div>
 
-        
+
         <div class="wy-menu wy-menu-vertical" data-spy="affix" role="navigation" aria-label="main navigation">
-          
-            
-            
-              
-            
-            
+
+
+
+
+
+
               <p class="caption" role="heading"><span class="caption-text">Geting Started</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="../getting_started/getting_started.html">Getting started</a></li>
@@ -132,29 +132,29 @@
 </li>
 </ul>
 
-            
-          
+
+
         </div>
-        
+
       </div>
     </nav>
 
     <section data-toggle="wy-nav-shift" class="wy-nav-content-wrap">
 
-      
+
       <nav class="wy-nav-top" aria-label="top navigation">
-        
+
           <i data-toggle="wy-nav-top" class="fa fa-bars"></i>
           <a href="../index.html">Aloception</a>
-        
+
       </nav>
 
 
       <div class="wy-nav-content">
-        
+
         <div class="rst-content">
-        
-          
+
+
 
 
 
@@ -175,33 +175,33 @@
 <div role="navigation" aria-label="breadcrumbs navigation">
 
   <ul class="wy-breadcrumbs">
-    
+
       <li><a href="../index.html" class="icon icon-home"></a> &raquo;</li>
-        
+
           <li><a href="alonet.html">Alonet</a> &raquo;</li>
-        
+
           <li><a href="detr.html">Detr</a> &raquo;</li>
-        
+
       <li>Detr: Data connectors</li>
-    
-    
+
+
       <li class="wy-breadcrumbs-aside">
-        
-          
+
+
             <a href="../_sources/alonet/detr_connectors.rst.txt" rel="nofollow"> View page source</a>
-          
-        
+
+
       </li>
-    
+
   </ul>
 
-  
+
   <hr/>
 </div>
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
-            
-  
+
+
 <style>
 /* CSS overrides for sphinx_rtd_theme */
 
@@ -400,7 +400,7 @@ using DDP.</p>
 </div>
 <div class="section" id="cocodetection2detr">
 <h2>CocoDetection2Detr<a class="headerlink" href="#cocodetection2detr" title="Permalink to this headline">¶</a></h2>
-<p>LightningDataModule that make the connection between <a class="reference internal" href="../alodataset/coco_detection_dataset.html#module-alodataset.coco_detection_dataset" title="alodataset.coco_detection_dataset"><code class="xref py py-mod docutils literal notranslate"><span class="pre">CocoDetectionDataset</span></code></a>
+<p>LightningDataModule that make the connection between <a class="reference internal" href="../alodataset/coco_detection_dataset.html#module-alodataset.coco_detection_dataset" title="alodataset.coco_detection_dataset"><code class="xref py py-mod docutils literal notranslate"><span class="pre">CocoBaseDataset</span></code></a>
 and <a class="reference internal" href="detr_training.html#module-alonet.detr.train" title="alonet.detr.train"><code class="xref py py-mod docutils literal notranslate"><span class="pre">LitDetr</span></code></a> modules. See <a class="reference internal" href="#module-alonet.detr.data_modules.data2detr" title="alonet.detr.data_modules.data2detr"><code class="xref py py-mod docutils literal notranslate"><span class="pre">Data2Detr</span></code></a>
 to more information about the methods and configurations.</p>
 <div class="section" id="examples">
@@ -519,7 +519,7 @@ in class declaration. Also, set <code class="xref py py-attr docutils literal no
 
 
            </div>
-           
+
           </div>
           <footer>
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
@@ -535,14 +535,14 @@ in class declaration. Also, set <code class="xref py py-attr docutils literal no
 
     </p>
   </div>
-    
-    
-    
+
+
+
     Built with <a href="https://www.sphinx-doc.org/">Sphinx</a> using a
-    
+
     <a href="https://github.com/readthedocs/sphinx_rtd_theme">theme</a>
-    
-    provided by <a href="https://readthedocs.org">Read the Docs</a>. 
+
+    provided by <a href="https://readthedocs.org">Read the Docs</a>.
 
 </footer>
         </div>
@@ -551,7 +551,7 @@ in class declaration. Also, set <code class="xref py py-attr docutils literal no
     </section>
 
   </div>
-  
+
 
   <script type="text/javascript">
       jQuery(function () {
@@ -559,10 +559,10 @@ in class declaration. Also, set <code class="xref py py-attr docutils literal no
       });
   </script>
 
-  
-  
-    
-   
+
+
+
+
 
 </body>
 </html>
