+++ docs/genindex.html
@@ -4,69 +4,69 @@
 <html class="writer-html5" lang="en" >
 <head>
   <meta charset="utf-8" />
-  
+
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  
+
   <title>Index &mdash; Aloception 0.0.1 documentation</title>
-  
 
-  
+
+
   <link rel="stylesheet" href="_static/css/theme.css" type="text/css" />
   <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
   <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
   <link rel="stylesheet" href="_static/css/theme.css" type="text/css" />
 
-  
-  
 
-  
-  
 
-  
 
-  
+
+
+
+
+
+
   <!--[if lt IE 9]>
     <script src="_static/js/html5shiv.min.js"></script>
   <![endif]-->
-  
-    
+
+
       <script type="text/javascript" id="documentation_options" data-url_root="./" src="_static/documentation_options.js"></script>
         <script data-url_root="./" id="documentation_options" src="_static/documentation_options.js"></script>
         <script src="_static/jquery.js"></script>
         <script src="_static/underscore.js"></script>
         <script src="_static/doctools.js"></script>
         <script crossorigin="anonymous" integrity="sha256-Ae2Vz/4ePdIu6ZyI/5ZGsYnb+m0JlOmKPjt6XZ9JJkA=" src="https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js"></script>
-    
+
     <script type="text/javascript" src="_static/js/theme.js"></script>
 
-    
+
     <link rel="index" title="Index" href="#" />
-    <link rel="search" title="Search" href="search.html" /> 
+    <link rel="search" title="Search" href="search.html" />
 </head>
 
 <body class="wy-body-for-nav">
 
-   
+
   <div class="wy-grid-for-nav">
-    
+
     <nav data-toggle="wy-nav-shift" class="wy-nav-side">
       <div class="wy-side-scroll">
         <div class="wy-side-nav-search" >
-          
 
-          
+
+
             <a href="index.html" class="icon icon-home"> Aloception
-          
 
-          
+
+
           </a>
 
-          
-            
-            
-          
 
-          
+
+
+
+
+
 <div role="search">
   <form id="rtd-search-form" class="wy-form" action="search.html" method="get">
     <input type="text" name="q" placeholder="Search docs" />
@@ -75,17 +75,17 @@
   </form>
 </div>
 
-          
+
         </div>
 
-        
+
         <div class="wy-menu wy-menu-vertical" data-spy="affix" role="navigation" aria-label="main navigation">
-          
-            
-            
-              
-            
-            
+
+
+
+
+
+
               <p class="caption" role="heading"><span class="caption-text">Geting Started</span></p>
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="getting_started/getting_started.html">Getting started</a></li>
@@ -111,29 +111,29 @@
 <li class="toctree-l1"><a class="reference internal" href="alonet/alonet.html">Alonet</a></li>
 </ul>
 
-            
-          
+
+
         </div>
-        
+
       </div>
     </nav>
 
     <section data-toggle="wy-nav-shift" class="wy-nav-content-wrap">
 
-      
+
       <nav class="wy-nav-top" aria-label="top navigation">
-        
+
           <i data-toggle="wy-nav-top" class="fa fa-bars"></i>
           <a href="index.html">Aloception</a>
-        
+
       </nav>
 
 
       <div class="wy-nav-content">
-        
+
         <div class="rst-content">
-        
-          
+
+
 
 
 
@@ -154,26 +154,26 @@
 <div role="navigation" aria-label="breadcrumbs navigation">
 
   <ul class="wy-breadcrumbs">
-    
+
       <li><a href="index.html" class="icon icon-home"></a> &raquo;</li>
-        
+
       <li>Index</li>
-    
-    
+
+
       <li class="wy-breadcrumbs-aside">
-        
-          
-        
+
+
+
       </li>
-    
+
   </ul>
 
-  
+
   <hr/>
 </div>
           <div role="main" class="document" itemscope="itemscope" itemtype="http://schema.org/Article">
            <div itemprop="articleBody">
-            
+
 
 <h1 id="index">Index</h1>
 
@@ -201,7 +201,7 @@
  | <a href="#W"><strong>W</strong></a>
  | <a href="#X"><strong>X</strong></a>
  | <a href="#Y"><strong>Y</strong></a>
- 
+
 </div>
 <h2 id="_">_</h2>
 <table style="width: 100%" class="indextable genindextable"><tr>
@@ -927,7 +927,7 @@
   <td style="width: 33%; vertical-align: top;"><ul>
       <li><a href="alonet/detr_connectors.html#alonet.detr.data_modules.coco_detection2detr.CocoDetection2Detr">CocoDetection2Detr (class in alonet.detr.data_modules.coco_detection2detr)</a>
 </li>
-      <li><a href="alodataset/coco_detection_dataset.html#alodataset.coco_detection_dataset.CocoDetectionDataset">CocoDetectionDataset (class in alodataset.coco_detection_dataset)</a>
+      <li><a href="alodataset/coco_detection_dataset.html#alodataset.coco_detection_dataset.CocoBaseDataset">CocoBaseDataset (class in alodataset.coco_detection_dataset)</a>
 </li>
       <li><a href="alonet/detr_connectors.html#alonet.detr.data_modules.coco_panoptic2detr.CocoPanoptic2Detr">CocoPanoptic2Detr (class in alonet.detr.data_modules.coco_panoptic2detr)</a>
 </li>
@@ -1262,7 +1262,7 @@
       <ul>
         <li><a href="alodataset/chairssdhom_dataset.html#alodataset.chairssdhom_dataset.ChairsSDHomDataset.getitem">(alodataset.chairssdhom_dataset.ChairsSDHomDataset method)</a>
 </li>
-        <li><a href="alodataset/coco_detection_dataset.html#alodataset.coco_detection_dataset.CocoDetectionDataset.getitem">(alodataset.coco_detection_dataset.CocoDetectionDataset method)</a>
+        <li><a href="alodataset/coco_detection_dataset.html#alodataset.coco_detection_dataset.CocoBaseDataset.getitem">(alodataset.coco_detection_dataset.CocoBaseDataset method)</a>
 </li>
         <li><a href="alodataset/crowd_human_dataset.html#alodataset.crowd_human_dataset.CrowdHumanDataset.getitem">(alodataset.crowd_human_dataset.CrowdHumanDataset method)</a>
 </li>
@@ -2120,7 +2120,7 @@
 
 
            </div>
-           
+
           </div>
           <footer>
 
@@ -2132,14 +2132,14 @@
 
     </p>
   </div>
-    
-    
-    
+
+
+
     Built with <a href="https://www.sphinx-doc.org/">Sphinx</a> using a
-    
+
     <a href="https://github.com/readthedocs/sphinx_rtd_theme">theme</a>
-    
-    provided by <a href="https://readthedocs.org">Read the Docs</a>. 
+
+    provided by <a href="https://readthedocs.org">Read the Docs</a>.
 
 </footer>
         </div>
@@ -2148,7 +2148,7 @@
     </section>
 
   </div>
-  
+
 
   <script type="text/javascript">
       jQuery(function () {
@@ -2156,10 +2156,10 @@
       });
   </script>
 
-  
-  
-    
-   
+
+
+
+
 
 </body>
 </html>
