+++ docs/searchindex.js
@@ -1,1 +1,1 @@
-Search.setIndex({docnames:["alodataset/alodataset","alodataset/base_dataset","alodataset/chairssdhom_dataset","alodataset/coco_detection_dataset","alodataset/crowd_human_dataset","alodataset/flying_chairs2_dataset","alodataset/flyingthings3D_subset_dataset","alodataset/merge_dataset","alodataset/mot17","alodataset/sintel_disparity_dataset","alodataset/sintel_flow_dataset","alodataset/sintel_multi_dataset","alodataset/transform","alodataset/waymo_dataset","alonet/alonet","alonet/alonet.callbacks","alonet/alonet.metrics","alonet/alonet.transformers","alonet/deformable","alonet/deformable_architecture","alonet/deformable_models","alonet/deformable_training","alonet/detr","alonet/detr_architecture","alonet/detr_connectors","alonet/detr_models","alonet/detr_training","alonet/panoptic","alonet/panoptic_models","alonet/panoptic_training","aloscene/aloscene","aloscene/augmented_tensor","aloscene/bounding_boxes_2d","aloscene/bounding_boxes_3d","aloscene/camera_calib","aloscene/disparity","aloscene/flow","aloscene/frame","aloscene/labels","aloscene/mask","aloscene/notebooks/bounding_boxes_3d","aloscene/oriented_boxes_2d","aloscene/points2d","aloscene/spatial_augmented_tensor","getting_started/alodataset","getting_started/alonet","getting_started/aloscene","getting_started/augmented_tensor","getting_started/getting_started","index","tutorials/data_setup","tutorials/export_tensorrt/Tutorial - Export DETR to TensorRT and inference","tutorials/export_tensorrt/Tutorial - Export Deformable DETR to TensorRT and inference","tutorials/finetuning_deformable_detr","tutorials/finetuning_detr","tutorials/tensort_inference","tutorials/training_deformable_detr","tutorials/training_detr","tutorials/training_panoptic"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,nbsphinx:3,sphinx:56},filenames:["alodataset/alodataset.rst","alodataset/base_dataset.rst","alodataset/chairssdhom_dataset.rst","alodataset/coco_detection_dataset.rst","alodataset/crowd_human_dataset.rst","alodataset/flying_chairs2_dataset.rst","alodataset/flyingthings3D_subset_dataset.rst","alodataset/merge_dataset.rst","alodataset/mot17.rst","alodataset/sintel_disparity_dataset.rst","alodataset/sintel_flow_dataset.rst","alodataset/sintel_multi_dataset.rst","alodataset/transform.rst","alodataset/waymo_dataset.rst","alonet/alonet.rst","alonet/alonet.callbacks.rst","alonet/alonet.metrics.rst","alonet/alonet.transformers.rst","alonet/deformable.rst","alonet/deformable_architecture.rst","alonet/deformable_models.rst","alonet/deformable_training.rst","alonet/detr.rst","alonet/detr_architecture.rst","alonet/detr_connectors.rst","alonet/detr_models.rst","alonet/detr_training.rst","alonet/panoptic.rst","alonet/panoptic_models.rst","alonet/panoptic_training.rst","aloscene/aloscene.rst","aloscene/augmented_tensor.rst","aloscene/bounding_boxes_2d.rst","aloscene/bounding_boxes_3d.rst","aloscene/camera_calib.rst","aloscene/disparity.rst","aloscene/flow.rst","aloscene/frame.rst","aloscene/labels.rst","aloscene/mask.rst","aloscene/notebooks/bounding_boxes_3d.ipynb","aloscene/oriented_boxes_2d.rst","aloscene/points2d.rst","aloscene/spatial_augmented_tensor.rst","getting_started/alodataset.ipynb","getting_started/alonet.ipynb","getting_started/aloscene.ipynb","getting_started/augmented_tensor.ipynb","getting_started/getting_started.rst","index.rst","tutorials/data_setup.rst","tutorials/export_tensorrt/Tutorial - Export DETR to TensorRT and inference.ipynb","tutorials/export_tensorrt/Tutorial - Export Deformable DETR to TensorRT and inference.ipynb","tutorials/finetuning_deformable_detr.ipynb","tutorials/finetuning_detr.ipynb","tutorials/tensort_inference.rst","tutorials/training_deformable_detr.ipynb","tutorials/training_detr.ipynb","tutorials/training_panoptic.ipynb"],objects:{"alodataset.base_dataset":{BaseDataset:[1,1,1,""],Split:[1,1,1,""],rename_data_to_none:[1,5,1,""],stream_loader:[1,5,1,""],train_loader:[1,5,1,""]},"alodataset.base_dataset.BaseDataset":{download_sample:[1,2,1,""],get:[1,2,1,""],get_dataset_dir:[1,2,1,""],getitem:[1,2,1,""],getitem_ignore_errors:[1,2,1,""],prepare:[1,2,1,""],set_dataset_dir:[1,2,1,""],stream_loader:[1,2,1,""],train_loader:[1,2,1,""],vb_folder:[1,3,1,""]},"alodataset.base_dataset.Split":{TEST:[1,4,1,""],TRAIN:[1,4,1,""],Type:[1,4,1,""],VAL:[1,4,1,""]},"alodataset.chairssdhom_dataset":{ChairsSDHomDataset:[2,1,1,""]},"alodataset.chairssdhom_dataset.ChairsSDHomDataset":{SPLIT_FOLDERS:[2,4,1,""],get_file_ids:[2,2,1,""],get_frames:[2,2,1,""],get_sequences:[2,2,1,""],getitem:[2,2,1,""]},"alodataset.coco_detection_dataset":{CocoDetectionDataset:[3,1,1,""]},"alodataset.coco_detection_dataset.CocoDetectionDataset":{getitem:[3,2,1,""]},"alodataset.crowd_human_dataset":{CrowdHumanDataset:[4,1,1,""],main:[4,5,1,""]},"alodataset.crowd_human_dataset.CrowdHumanDataset":{getitem:[4,2,1,""],load_gt:[4,2,1,""],load_json_lines:[4,2,1,""],prepare:[4,2,1,""]},"alodataset.flying_chairs2_dataset":{FlyingChairs2Dataset:[5,1,1,""]},"alodataset.flying_chairs2_dataset.FlyingChairs2Dataset":{SPLIT_FOLDERS:[5,4,1,""],getitem:[5,2,1,""]},"alodataset.flyingthings3D_subset_dataset":{FlyingThings3DSubsetDataset:[6,1,1,""]},"alodataset.flyingthings3D_subset_dataset.FlyingThings3DSubsetDataset":{CAMERAS:[6,4,1,""],LABELS:[6,4,1,""],get_featlabels_element:[6,2,1,""],get_features_labels:[6,2,1,""],get_frames:[6,2,1,""],getitem:[6,2,1,""],id_sf_path:[6,2,1,""]},"alodataset.merge_dataset":{MergeDataset:[7,1,1,""]},"alodataset.merge_dataset.MergeDataset":{stream_loader:[7,2,1,""],train_loader:[7,2,1,""]},"alodataset.mot17":{Mot17:[8,1,1,""],main:[8,5,1,""]},"alodataset.mot17.Mot17":{DPM_SEQUENCES:[8,4,1,""],FRCNN_SEQUENCES:[8,4,1,""],SDP_SEQUENCES:[8,4,1,""],SEQUENCES:[8,4,1,""],SPLIT_FOLDERS:[8,4,1,""],getitem:[8,2,1,""]},"alodataset.sintel_disparity_dataset":{SintelDisparityDataset:[9,1,1,""]},"alodataset.sintel_disparity_dataset.SintelDisparityDataset":{CAMERAS:[9,4,1,""],LABELS:[9,4,1,""],PASSES:[9,4,1,""]},"alodataset.sintel_flow_dataset":{SintelFlowDataset:[10,1,1,""]},"alodataset.sintel_flow_dataset.SintelFlowDataset":{CAMERAS:[10,4,1,""],LABELS:[10,4,1,""],PASSES:[10,4,1,""]},"alodataset.sintel_multi_dataset":{SintelMultiDataset:[11,1,1,""]},"alodataset.sintel_multi_dataset.SintelMultiDataset":{CAMERAS:[11,4,1,""],DISP_LABELS:[11,4,1,""],FLOW_LABELS:[11,4,1,""],LABELS:[11,4,1,""],PASSES:[11,4,1,""],SINTEL_SEQUENCES:[11,4,1,""],get_dataset_dir:[11,2,1,""],getitem:[11,2,1,""],set_dataset_dir:[11,2,1,""]},"alodataset.transforms":{AloTransform:[12,1,1,""],ColorJitter:[12,1,1,""],Compose:[12,1,1,""],CustomRandomColoring:[12,1,1,""],RandomCrop:[12,1,1,""],RandomHorizontalFlip:[12,1,1,""],RandomResizeWithAspectRatio:[12,1,1,""],RandomSelect:[12,1,1,""],RandomSizeCrop:[12,1,1,""],RealisticNoise:[12,1,1,""],Resize:[12,1,1,""],SpatialShift:[12,1,1,""]},"alodataset.transforms.AloTransform":{sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.ColorJitter":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""],training:[12,4,1,""]},"alodataset.transforms.Compose":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.CustomRandomColoring":{apply:[12,2,1,""],sample_params:[12,2,1,""]},"alodataset.transforms.RandomCrop":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomHorizontalFlip":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomResizeWithAspectRatio":{apply:[12,2,1,""],get_size_with_aspect_ratio:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomSelect":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomSizeCrop":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RealisticNoise":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.Resize":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.SpatialShift":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.waymo_dataset":{WaymoDataset:[13,1,1,""],main:[13,5,1,""]},"alodataset.waymo_dataset.WaymoDataset":{CAMERAS:[13,4,1,""],CLASSES:[13,4,1,""],LABELS:[13,4,1,""],SPLIT_FOLDERS:[13,4,1,""],__init__:[13,2,1,""],get_frame_boxes2d:[13,2,1,""],get_frame_boxes3d:[13,2,1,""],get_frame_camera_parameters:[13,2,1,""],get_frames:[13,2,1,""],get_segments:[13,2,1,""],getitem:[13,2,1,""],np_convert_waymo_to_aloception_coordinate_system:[13,2,1,""],prepare:[13,2,1,""],read_sequences:[13,2,1,""]},"alonet.callbacks":{base_metrics_callback:[15,0,0,"-"],map_metrics_callback:[15,0,0,"-"],metrics_callback:[15,0,0,"-"],object_detector_callback:[15,0,0,"-"],pq_metrics_callback:[15,0,0,"-"]},"alonet.callbacks.base_metrics_callback":{InstancesBaseMetricsCallback:[15,1,1,""]},"alonet.callbacks.base_metrics_callback.InstancesBaseMetricsCallback":{add_sample:[15,2,1,""],inference:[15,2,1,""],on_validation_batch_end:[15,2,1,""],on_validation_end:[15,2,1,""]},"alonet.callbacks.map_metrics_callback":{ApMetricsCallback:[15,1,1,""]},"alonet.callbacks.map_metrics_callback.ApMetricsCallback":{on_validation_end:[15,2,1,""]},"alonet.callbacks.metrics_callback":{MetricsCallback:[15,1,1,""],get_rank:[15,5,1,""],is_dist_avail_and_initialized:[15,5,1,""]},"alonet.callbacks.metrics_callback.MetricsCallback":{on_train_batch_end:[15,2,1,""],on_train_epoch_end:[15,2,1,""],on_validation_batch_end:[15,2,1,""],on_validation_epoch_end:[15,2,1,""]},"alonet.callbacks.object_detector_callback":{ObjectDetectorCallback:[15,1,1,""]},"alonet.callbacks.object_detector_callback.ObjectDetectorCallback":{log_boxes_2d:[15,2,1,""],log_boxes_3d:[15,2,1,""],log_masks:[15,2,1,""],on_train_batch_end:[15,2,1,""],on_validation_epoch_end:[15,2,1,""]},"alonet.callbacks.pq_metrics_callback":{PQMetricsCallback:[15,1,1,""]},"alonet.callbacks.pq_metrics_callback.PQMetricsCallback":{on_validation_end:[15,2,1,""]},"alonet.deformable_detr":{backbone:[19,0,0,"-"],callbacks:[21,0,0,"-"],criterion:[21,0,0,"-"],deformable_detr:[20,0,0,"-"],deformable_detr_r50:[20,0,0,"-"],deformable_detr_r50_finetune:[20,0,0,"-"],deformable_detr_r50_refinement:[20,0,0,"-"],deformable_transformer:[19,0,0,"-"],matcher:[21,0,0,"-"],train:[21,0,0,"-"]},"alonet.deformable_detr.backbone":{Backbone:[19,1,1,""],BackboneBase:[19,1,1,""],Joiner:[19,1,1,""]},"alonet.deformable_detr.backbone.Backbone":{training:[19,4,1,""]},"alonet.deformable_detr.backbone.BackboneBase":{training:[19,4,1,""]},"alonet.deformable_detr.backbone.Joiner":{training:[19,4,1,""]},"alonet.deformable_detr.criterion":{DeformableCriterion:[21,1,1,""],sigmoid_focal_loss:[21,5,1,""]},"alonet.deformable_detr.criterion.DeformableCriterion":{get_metrics:[21,2,1,""],loss_labels:[21,2,1,""],training:[21,4,1,""]},"alonet.deformable_detr.deformable_detr":{DeformableDETR:[20,1,1,""],build_deformable_detr_r50:[20,5,1,""]},"alonet.deformable_detr.deformable_detr.DeformableDETR":{INPUT_MEAN_STD:[20,4,1,""],build_backbone:[20,2,1,""],build_decoder:[20,2,1,""],build_decoder_layer:[20,2,1,""],build_positional_encoding:[20,2,1,""],build_transformer:[20,2,1,""],forward:[20,2,1,""],forward_class_heads:[20,2,1,""],forward_heads:[20,2,1,""],forward_position_heads:[20,2,1,""],get_outs_filter:[20,2,1,""],get_outs_labels:[20,2,1,""],inference:[20,2,1,""],training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50":{DeformableDetrR50:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50.DeformableDetrR50":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune":{DeformableDetrR50Finetune:[20,1,1,""],DeformableDetrR50RefinementFinetune:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune.DeformableDetrR50Finetune":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune.DeformableDetrR50RefinementFinetune":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_refinement":{DeformableDetrR50Refinement:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50_refinement.DeformableDetrR50Refinement":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_transformer":{DeformableTransformer:[19,1,1,""],DeformableTransformerDecoder:[19,1,1,""],DeformableTransformerDecoderLayer:[19,1,1,""],DeformableTransformerEncoder:[19,1,1,""],DeformableTransformerEncoderLayer:[19,1,1,""],build_deforamble_transformer:[19,5,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformer":{forward:[19,2,1,""],gen_encoder_output_proposals:[19,2,1,""],get_proposal_pos_embed:[19,2,1,""],get_valid_ratio:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerDecoder":{decoder_forward:[19,2,1,""],forward:[19,2,1,""],pre_process_tgt:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerDecoderLayer":{decoder_layer_forward:[19,2,1,""],forward:[19,2,1,""],forward_ffn:[19,2,1,""],pre_process_tgt:[19,2,1,""],training:[19,4,1,""],with_pos_embed:[19,2,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerEncoder":{forward:[19,2,1,""],get_reference_points:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerEncoderLayer":{forward:[19,2,1,""],forward_ffn:[19,2,1,""],training:[19,4,1,""],with_pos_embed:[19,2,1,""]},"alonet.deformable_detr.matcher":{DeformableDetrHungarianMatcher:[21,1,1,""]},"alonet.deformable_detr.matcher.DeformableDetrHungarianMatcher":{hungarian_cost_class:[21,2,1,""],training:[21,4,1,""]},"alonet.deformable_detr.train":{LitDeformableDetr:[21,1,1,""]},"alonet.deformable_detr.train.LitDeformableDetr":{add_argparse_args:[21,2,1,""],build_criterion:[21,2,1,""],build_matcher:[21,2,1,""],build_model:[21,2,1,""],configure_optimizers:[21,2,1,""],training:[21,4,1,""]},"alonet.detr":{backbone:[23,0,0,"-"],callbacks:[26,0,0,"-"],criterion:[26,0,0,"-"],detr:[25,0,0,"-"],detr_r50:[25,0,0,"-"],detr_r50_finetune:[25,0,0,"-"],matcher:[26,0,0,"-"],train:[26,0,0,"-"],transformer:[23,0,0,"-"]},"alonet.detr.backbone":{Backbone:[23,1,1,""],BackboneBase:[23,1,1,""],FrozenBatchNorm2d:[23,1,1,""],Joiner:[23,1,1,""],build_backbone:[23,5,1,""],get_rank:[23,5,1,""],get_world_size:[23,5,1,""],is_dist_avail_and_initialized:[23,5,1,""],is_main_process:[23,5,1,""]},"alonet.detr.backbone.Backbone":{training:[23,4,1,""]},"alonet.detr.backbone.BackboneBase":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.backbone.FrozenBatchNorm2d":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.backbone.Joiner":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.callbacks":{DetrObjectDetectorCallback:[26,1,1,""]},"alonet.detr.callbacks.DetrObjectDetectorCallback":{on_train_batch_end:[26,2,1,""],on_validation_epoch_end:[26,2,1,""]},"alonet.detr.criterion":{DetrCriterion:[26,1,1,""]},"alonet.detr.criterion.DetrCriterion":{forward:[26,2,1,""],get_loss:[26,2,1,""],get_metrics:[26,2,1,""],get_statistical_metrics:[26,2,1,""],loss_boxes:[26,2,1,""],loss_labels:[26,2,1,""],training:[26,4,1,""]},"alonet.detr.data_modules":{coco_detection2detr:[24,0,0,"-"],coco_panoptic2detr:[24,0,0,"-"],data2detr:[24,0,0,"-"]},"alonet.detr.data_modules.coco_detection2detr":{CocoDetection2Detr:[24,1,1,""]},"alonet.detr.data_modules.coco_detection2detr.CocoDetection2Detr":{setup:[24,2,1,""]},"alonet.detr.data_modules.coco_panoptic2detr":{CocoPanoptic2Detr:[24,1,1,""]},"alonet.detr.data_modules.coco_panoptic2detr.CocoPanoptic2Detr":{setup:[24,2,1,""]},"alonet.detr.data_modules.data2detr":{Data2Detr:[24,1,1,""]},"alonet.detr.data_modules.data2detr.Data2Detr":{add_argparse_args:[24,2,1,""],setup:[24,2,1,""],train_dataloader:[24,2,1,""],train_transform:[24,2,1,""],val_dataloader:[24,2,1,""],val_transform:[24,2,1,""]},"alonet.detr.detr":{Detr:[25,1,1,""]},"alonet.detr.detr.Detr":{INPUT_MEAN_STD:[25,4,1,""],build_backbone:[25,2,1,""],build_bbox_embed:[25,2,1,""],build_class_embed:[25,2,1,""],build_decoder:[25,2,1,""],build_decoder_layer:[25,2,1,""],build_positional_encoding:[25,2,1,""],build_transformer:[25,2,1,""],forward:[25,2,1,""],forward_class_heads:[25,2,1,""],forward_heads:[25,2,1,""],forward_position_heads:[25,2,1,""],get_outs_filter:[25,2,1,""],get_outs_labels:[25,2,1,""],inference:[25,2,1,""],training:[25,4,1,""]},"alonet.detr.detr_r50":{DetrR50:[25,1,1,""],main:[25,5,1,""]},"alonet.detr.detr_r50.DetrR50":{training:[25,4,1,""]},"alonet.detr.detr_r50_finetune":{DetrR50Finetune:[25,1,1,""]},"alonet.detr.detr_r50_finetune.DetrR50Finetune":{training:[25,4,1,""]},"alonet.detr.matcher":{DetrHungarianMatcher:[26,1,1,""],build_matcher:[26,5,1,""]},"alonet.detr.matcher.DetrHungarianMatcher":{forward:[26,2,1,""],hungarian:[26,2,1,""],hungarian_cost_class:[26,2,1,""],hungarian_cost_giou_boxes:[26,2,1,""],hungarian_cost_l1_boxes:[26,2,1,""],training:[26,4,1,""]},"alonet.detr.train":{LitDetr:[26,1,1,""]},"alonet.detr.train.LitDetr":{add_argparse_args:[26,2,1,""],assert_input:[26,2,1,""],build_criterion:[26,2,1,""],build_matcher:[26,2,1,""],build_model:[26,2,1,""],callbacks:[26,2,1,""],configure_optimizers:[26,2,1,""],forward:[26,2,1,""],inference:[26,2,1,""],run_train:[26,2,1,""],training:[26,4,1,""],training_step:[26,2,1,""],validation_step:[26,2,1,""]},"alonet.detr.transformer":{Transformer:[23,1,1,""],TransformerDecoder:[23,1,1,""],TransformerDecoderLayer:[23,1,1,""],TransformerEncoder:[23,1,1,""],TransformerEncoderLayer:[23,1,1,""],build_transformer:[23,5,1,""]},"alonet.detr.transformer.Transformer":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerDecoder":{decoder_forward:[23,2,1,""],forward:[23,2,1,""],pre_process_tgt:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerDecoderLayer":{decoder_layer_forward:[23,2,1,""],forward:[23,2,1,""],forward_post:[23,2,1,""],forward_pre:[23,2,1,""],pre_process_tgt:[23,2,1,""],training:[23,4,1,""],with_pos_embed:[23,2,1,""]},"alonet.detr.transformer.TransformerEncoder":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerEncoderLayer":{forward:[23,2,1,""],forward_post:[23,2,1,""],forward_pre:[23,2,1,""],training:[23,4,1,""],with_pos_embed:[23,2,1,""]},"alonet.detr_panoptic":{callbacks:[29,0,0,"-"],detr_panoptic:[28,0,0,"-"],detr_r50_panoptic_finetune:[28,0,0,"-"],train:[29,0,0,"-"]},"alonet.detr_panoptic.callbacks":{PanopticApMetricsCallbacks:[29,1,1,""],PanopticObjectDetectorCallback:[29,1,1,""]},"alonet.detr_panoptic.callbacks.PanopticApMetricsCallbacks":{add_sample:[29,2,1,""]},"alonet.detr_panoptic.callbacks.PanopticObjectDetectorCallback":{on_train_batch_end:[29,2,1,""],on_validation_epoch_end:[29,2,1,""]},"alonet.detr_panoptic.detr_panoptic":{PanopticHead:[28,1,1,""],main:[28,5,1,""]},"alonet.detr_panoptic.detr_panoptic.PanopticHead":{INPUT_MEAN_STD:[28,4,1,""],forward:[28,2,1,""],inference:[28,2,1,""],training:[28,4,1,""]},"alonet.detr_panoptic.detr_r50_panoptic_finetune":{DetrR50PanopticFinetune:[28,1,1,""]},"alonet.detr_panoptic.detr_r50_panoptic_finetune.DetrR50PanopticFinetune":{training:[28,4,1,""]},"alonet.detr_panoptic.train":{LitPanopticDetr:[29,1,1,""]},"alonet.detr_panoptic.train.LitPanopticDetr":{add_argparse_args:[29,2,1,""],build_criterion:[29,2,1,""],build_model:[29,2,1,""],callbacks:[29,2,1,""],precision:[29,4,1,""],run_train:[29,2,1,""],training:[29,4,1,""],training_step:[29,2,1,""],use_amp:[29,4,1,""],validation_step:[29,2,1,""]},"alonet.metrics":{compute_map:[16,0,0,"-"],compute_pq:[16,0,0,"-"]},"alonet.metrics.compute_map":{APDataObject:[16,1,1,""],ApMetrics:[16,1,1,""]},"alonet.metrics.compute_map.APDataObject":{add_gt_positives:[16,2,1,""],get_metrics:[16,2,1,""],is_empty:[16,2,1,""],push:[16,2,1,""]},"alonet.metrics.compute_map.ApMetrics":{add_sample:[16,2,1,""],calc_map:[16,2,1,""],init_data_objects:[16,2,1,""]},"alonet.metrics.compute_pq":{PQMetrics:[16,1,1,""],PQStatCat:[16,1,1,""]},"alonet.metrics.compute_pq.PQMetrics":{add_sample:[16,2,1,""],calc_map:[16,2,1,""],pq_average:[16,2,1,""],update_data_objects:[16,2,1,""]},"alonet.transformers":{mlp:[17,0,0,"-"],position_encoding:[17,0,0,"-"]},"alonet.transformers.mlp":{MLP:[17,1,1,""]},"alonet.transformers.mlp.MLP":{forward:[17,2,1,""],training:[17,4,1,""]},"alonet.transformers.position_encoding":{PositionEmbeddingSine:[17,1,1,""],build_position_encoding:[17,5,1,""]},"alonet.transformers.position_encoding.PositionEmbeddingSine":{forward:[17,2,1,""],training:[17,4,1,""]},"aloscene.bounding_boxes_2d":{BoundingBoxes2D:[32,1,1,""]},"aloscene.bounding_boxes_2d.BoundingBoxes2D":{FORMATS:[32,4,1,""],GLOBAL_COLOR_SET:[32,4,1,""],abs_area:[32,2,1,""],abs_pos:[32,2,1,""],append_labels:[32,2,1,""],area:[32,2,1,""],as_boxes:[32,2,1,""],boxes2abspos:[32,2,1,""],boxes2relpos:[32,2,1,""],boxes2xcyc:[32,2,1,""],boxes2xyxy:[32,2,1,""],boxes2yxyx:[32,2,1,""],boxes_hflip:[32,2,1,""],fit_to_padded_size:[32,2,1,""],get_view:[32,2,1,""],get_with_format:[32,2,1,""],giou:[32,2,1,""],giou_with:[32,2,1,""],iou:[32,2,1,""],iou_with:[32,2,1,""],nms:[32,2,1,""],rel_area:[32,2,1,""],rel_pos:[32,2,1,""],remove_padding:[32,2,1,""],xcyc:[32,2,1,""],xyxy:[32,2,1,""],yxyx:[32,2,1,""]},"aloscene.bounding_boxes_3d":{BoundingBoxes3D:[33,1,1,""]},"aloscene.bounding_boxes_3d.BoundingBoxes3D":{bev_boxes:[33,2,1,""],boxes_3d_hflip:[33,2,1,""],get_enclosing_box_2d:[33,2,1,""],get_vertices_3d:[33,2,1,""],get_vertices_3d_proj:[33,2,1,""],get_view:[33,2,1,""],giou3d:[33,2,1,""],giou3d_with:[33,2,1,""],iou3d:[33,2,1,""],iou3d_with:[33,2,1,""]},"aloscene.camera_calib":{CameraExtrinsic:[34,1,1,""],CameraIntrinsic:[34,1,1,""]},"aloscene.camera_calib.CameraExtrinsic":{get_view:[34,2,1,""]},"aloscene.camera_calib.CameraIntrinsic":{get_view:[34,2,1,""]},"aloscene.disparity":{Disparity:[35,1,1,""]},"aloscene.disparity.Disparity":{append_occlusion:[35,2,1,""],signed:[35,2,1,""],unsigned:[35,2,1,""]},"aloscene.flow":{Flow:[36,1,1,""]},"aloscene.flow.Flow":{append_occlusion:[36,2,1,""]},"aloscene.frame":{Frame:[37,1,1,""]},"aloscene.frame.Frame":{append_boxes2d:[37,2,1,""],append_boxes3d:[37,2,1,""],append_disparity:[37,2,1,""],append_flow:[37,2,1,""],append_labels:[37,2,1,""],append_points2d:[37,2,1,""],append_segmentation:[37,2,1,""],mean_std_norm:[37,2,1,""],norm01:[37,2,1,""],norm255:[37,2,1,""],norm_as:[37,2,1,""],norm_minmax_sym:[37,2,1,""],norm_resnet:[37,2,1,""],save:[37,2,1,""]},"aloscene.labels":{Labels:[38,1,1,""]},"aloscene.labels.Labels":{get_view:[38,2,1,""]},"aloscene.mask":{Mask:[39,1,1,""]},"aloscene.mask.Mask":{append_labels:[39,2,1,""],get_view:[39,2,1,""],iou_with:[39,2,1,""],mask2id:[39,2,1,""]},"aloscene.oriented_boxes_2d":{OrientedBoxes2D:[41,1,1,""]},"aloscene.oriented_boxes_2d.OrientedBoxes2D":{GLOBAL_COLOR_SET:[41,4,1,""],abs_pos:[41,2,1,""],append_labels:[41,2,1,""],boxes2abspos:[41,2,1,""],boxes2relpos:[41,2,1,""],corners:[41,2,1,""],get_view:[41,2,1,""],rel_pos:[41,2,1,""],rotated_giou:[41,2,1,""],rotated_giou_with:[41,2,1,""],rotated_iou:[41,2,1,""],rotated_iou_with:[41,2,1,""]},"aloscene.points_2d":{Points2D:[42,1,1,""]},"aloscene.points_2d.Points2D":{FORMATS:[42,4,1,""],abs_pos:[42,2,1,""],append_labels:[42,2,1,""],as_points:[42,2,1,""],fit_to_padded_size:[42,2,1,""],get_view:[42,2,1,""],get_with_format:[42,2,1,""],rel_pos:[42,2,1,""],remove_padding:[42,2,1,""],xy:[42,2,1,""],yx:[42,2,1,""]},"aloscene.tensors":{augmented_tensor:[31,0,0,"-"],spatial_augmented_tensor:[43,0,0,"-"]},"aloscene.tensors.augmented_tensor":{AugmentedTensor:[31,1,1,""]},"aloscene.tensors.augmented_tensor.AugmentedTensor":{COMMON_DIM_NAMES:[31,4,1,""],add_label:[31,2,1,""],add_property:[31,2,1,""],apply_on_label:[31,2,1,""],as_tensor:[31,2,1,""],clone:[31,2,1,""],cpu:[31,2,1,""],crop:[31,2,1,""],cuda:[31,2,1,""],drop_labels:[31,2,1,""],get_labels:[31,2,1,""],get_slices:[31,2,1,""],hflip:[31,2,1,""],pad:[31,2,1,""],recursive_apply_on_labels_:[31,2,1,""],rename:[31,2,1,""],rename_:[31,2,1,""],reset_names:[31,2,1,""],resize:[31,2,1,""],set_labels:[31,2,1,""],spatial_shift:[31,2,1,""],to:[31,2,1,""],vflip:[31,2,1,""]},"aloscene.tensors.spatial_augmented_tensor":{SpatialAugmentedTensor:[43,1,1,""]},"aloscene.tensors.spatial_augmented_tensor.SpatialAugmentedTensor":{H:[43,3,1,""],HW:[43,3,1,""],W:[43,3,1,""],append_cam_extrinsic:[43,2,1,""],append_cam_intrinsic:[43,2,1,""],append_mask:[43,2,1,""],batch:[43,2,1,""],batch_list:[43,2,1,""],get_view:[43,2,1,""],relative_to_absolute:[43,2,1,""],temporal:[43,2,1,""]},alodataset:{base_dataset:[1,0,0,"-"],chairssdhom_dataset:[2,0,0,"-"],coco_detection_dataset:[3,0,0,"-"],crowd_human_dataset:[4,0,0,"-"],flying_chairs2_dataset:[5,0,0,"-"],flyingthings3D_subset_dataset:[6,0,0,"-"],merge_dataset:[7,0,0,"-"],mot17:[8,0,0,"-"],sintel_disparity_dataset:[9,0,0,"-"],sintel_flow_dataset:[10,0,0,"-"],sintel_multi_dataset:[11,0,0,"-"],transforms:[12,0,0,"-"],waymo_dataset:[13,0,0,"-"]},alonet:{transformers:[17,0,0,"-"]},aloscene:{bounding_boxes_2d:[32,0,0,"-"],bounding_boxes_3d:[33,0,0,"-"],camera_calib:[34,0,0,"-"],disparity:[35,0,0,"-"],flow:[36,0,0,"-"],frame:[37,0,0,"-"],labels:[38,0,0,"-"],mask:[39,0,0,"-"],oriented_boxes_2d:[41,0,0,"-"],points_2d:[42,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","property","Python property"],"4":["py","attribute","Python attribute"],"5":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:property","4":"py:attribute","5":"py:function"},terms:{"0":[3,6,12,16,19,20,21,23,25,26,28,29,31,32,34,36,37,39,40,41,42,44,45,46,47,48,50,51,52,53,54,56,57,58],"0000":47,"00051602e":41,"00070727e":41,"00119355e":32,"00133255e":32,"00159851e":32,"00204472e":32,"00215137e":32,"00252999e":32,"00265454e":41,"00416320e":41,"00508215e":32,"00565671e":41,"00572965e":32,"00630891e":32,"00671225e":41,"00819157e":32,"01":[32,37,41,46,50],"01023613e":41,"01064562e":41,"01078656e":41,"01085460e":41,"01117099e":41,"01284568e":32,"01389831e":41,"01412360e":32,"01414503e":32,"01447680e":41,"01489084e":32,"01573518e":32,"0164":37,"01693818e":32,"01715315e":41,"01773214e":41,"01826448e":41,"02":[8,12,32,41],"02008869e":32,"02071244e":41,"02099545e":41,"02148027e":41,"02181674e":41,"02201126e":32,"02216669e":41,"02224198e":41,"02320372e":41,"02340613e":32,"02524928e":41,"02542744e":41,"02609146e":41,"02637461e":32,"02712669e":32,"02730001e":32,"02754232e":41,"02925814e":41,"02968636e":32,"03":[32,41],"0300":47,"03028475e":32,"03053516e":41,"03128608e":32,"03145841e":32,"03243918e":32,"03257652e":41,"0330":47,"03335943e":32,"03398554e":41,"03478593e":32,"03496304e":41,"03540064e":41,"03600335e":41,"03750293e":41,"03760871e":32,"03872254e":32,"03880605e":32,"04":[8,32,41],"04081079e":32,"04119481e":32,"04159":[19,20],"04167677e":41,"04227657e":41,"04303905e":41,"04376091e":41,"04388196e":32,"04425869e":41,"04437640e":32,"04463429e":41,"04551275e":41,"04587221e":32,"04680798e":41,"04717023e":41,"04761188e":32,"04761873e":41,"04771734e":41,"04772201e":32,"04795949e":32,"04809226e":41,"04822710e":32,"04852029e":41,"04874016e":32,"04947969e":41,"05":[8,44],"05019362e":32,"05144850e":32,"05169697e":32,"05185719e":41,"05273729e":41,"05301257e":41,"05314606e":41,"05625390e":32,"05640656e":32,"05667166e":41,"05756191e":41,"05776466e":41,"05919485e":32,"05920354e":41,"05957368e":41,"05972891e":32,"05994362e":32,"0600":47,"06002222e":32,"06026421e":41,"06148241e":32,"06148329e":32,"06210334e":41,"06325925e":32,"06413557e":32,"06576427e":32,"06577017e":32,"06581083e":41,"06702779e":41,"06757526e":41,"06783829e":32,"06792355e":32,"06797182e":32,"06819102e":32,"06850095e":41,"06857690e":41,"06963077e":41,"0700":47,"07021418e":32,"07183093e":32,"07209305e":41,"07346811e":41,"07361911e":41,"07424044e":32,"07534600e":32,"07556239e":32,"07561498e":41,"07725342e":32,"07753378e":32,"07776551e":32,"0781":40,"07870363e":32,"07880044e":41,"07924467e":41,"08":34,"0800":47,"08027281e":32,"08153991e":32,"08175227e":32,"08308663e":41,"08317980e":41,"08348359e":32,"08448056e":41,"08482371e":32,"08526510e":32,"08542211e":32,"0868":37,"08703782e":32,"08778544e":32,"08817792e":32,"08897503e":32,"08912699e":41,"08938472e":32,"09":8,"09025812e":32,"09036772e":41,"09057900e":41,"09094712e":32,"09124164e":41,"09149733e":41,"09254644e":41,"09388471e":32,"09415695e":32,"09463034e":32,"09494858e":32,"09652805e":41,"09653721e":32,"09710224e":32,"09790113e":32,"09801983e":41,"09888444e":32,"09910303e":41,"09964232e":32,"09974838e":32,"1":[1,7,12,13,19,20,21,23,25,26,28,29,31,32,34,35,36,37,39,40,41,42,43,44,45,46,47,49,51,52],"10":[8,40,44,45,46,47],"100":[15,25,40,46,47],"1000":[44,47,54],"10000":17,"10021417e":41,"10034638e":32,"10047256e":32,"10088848e":32,"101":23,"1024":[19,20],"10292364e":41,"10305588e":32,"10343312e":32,"10596702e":41,"106":47,"10611510e":41,"10633622e":41,"10671862e":41,"107":46,"1080":[56,57],"10899068e":32,"10948697e":41,"11":[8,40,44,48],"110":46,"1100":47,"11124226e":41,"11133025e":41,"11300245e":41,"11347827e":41,"11365090e":32,"114":46,"11412094e":41,"11456517e":41,"11509175e":41,"11572169e":41,"11691907e":32,"11817739e":41,"11863718e":41,"11874017e":41,"11886488e":32,"119":46,"11913570e":32,"12":[16,40,47],"120":[46,58],"1200":[42,47],"12021770e":41,"12169042e":41,"123":47,"12349890e":41,"124":46,"125":[46,47],"12538061e":41,"12544761e":41,"12556038e":32,"126":46,"12659113e":32,"127":46,"12716393e":41,"12764148e":41,"1280":[40,44,51,52],"12865234e":32,"12872":25,"12891664e":32,"12921613e":32,"12997630e":41,"13":[8,34,46],"13072216e":32,"13084059e":32,"13108735e":41,"13168020e":41,"133":[16,46],"13323163e":41,"1333":[44,45],"13334349e":41,"13351250e":41,"13454091e":32,"135":46,"13520974e":32,"13596292e":41,"13632814e":41,"13658118e":41,"138":46,"13817184e":32,"13832005e":41,"13842775e":41,"139":46,"13952673e":41,"14":16,"1400":47,"14032232e":32,"141":46,"14109694e":41,"14128119e":32,"14128824e":32,"142":46,"14213758e":32,"14247333e":32,"143":46,"14312282e":32,"14319587e":41,"14345273e":41,"14361201e":32,"14363851e":32,"14391735e":32,"14424412e":32,"14461673e":41,"145":46,"14539692e":32,"14662452e":32,"147":46,"14770915e":41,"148":46,"14851037e":41,"14901921e":32,"14911471e":41,"14935338e":32,"150":40,"15011014e":32,"15011558e":32,"15016455e":32,"15059445e":32,"15088625e":32,"151":46,"15179000e":41,"152":46,"15205015e":32,"15210213e":32,"153":46,"15333085e":41,"15350205e":41,"154":46,"15448396e":41,"155":46,"15506293e":41,"15524881e":32,"15566470e":41,"15570085e":41,"156":47,"15638290e":32,"15653182e":32,"15730277e":41,"15752975e":41,"15864931e":32,"15872935e":41,"159":47,"15981518e":41,"16":46,"160":[46,47],"1600":47,"16018163e":41,"16088879e":32,"16145634e":41,"16156656e":41,"16201906e":32,"163":46,"16348874e":32,"16391443e":41,"164":46,"16421821e":41,"16431226e":41,"16490777e":41,"16517083e":32,"16562790e":32,"166":[46,47],"16624872e":32,"16665622e":32,"167":46,"16817471e":41,"16826625e":32,"16886530e":41,"16887195e":41,"16902778e":32,"170":46,"17055869e":41,"17066111e":32,"17100105e":32,"17129627e":41,"17260859e":41,"17301763e":41,"17315123e":32,"17322725e":41,"17367997e":32,"17386789e":41,"17392880e":41,"17454702e":32,"17480377e":41,"17488391e":41,"175":47,"176":46,"17600741e":32,"17604837e":32,"17828159e":32,"179":[46,47],"17928680e":32,"17957460e":32,"18":23,"180":47,"18010833e":32,"181":46,"18129853e":41,"18177504e":32,"18192625e":41,"182":46,"18296965e":41,"18372489e":41,"184":46,"185":46,"18546802e":41,"18551544e":32,"18587281e":32,"18590377e":41,"186":46,"18610614e":41,"18648595e":32,"187":46,"18733375e":32,"18753064e":41,"18762107e":41,"18762606e":32,"188":46,"18838328e":32,"18845964e":41,"18892872e":32,"189":46,"18907316e":32,"18926775e":32,"18945779e":41,"18981384e":32,"19":47,"1900":47,"191":46,"19146957e":32,"192":46,"1920":[40,44,51,52,56,57],"19242426e":32,"193":46,"19328787e":32,"194":46,"195":46,"19576752e":32,"19592251e":32,"196":46,"19628180e":41,"19658102e":41,"197":46,"19779656e":32,"19797095e":32,"198":[46,47],"19818748e":41,"19879779e":41,"19900324e":41,"19934985e":41,"2":[1,3,5,7,12,13,15,19,20,21,26,29,32,33,36,37,40,41,42,44,45,46,47,49,51,52],"20":[1,16,40,42],"200":[40,46,54],"2000":47,"2005":25,"20071020e":32,"2010":[19,20],"20113772e":32,"2013":34,"2017":[3,50,54,57,58],"2018":58,"202":46,"20267645e":32,"203":46,"20330610e":41,"20360519e":32,"20390712e":32,"204":46,"20466817e":32,"2048":[23,25],"20483222e":32,"20574823e":41,"20592140e":41,"20625557e":32,"20627365e":41,"207":46,"20788081e":32,"208":46,"20816324e":41,"20873807e":41,"209":46,"20973016e":41,"20979209e":41,"21":16,"210":47,"211":46,"21105030e":32,"212":46,"2120":37,"21200398e":41,"2126":40,"21373030e":32,"214":46,"21516043e":41,"21535104e":32,"21552847e":41,"21621182e":41,"217":47,"21776547e":32,"2178":40,"21874327e":41,"21884820e":41,"219":47,"21935562e":41,"22":37,"220":[46,47],"22074427e":32,"221":46,"22128649e":32,"22136800e":32,"22140182e":41,"22236057e":41,"22349080e":41,"22355425e":32,"224":[20,25,28,37,46],"22413266e":41,"225":[20,25,28,37],"22500":47,"22549850e":41,"22604942e":32,"22653131e":32,"22662459e":41,"227":46,"22738850e":32,"22795155e":32,"22852834e":41,"22862365e":32,"229":[20,25,28,37,46],"22904466e":32,"22959938e":41,"23":16,"2300":47,"23045578e":41,"23059211e":32,"231":47,"23125445e":32,"23190917e":32,"23199876e":32,"23247777e":32,"23266388e":32,"23298118e":41,"23325032e":32,"23332853e":41,"23381958e":41,"235":46,"23599575e":32,"236":46,"23621722e":41,"23647925e":41,"23674810e":41,"238":46,"23802378e":32,"23885994e":32,"239":46,"23904364e":41,"23983146e":41,"23990757e":41,"23998607e":41,"24":42,"240":46,"2400":47,"24025177e":41,"241":46,"24135251e":32,"24157363e":41,"24232008e":41,"24270324e":32,"24286011e":32,"243":46,"24386672e":32,"24405108e":32,"24447882e":41,"24458203e":41,"24510975e":32,"24537933e":41,"24551036e":41,"24568951e":32,"24572259e":32,"246":46,"24603436e":32,"24609747e":32,"24663319e":41,"24813112e":41,"24834611e":32,"24873060e":32,"24927253e":32,"24974845e":41,"24997086e":41,"25":21,"250":[24,28,29],"25138723e":41,"25226439e":41,"25288036e":32,"25394419e":41,"25403250e":32,"25409470e":41,"25432243e":41,"25454067e":32,"25488366e":41,"25493990e":41,"255":[37,44,46,47],"25555607e":41,"256":[19,20,25,37],"257":47,"258":47,"25918035e":32,"2600":47,"26016131e":41,"26040759e":41,"26152547e":32,"26236267e":41,"26250020e":32,"263":47,"265":47,"266":47,"26601184e":32,"26812553e":32,"26835802e":32,"26910887e":32,"26920870e":32,"26956763e":41,"26990011e":41,"270":47,"2700":47,"27099208e":32,"27192841e":41,"27377255e":32,"27495814e":41,"275":47,"27503982e":32,"27529135e":32,"27608733e":32,"27636232e":32,"27690540e":32,"27724104e":41,"27760922e":41,"2779":37,"27835129e":41,"279":47,"27989183e":32,"28":42,"28001460e":41,"28007186e":32,"28085192e":32,"28136816e":41,"28142564e":32,"28164126e":32,"282":47,"28205277e":32,"28238067e":32,"28260103e":32,"28285096e":32,"28407752e":32,"28413887e":32,"28435191e":41,"28471761e":41,"28508218e":32,"28545721e":32,"286":47,"28731766e":32,"28751525e":32,"28767839e":32,"28774609e":41,"28852497e":32,"289":47,"28959047e":41,"28976645e":41,"29":45,"290":47,"29011612e":32,"29149030e":32,"29159375e":41,"29201944e":32,"29264704e":41,"29333737e":32,"294":47,"295":47,"296":47,"29713198e":32,"29738768e":41,"298":47,"29956402e":41,"29976252e":32,"2d":[13,33,34,36,40,41,42,44,46,47,48],"3":[1,13,20,21,25,32,33,34,37,40,41,42,44,46,47,49,51,52],"30":16,"300":[19,20,41,46,47,50],"3000":47,"30000":47,"30150565e":32,"30190983e":41,"302":47,"30243579e":41,"30263620e":41,"30273065e":41,"303":47,"30311304e":41,"30397866e":32,"30646462e":41,"30647157e":41,"30702881e":32,"30704639e":32,"30764343e":41,"30773518e":41,"31":47,"310":47,"3100":47,"31001691e":32,"31076889e":41,"31128677e":41,"31176982e":32,"31183136e":41,"31229004e":41,"3125":47,"315":47,"31666474e":32,"31696089e":32,"31799552e":32,"318":47,"31807037e":32,"31839992e":41,"31857782e":41,"31877959e":41,"31930795e":41,"31939487e":41,"32":47,"3200":47,"32080311e":41,"32133632e":32,"32206036e":32,"32284414e":41,"32289133e":41,"32321250e":41,"32341182e":32,"32366280e":41,"32369469e":41,"324":47,"32434518e":32,"32439932e":32,"32449319e":41,"325":47,"32575747e":41,"32596057e":41,"32635855e":41,"32765525e":41,"32853739e":32,"32861016e":32,"32861072e":41,"32959088e":32,"32966981e":41,"33":[16,47],"33000444e":32,"33112482e":41,"33177747e":32,"33192793e":41,"33229234e":32,"33270789e":32,"33455423e":41,"33465786e":41,"33531934e":41,"33592468e":41,"33595493e":32,"33667201e":32,"33678078e":41,"33735841e":32,"33760713e":41,"33810312e":32,"34":[23,47],"3400":47,"34068612e":41,"34103563e":32,"34146171e":41,"34189645e":41,"34305328e":32,"34473480e":41,"34515366e":41,"34645829e":41,"34748716e":32,"34829280e":32,"34837902e":32,"34851278e":41,"34916476e":41,"34930565e":41,"34949801e":41,"35":47,"3500":47,"35023717e":32,"35038607e":41,"35063551e":32,"35093884e":41,"35117482e":41,"35205932e":41,"35281259e":41,"35338970e":41,"35368218e":41,"35414255e":32,"355":47,"35567359e":32,"35577368e":32,"35584003e":41,"35621356e":41,"35636532e":41,"35703430e":41,"35719585e":41,"35765215e":41,"35784922e":32,"358":16,"35853428e":32,"359":47,"35962038e":41,"36":[16,47],"360":47,"36002306e":32,"36022002e":41,"36165975e":41,"36199434e":41,"362":47,"36231448e":41,"36241428e":32,"36244965e":32,"36276863e":41,"36280255e":41,"36290787e":32,"36317878e":32,"36335809e":32,"36422203e":32,"36431987e":32,"36445537e":41,"36521038e":32,"36553374e":41,"36569415e":41,"36606170e":32,"368":7,"36827042e":41,"36962272e":32,"36981317e":32,"37":47,"3700":47,"37088945e":32,"371":47,"37103539e":41,"37155793e":41,"37174634e":41,"37310897e":32,"37340015e":41,"37402211e":32,"37432704e":41,"37436600e":41,"37532005e":32,"37535755e":32,"37588967e":32,"37594808e":41,"376":47,"37664331e":32,"37664382e":32,"37693244e":32,"377":47,"37711062e":32,"37727856e":41,"37745020e":32,"37782600e":32,"37793197e":32,"37835736e":41,"37844146e":32,"37934865e":32,"38":16,"3800":47,"38015272e":32,"38051907e":32,"38175275e":41,"382":47,"38228521e":41,"38234810e":32,"38256748e":41,"384":[44,47],"38471419e":41,"38481435e":32,"385":47,"38641228e":41,"38698314e":41,"38740534e":32,"388":47,"38854459e":32,"38933458e":41,"38938577e":32,"38944222e":32,"38944719e":32,"38949338e":32,"38963467e":32,"39":[40,44,46,47],"39018344e":41,"39137950e":41,"39148222e":41,"39254133e":32,"393":47,"39320385e":32,"39329815e":41,"39344775e":32,"394":47,"395":47,"39508256e":32,"39590570e":41,"396":47,"39600003e":41,"39606577e":32,"39638784e":41,"39674161e":41,"39763363e":41,"39778635e":41,"39815691e":32,"39864336e":32,"39977294e":32,"3d":[10,13,30,33,34,37,46,48,49],"4":[7,13,19,20,21,25,26,28,29,32,33,34,37,40,41,42,44,46,47,48,50,51,52,56],"40":[16,46],"400":[35,36,39,44,46],"4000":47,"40034328e":41,"40113234e":32,"40114690e":32,"40357930e":32,"40482958e":32,"40582895e":32,"406":[20,25,28,37],"40639729e":32,"40649491e":32,"40667169e":32,"40729489e":32,"40735860e":32,"40752254e":32,"40962720e":41,"41":46,"41002649e":32,"41071030e":32,"4111":37,"41128808e":41,"41241494e":41,"41645203e":32,"41729016e":32,"41786611e":41,"41820729e":41,"41830554e":32,"41831862e":41,"41883656e":32,"41889615e":41,"41894734e":32,"41957316e":32,"42":[37,46],"4200":47,"42163535e":32,"422":46,"42204493e":32,"42271943e":41,"42281033e":32,"42288814e":41,"42297352e":41,"42312343e":41,"42385955e":32,"42528038e":41,"42606505e":32,"42685751e":32,"42730215e":41,"42735322e":41,"42819821e":32,"42853762e":32,"42895935e":41,"42904727e":32,"42925604e":32,"42938804e":32,"42956539e":32,"43":16,"4300":47,"43023087e":41,"43042841e":32,"43075657e":41,"43130401e":32,"43141384e":41,"43194812e":41,"432":16,"43413371e":41,"43423731e":32,"43444735e":41,"43445851e":32,"43492379e":41,"43538160e":41,"43570700e":41,"43597600e":41,"43636328e":41,"43676481e":32,"43878267e":32,"43905462e":32,"43952490e":32,"44":[16,47],"4400":47,"44063067e":32,"44081660e":41,"44172397e":32,"44270017e":41,"44286947e":32,"44385708e":32,"44392263e":32,"44398756e":32,"44399895e":32,"445":47,"44551811e":41,"44630692e":41,"44641618e":41,"44673896e":32,"44688566e":41,"44726311e":32,"44732274e":41,"44734519e":41,"44783429e":41,"44807442e":41,"44810261e":41,"449":16,"44939437e":41,"45":46,"45021435e":41,"45065270e":41,"45066351e":32,"45093133e":32,"45196455e":41,"45234899e":32,"45310316e":32,"45398947e":32,"45454722e":32,"45599786e":32,"456":[20,25,28,37],"45611063e":32,"45633296e":41,"45654602e":41,"45726122e":32,"45766837e":41,"458":42,"45811637e":41,"45848217e":32,"45859593e":32,"45876408e":41,"46":[16,46,47],"46008537e":41,"46064335e":41,"46084398e":32,"46232387e":41,"46315548e":41,"46315707e":41,"46321268e":32,"46385916e":41,"46388149e":32,"46406351e":32,"46417379e":32,"46599072e":32,"46606147e":41,"46674042e":41,"46745001e":41,"46897770e":41,"46924197e":32,"46929039e":41,"47":[16,47],"47057405e":41,"47103447e":32,"47156683e":32,"47200802e":32,"47289682e":32,"47393835e":32,"474":47,"47478223e":32,"47495599e":32,"47538196e":32,"47594311e":41,"476":47,"47603517e":41,"47700368e":41,"47708146e":32,"47869084e":41,"47907059e":41,"47917999e":41,"47944036e":32,"480":[44,45,47],"4800":47,"48028846e":32,"48030262e":41,"48033253e":32,"48081258e":41,"481":16,"48104179e":32,"48147176e":41,"48153055e":41,"48213761e":41,"48244354e":41,"48255898e":32,"485":[20,25,28,37],"48585800e":32,"48598628e":32,"48600382e":32,"48744851e":32,"48818781e":32,"48834808e":41,"48855822e":41,"48886058e":41,"48938736e":32,"48954334e":41,"48989136e":41,"49":[16,42],"4900":47,"49046972e":32,"49128421e":41,"49172560e":41,"49187534e":41,"49219411e":41,"49314570e":41,"49325455e":41,"49378200e":41,"49435524e":32,"49493830e":41,"49555979e":41,"49573932e":32,"49588511e":32,"496":7,"49621867e":41,"49626810e":41,"497":47,"49703236e":32,"49710231e":32,"49713774e":41,"49803308e":32,"49808423e":41,"49885566e":32,"49940534e":41,"49983560e":32,"4k":4,"5":[12,16,21,26,28,29,32,35,37,40,41,42,47,53,54,56],"50":[16,23,40],"500":[44,46,47,53],"5000":47,"50006558e":41,"50085920e":41,"50138889e":32,"50165707e":32,"50186277e":32,"50199273e":41,"50207678e":32,"50293453e":41,"50510245e":32,"50537991e":32,"50632170e":41,"50727936e":41,"50751731e":32,"50835760e":32,"50882574e":41,"50888647e":41,"5095":37,"50996342e":32,"51":[42,46],"5100":47,"51005229e":32,"51024932e":32,"51097186e":32,"51137867e":32,"512":[23,37,42,44,45],"51228718e":32,"51248612e":41,"51259558e":32,"51268173e":32,"51292242e":41,"51396928e":41,"51441696e":32,"51493693e":32,"51499297e":32,"51513787e":32,"51539646e":32,"5155":40,"51613569e":32,"51674214e":41,"51688990e":32,"51800828e":41,"51831804e":41,"51860355e":41,"51887899e":32,"51926796e":32,"51955247e":41,"51955534e":32,"52001675e":32,"52037835e":32,"52071569e":41,"52142056e":32,"52465819e":32,"52506672e":32,"52569220e":41,"52606388e":41,"52658803e":41,"52747529e":41,"52754593e":32,"52757911e":41,"52768676e":41,"52829437e":32,"52851087e":32,"52855277e":41,"52891400e":32,"52902272e":41,"52993649e":41,"53":[16,47],"53208562e":41,"53285631e":32,"53359910e":41,"53381575e":41,"53381805e":41,"53438169e":32,"53471212e":32,"53631860e":41,"537":16,"53742844e":41,"53758957e":32,"53860176e":41,"53878839e":32,"53923576e":32,"53975753e":32,"53979624e":41,"54008178e":41,"54110357e":41,"54207130e":32,"54268708e":32,"54319462e":41,"54334671e":41,"544":[44,45],"54466717e":41,"54512593e":41,"54525691e":32,"54569385e":32,"54574824e":41,"54686235e":41,"54779696e":32,"54782298e":41,"54862369e":41,"54995590e":32,"55":[16,47],"5500":47,"55054480e":41,"55063802e":32,"55228428e":32,"55293852e":41,"55350342e":41,"55354685e":41,"55362918e":41,"55396186e":32,"55398361e":41,"55427503e":41,"55437307e":41,"55583297e":32,"55609667e":32,"55835216e":32,"55939103e":41,"55966798e":41,"55986460e":32,"55989053e":32,"55989290e":32,"56010235e":41,"56051858e":41,"56194948e":32,"56211088e":41,"56224417e":32,"56350143e":32,"56414250e":41,"56469828e":41,"56527501e":41,"56560266e":32,"56565124e":41,"56720742e":32,"56725107e":32,"56746286e":41,"56804765e":41,"56809213e":32,"56877725e":32,"56884720e":41,"56892042e":41,"56978680e":41,"5700":47,"57129571e":41,"57217245e":41,"57303626e":41,"57429174e":41,"57439643e":32,"57449467e":32,"57587518e":32,"576":[44,45],"57713193e":32,"57783664e":41,"57817782e":32,"57948552e":41,"57962473e":32,"5800":47,"58061958e":41,"58137991e":32,"58148635e":32,"58234590e":41,"58270290e":32,"58273479e":32,"58316872e":41,"58318236e":32,"58339510e":41,"58376747e":41,"58519625e":32,"58551430e":32,"58569475e":32,"58687131e":32,"58765499e":41,"58785518e":41,"58853357e":41,"5900":47,"59014283e":41,"59044999e":32,"59054548e":41,"59073593e":32,"59134616e":32,"59178390e":41,"59196241e":41,"59210991e":32,"59247284e":32,"59308577e":41,"59315360e":32,"59333520e":32,"59462401e":41,"59469003e":32,"59475331e":32,"595":16,"59544810e":41,"59565306e":32,"59610660e":32,"59693846e":41,"59758080e":41,"59768601e":32,"59967107e":32,"59970786e":41,"5th":40,"6":[16,19,20,21,23,25,26,29,32,37,40,41,57],"60":16,"600":[35,36,39,44],"6000":47,"60041588e":41,"60048677e":32,"60049439e":41,"60173608e":32,"60185582e":32,"60189655e":32,"60212139e":41,"60247861e":41,"60250229e":41,"60351173e":41,"60356952e":41,"60495137e":32,"60587540e":41,"60610769e":41,"60780852e":32,"60788182e":41,"608":[44,45],"60953736e":32,"61018186e":32,"61021307e":32,"61022231e":32,"61079053e":41,"61103382e":32,"61237738e":41,"61368908e":32,"61420213e":32,"61428914e":32,"61433931e":41,"61495355e":41,"61506634e":41,"61564484e":41,"61587099e":41,"61600747e":32,"61617961e":41,"61637559e":32,"61654997e":32,"61665926e":41,"61676887e":32,"61688289e":32,"61717298e":32,"61718622e":32,"61773856e":41,"61922468e":41,"61952807e":32,"6200":47,"62113942e":32,"62143484e":41,"62263973e":32,"62274267e":41,"62281202e":41,"62404387e":32,"62477445e":32,"62584241e":41,"62588816e":41,"62596573e":32,"62599278e":32,"62633722e":41,"62679464e":41,"62688509e":32,"62728852e":41,"62787151e":41,"63336631e":32,"63396338e":32,"63399274e":41,"63399427e":32,"63423042e":41,"63503936e":41,"63510266e":41,"63545749e":32,"63574387e":32,"63692683e":41,"63707289e":41,"63736254e":41,"63906017e":41,"63913892e":41,"63981616e":32,"64":[17,46],"640":[44,45,46,47],"6400":47,"64021870e":32,"64131004e":41,"64142216e":41,"64166583e":41,"64243515e":32,"64344971e":32,"64359957e":41,"64385619e":41,"64421797e":41,"64427192e":41,"64432783e":32,"64600468e":41,"64662324e":41,"64701496e":41,"64750741e":32,"64756788e":32,"64870111e":41,"64894611e":41,"64911196e":41,"64915846e":41,"64933322e":32,"65":[16,47],"65067771e":32,"65088251e":41,"65098195e":41,"65109356e":32,"65119162e":32,"65147551e":41,"65152192e":41,"65204624e":32,"65284284e":32,"65303336e":32,"65400850e":32,"65421511e":32,"65483130e":41,"65610983e":32,"65652017e":32,"65752488e":32,"65766207e":41,"65893743e":32,"65907715e":32,"66":[16,46],"6600":47,"66010074e":32,"66092573e":32,"66148970e":32,"66270462e":41,"66285481e":41,"66434807e":32,"66452541e":32,"66494927e":32,"66543838e":41,"66659251e":41,"66659612e":41,"66723721e":41,"66735310e":41,"66822109e":32,"66857574e":32,"66905130e":41,"66917036e":32,"66917625e":32,"66960705e":41,"66971043e":41,"67186664e":41,"672":[44,45],"67294100e":32,"67331559e":32,"67376126e":41,"67384841e":32,"67483186e":41,"67483792e":41,"67499068e":41,"67527258e":41,"67567753e":41,"67653709e":41,"67672440e":32,"67693356e":32,"67700563e":41,"67713608e":32,"67761680e":41,"67807741e":32,"67831527e":32,"67877079e":32,"67928400e":41,"67975352e":41,"67979530e":41,"68":16,"68000022e":41,"68046141e":41,"68192688e":32,"68211478e":41,"68220504e":32,"68236094e":41,"68399237e":32,"68479265e":41,"68507392e":41,"68516922e":41,"68583363e":32,"68618004e":32,"68693320e":32,"68705006e":32,"68732776e":41,"68743685e":41,"68751827e":32,"68811191e":41,"68928579e":41,"68971811e":41,"68983008e":32,"69041584e":32,"69090609e":41,"69146903e":32,"69166225e":32,"69224194e":41,"69284451e":41,"69303778e":32,"69435785e":32,"69470848e":32,"69479280e":41,"69482472e":32,"69515134e":41,"69743604e":32,"69853080e":32,"69859848e":32,"69879433e":41,"69946622e":32,"69998771e":32,"7":[13,16,32,33,40,41],"70":16,"700":[44,53],"70002547e":41,"70008731e":41,"70072899e":32,"70087830e":41,"70103464e":41,"70163010e":41,"70273302e":41,"70313620e":41,"70339913e":32,"70343670e":32,"704":[44,45],"70430650e":41,"70582914e":41,"70594696e":32,"70613403e":32,"70685630e":32,"70712863e":32,"70718343e":41,"70738131e":41,"70792886e":32,"70804075e":32,"70921545e":41,"70954888e":32,"71":47,"7100":47,"71011013e":32,"71052133e":41,"71176431e":32,"71190578e":32,"71259938e":41,"71267741e":41,"71332358e":32,"71382250e":41,"71434822e":41,"71457265e":41,"71499615e":41,"71508068e":32,"71521833e":32,"71522226e":32,"71602391e":32,"71614282e":41,"71659425e":41,"71664230e":32,"71677147e":41,"71780270e":32,"71791272e":41,"71857940e":32,"72":47,"7200":47,"72007854e":41,"72028956e":41,"72103431e":32,"72150734e":32,"72187218e":41,"72226359e":41,"72271830e":41,"72366775e":32,"72665983e":32,"72679423e":41,"72939716e":41,"73100681e":41,"73111726e":32,"73210450e":41,"73226457e":41,"73304404e":41,"73455073e":32,"73455615e":41,"73510326e":32,"73568654e":41,"736":[44,45],"73635007e":32,"73657264e":41,"73661135e":41,"73737889e":32,"73823287e":32,"73981643e":41,"73997584e":32,"74":45,"74112186e":32,"74129347e":41,"74143943e":32,"74146505e":41,"74192287e":41,"74220980e":32,"74268780e":32,"74578149e":32,"74845295e":32,"74953936e":41,"74958449e":41,"74985858e":41,"75":16,"7500":47,"75011383e":41,"75019079e":41,"75021987e":32,"75101960e":41,"75140731e":41,"75259072e":32,"75654494e":32,"75667244e":41,"75735485e":32,"75753328e":41,"75873084e":41,"75963803e":32,"76":47,"7600":47,"76006581e":41,"76019975e":41,"76141550e":32,"76275759e":32,"76323899e":41,"76334942e":32,"76360560e":32,"76368554e":32,"76438300e":41,"76451505e":32,"76487271e":41,"76610047e":41,"768":[44,45],"76964335e":32,"77":[46,47],"77243902e":32,"77297886e":32,"77308245e":32,"77310918e":32,"77407023e":41,"77438827e":32,"77441592e":41,"77486195e":41,"77529370e":32,"77586006e":41,"77598847e":32,"77616067e":32,"77629993e":41,"77646448e":32,"77698357e":32,"77751604e":41,"77756810e":32,"77773580e":41,"77797875e":32,"77801385e":41,"77826693e":41,"77856308e":32,"779":16,"77914277e":41,"77931165e":32,"77946945e":32,"78":[16,47],"78037319e":41,"78091111e":41,"78129718e":41,"78176969e":32,"78197124e":32,"78372089e":41,"78424306e":41,"78438102e":32,"78455093e":41,"78460247e":32,"78463067e":41,"78496098e":41,"78535344e":32,"78593137e":41,"78637232e":41,"78722001e":32,"78777501e":32,"78862323e":41,"78869568e":41,"789":16,"79035263e":41,"79110819e":32,"79143252e":41,"79162661e":41,"79185960e":32,"79207557e":32,"79241608e":41,"79270662e":41,"79300919e":41,"79319073e":32,"79384543e":32,"79448105e":41,"79474811e":32,"79491331e":32,"79491716e":41,"795":16,"79509705e":32,"79603274e":32,"79725511e":32,"79792720e":41,"79839197e":41,"79853955e":41,"79890251e":32,"79909151e":41,"79979004e":41,"8":[3,12,16,19,20,23,25,32,33,40,41],"80":[16,40],"800":[44,45],"80122702e":32,"80164268e":41,"80249115e":41,"80379387e":41,"80387101e":32,"80421045e":32,"80481041e":32,"80484864e":41,"80548756e":41,"80576126e":41,"80758245e":32,"80799150e":41,"80824122e":41,"80945642e":41,"81":47,"8100":47,"81045856e":32,"81066590e":41,"81130519e":41,"81157385e":32,"81346552e":41,"81514809e":32,"81639375e":41,"81666755e":41,"81685497e":32,"81706826e":32,"81749518e":41,"81779440e":32,"81817431e":41,"82":47,"82050048e":32,"82143918e":41,"82177723e":41,"82201724e":32,"82227314e":32,"82237027e":41,"82255091e":41,"82357466e":41,"82402189e":41,"82422198e":32,"82470218e":41,"82508046e":41,"82511740e":32,"82536143e":32,"82588178e":32,"82621410e":41,"82679031e":32,"82708323e":32,"82773104e":32,"82937714e":32,"83030218e":32,"83051399e":32,"83091248e":41,"83101718e":41,"83109898e":41,"83324946e":32,"83329206e":41,"83478347e":32,"83557829e":32,"83647122e":32,"83708934e":41,"83763381e":32,"83780808e":32,"83815209e":41,"83872906e":41,"83890262e":32,"83936217e":32,"83961364e":32,"83987536e":41,"83995094e":32,"84":[46,47],"84010504e":41,"84090080e":41,"84101993e":32,"84146473e":41,"84158592e":32,"84164432e":32,"84251230e":32,"84268810e":41,"84279112e":32,"84296254e":41,"84301525e":32,"84331772e":32,"84339841e":32,"84388865e":41,"84463618e":41,"84465717e":32,"84538982e":41,"84682345e":41,"84749912e":41,"84768479e":32,"84845043e":41,"84866116e":41,"84896261e":41,"84996651e":32,"85":[16,47],"85021536e":32,"85091537e":32,"85120679e":41,"85211241e":41,"85238648e":32,"85366584e":41,"85393316e":41,"85620672e":41,"85671024e":41,"85685524e":32,"85841072e":41,"85921243e":41,"85940756e":32,"85944980e":32,"8600":47,"86003087e":32,"86123463e":32,"86166188e":32,"86326450e":41,"86393952e":41,"86460032e":41,"86501833e":41,"86584065e":41,"86661993e":32,"86730818e":32,"86814944e":41,"86865287e":41,"86865847e":41,"86893799e":41,"86910936e":41,"86956994e":41,"87143027e":41,"87310275e":41,"87415176e":41,"87435697e":32,"87451562e":41,"87564552e":32,"87583674e":41,"87755884e":32,"87812190e":32,"87899595e":32,"87901050e":32,"8800":47,"88070478e":41,"88312613e":41,"88357981e":41,"88372827e":41,"88466606e":32,"88564202e":32,"88586194e":32,"88630857e":41,"88660435e":32,"88764308e":32,"88766773e":32,"88837141e":32,"88881648e":41,"88891904e":32,"88915652e":41,"88978771e":41,"88997795e":41,"89":16,"89131362e":41,"89141990e":41,"89174282e":32,"89301891e":32,"89312950e":32,"89339351e":32,"89341429e":41,"89365069e":41,"89374835e":41,"89449456e":41,"89524910e":41,"89543529e":32,"89661277e":32,"89689550e":41,"89754352e":32,"89778210e":32,"89874894e":41,"89885559e":41,"89964217e":32,"9":[16,32,40,41,44],"90":16,"90004727e":41,"90063780e":32,"90076284e":32,"90100879e":32,"90206548e":32,"90246467e":41,"90291605e":32,"90313894e":41,"90345274e":32,"90623152e":41,"90666433e":41,"90748601e":32,"90754411e":32,"90760191e":32,"90879256e":32,"90948464e":32,"91":[20,21,25,26,51,52,54],"91010232e":41,"91144374e":41,"91153068e":41,"91191438e":32,"91256492e":41,"91299428e":41,"91303686e":32,"91348126e":41,"91404197e":32,"91447936e":32,"91559028e":32,"91583537e":32,"91605333e":32,"91632384e":41,"91752845e":32,"91784979e":41,"91825279e":32,"91838632e":41,"91866292e":32,"91948573e":32,"91977332e":41,"92020874e":41,"92035442e":32,"92058403e":41,"92086297e":32,"92103441e":41,"92270963e":41,"92275251e":32,"92363776e":32,"92396779e":32,"92420729e":32,"92519624e":41,"92566638e":32,"92764504e":41,"92827347e":41,"92906166e":32,"92927763e":41,"92941282e":41,"93008534e":32,"93014832e":41,"93024218e":32,"93101057e":32,"93132891e":32,"93189364e":32,"9325":37,"93250958e":32,"93299667e":32,"93357992e":32,"93365737e":32,"93404331e":41,"93574127e":32,"93591252e":32,"93610902e":32,"93714209e":32,"93715457e":32,"93814590e":41,"93820383e":41,"93923941e":41,"93950878e":41,"93970080e":41,"9400":47,"94003097e":32,"94014943e":41,"94024553e":41,"94055706e":32,"94087964e":32,"94134256e":41,"94207973e":41,"94254100e":41,"94374577e":32,"94398543e":32,"94429515e":32,"94497217e":32,"94516350e":41,"94535338e":41,"94554753e":41,"94617276e":32,"94652585e":41,"94653947e":41,"94669003e":32,"94687653e":32,"94712473e":32,"95":16,"9500":47,"95017101e":32,"95049659e":32,"95088060e":32,"95102486e":41,"95142190e":41,"95150689e":41,"95174015e":41,"95208665e":32,"95322363e":32,"95329289e":32,"95393683e":41,"95397235e":41,"95431976e":41,"95460682e":32,"95590970e":32,"95687222e":41,"95763808e":41,"95774269e":41,"95853576e":41,"95874170e":32,"95883835e":41,"95958340e":41,"95960441e":32,"96186448e":41,"96191997e":32,"96198589e":32,"96248508e":41,"96349881e":41,"96446913e":32,"96451715e":32,"96482586e":32,"96495936e":41,"96506289e":32,"96526082e":41,"96544465e":32,"96587537e":32,"96620758e":32,"96653278e":32,"96657595e":41,"96715933e":41,"96767380e":32,"96790655e":41,"9700":47,"97039598e":41,"97093255e":32,"97093540e":32,"97154195e":32,"97223866e":41,"97272138e":32,"97327524e":41,"97422870e":32,"97425823e":41,"97456217e":41,"97482060e":41,"97487153e":32,"97532977e":32,"97618669e":32,"97622434e":41,"97637064e":32,"97664087e":32,"97701028e":32,"97864437e":41,"97909964e":41,"97917295e":32,"97964669e":41,"98":[16,47],"9800":47,"98065324e":41,"98117998e":41,"98118201e":32,"98195025e":32,"98352541e":41,"98457295e":41,"98457887e":32,"98486173e":32,"98630104e":41,"98648200e":32,"98677499e":41,"98685712e":41,"98695788e":32,"98771755e":41,"98802021e":41,"98811528e":32,"99":46,"9900":47,"99009071e":32,"99061946e":41,"99098426e":41,"99201204e":32,"99214877e":41,"99223449e":32,"99253306e":41,"99328799e":32,"99377370e":41,"99413904e":41,"99460100e":32,"99507426e":32,"99540086e":41,"99682481e":32,"99706296e":32,"99775088e":41,"99808706e":41,"99813722e":32,"99829637e":32,"99835651e":32,"99921852e":32,"boolean":28,"break":44,"case":[3,13,26,34,37,39,44,51,52],"class":[0,1,2,4,5,6,7,8,9,10,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,34,35,36,37,38,39,40,41,42,43,45,46,49,50,51,52,53,54,56,57,58],"default":[0,1,5,7,13,15,16,20,21,24,25,26,28,29,31,32,33,35,39,40,41,42,43,44,46,49,50,54,56,57,58],"do":[1,7,24,26,31,43,47],"enum":1,"export":[48,49,51,52,53],"final":[9,10,11,15,16,25,33,42,43,54,57,58],"float":[16,20,21,25,26,28,29,31,32,39,44],"function":[3,7,13,15,17,19,20,21,23,24,25,26,28,29,31,33,35,36,39,50,52,53,54,56,57,58],"import":[3,7,13,16,20,24,25,28,35,36,39,40,41,42,44,45,46,47,50,51,52,53,54,56,57,58],"int":[1,3,5,7,9,10,12,13,15,16,20,21,24,25,26,28,29,31,33,43,50],"long":47,"new":[1,3,4,11,12,13,16,21,26,28,29,31,32,35,36,37,41,42,43,46,47,58],"return":[1,2,3,5,6,7,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,32,33,35,37,38,39,41,42,43,44,45,46,50,51,52,53,57],"short":10,"static":[6,12,13,19,21,24,26,29,31,32,33,41,43,53,57],"super":[45,50],"switch":35,"true":[1,3,6,12,16,20,21,23,24,25,26,28,29,31,33,34,35,36,37,40,41,42,44,45,46,47,48,50,53,54,56,57,58],"try":[1,15,40,44,47,54],"while":[15,17,19,23,26,32,41,42,44,47,51,52],A:[1,7,13,15,19,21,23,24,26,31,35,36,39,43,45,46,47,50,51,52,54],As:[47,50,51,52],Be:47,But:[1,35,36,47],By:[3,16,20,21,28,33,40,43,44,46,47,50,54,58],For:[2,3,5,6,13,15,19,20,24,25,26,29,34,35,36,37,40,44,46,47,48,50,51,52,53,54,56,57,58],If:[1,3,4,13,15,16,20,21,24,25,28,31,32,33,35,36,37,38,39,40,41,42,43,47,50,53],In:[3,13,25,26,35,36,42,44,45,46,47,50,51,52,53,54,56,57,58],Is:[28,47],It:[3,7,20,25,28,39,42,44,46,50,51,52,56,57,58],Its:54,NOT:26,No:[12,46],Not:1,OR:15,On:[50,58],One:[6,9,10,11,13,15,26,29,37,42,47,48],Or:[36,42,48],TO:[51,52],That:[20,21,26],The:[1,2,3,4,5,7,9,10,12,13,15,17,20,21,25,26,27,28,29,32,33,34,35,36,37,39,42,44,45,46,47,48,51,52,54,57,58],Then:[23,33,58],There:[50,57],These:[20,25,28,57],Theses:15,To:[1,2,3,4,5,6,9,10,11,15,20,25,28,32,37,45,47,56],Will:37,With:[33,35,36,45,50,53,58],_:[16,51,52],__:[13,51,52],__calib:13,__call__:[51,52],__camera_label:13,__getitem__:[1,31],__image0:13,__image1:13,__image2:13,__image3:13,__image4:13,__init__:[13,45,50,51,52],__lidar_label:13,__pose:13,__test:13,__train:13,__valid:13,__velodyn:13,_annot:[3,53],_h:12,_id:6,_load_ann:50,_load_imag:50,_match_image_ann:50,_prepar:13,_r:12,_size:12,_w:12,ab:[19,20,25],abl:[26,28,50,53,58],about:[2,5,6,24,26,42,45,49,50,53,54,56,57,58],abov:[20,25,26,28,40,42,44,47,51,52],abs_area:32,abs_box:41,abs_po:[32,41,42,47],absolut:[31,32,33,35,37,41,42,47,50],accel:13,acceler:13,access:[3,13,15,53],accord:[27,50,53,54,56],accordingli:[44,46],account:[26,29],accross:26,accumul:[26,29],accumulate_grad_batch:[26,29,56,57],achiev:[56,57,58],across:[15,21,26,43,46],activ:[19,20,21,23,25,28,52],activation_fn:[20,21,52,53],ad:[21,23,32,35,36,37,39,41,42,43,46],adamw:[21,26],adapt:[29,45],add:[13,15,16,21,26,29,37,43,44,47,56,57,58],add_argparse_arg:[21,24,26,29,53,54,56,57,58],add_gt_posit:16,add_label:31,add_properti:31,add_sampl:[15,16,29,56],addit:[13,24,25,26,29,36,46,50],addition:15,adit:[20,25],adjust:24,advantag:[40,54],advis:3,affin:23,after:[3,15,35,36,50,51,52],afterward:[17,19,23],aggreg:15,air:[51,52],airplan:47,albedo:[10,11],aldreadi:13,algorithm:[33,41],align:[33,41],all:[1,2,4,7,12,13,15,16,17,19,20,21,23,24,25,26,28,31,33,43,44,45,46,47,50,53,54,57,58],all_gt:8,alley_1:11,alley_2:11,allow:[3,25,27,28,39,50,53,56,57,58],alocept:[1,3,4,11,12,13,23,44,45,50,51,52,53,54,56,57,58],aloception_config:50,alodataset:[1,2,3,4,5,6,7,8,9,10,11,12,13,16,31,35,36,37,40,45,46,47,48,49,50],alodataset_config:[1,3,4,11,13],alonet:[15,16,17,19,20,21,23,24,25,26,28,29,48,49,51,52,53,54,56,57,58],along:[23,33],aloscen:[1,3,5,6,9,10,11,12,13,16,20,21,24,25,26,28,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,47,48,49,50,51,52,53,54],alotransform:12,alpha:21,alreadi:[4,13,20,25,28],also:[3,17,22,33,35,36,40,41,42,46,47,50,53,54,57,58],although:[17,19,23],alwai:[12,13],ambush_2:11,ambush_4:11,ambush_5:11,ambush_6:11,ambush_7:11,among:[21,26],amount:3,an:[1,3,9,15,16,20,25,26,28,35,36,37,39,40,42,44,45,47,50,51,52,53,54,57,58],angl:41,ani:[1,3,11,15,23,25,28,43,44,46,47,51,52],anim:[10,54],ann_fil:[3,4,45,50],ann_fold:50,ann_img_train_0:50,ann_img_train_1:50,ann_img_train_l:50,ann_img_val_0:50,ann_img_val_1:50,ann_img_val_m:50,ann_train:50,ann_valid:50,annot:[3,9,24,39,45,50,58],anoth:[3,39,40,41,46,47],ap50:16,ap:[14,29,49,56],apdataobject:16,api:[0,30,44,47,51,52],apli:12,apmask:16,apmetr:[15,16,56],apmetricscallback:[15,29],append:[16,24,25,26,29,37],append_boxes2d:[20,25,28,37,47,50],append_boxes3d:37,append_cam_extrins:43,append_cam_intrins:43,append_dispar:37,append_flow:37,append_label:[32,37,39,41,42,50],append_mask:43,append_occlus:[35,36],append_points2d:37,append_segment:[28,37],appl:47,appli:[7,12,16,20,21,25,26,31,35,36,43,46,50],applic:[3,50,53,54],apply_on_label:31,approach:56,approxim:[33,41],ar:[3,4,9,10,13,15,20,21,23,25,26,28,29,31,32,35,36,37,39,41,42,43,44,45,46,47,48,50,53,57,58],architectur:[14,18,22,24,25,28,45,48,49,53,54,56,57,58],area:[16,32,41,42,43,44,47],arg:[9,10,11,12,15,19,20,21,24,25,26,28,29,31,32,33,34,35,36,37,38,39,41,42,43,51,52,53,54,56,57,58],argpars:[45,53,54,56,57,58],argument:[20,21,24,25,26,29,49,56,58],argument_group:[26,29],argumentpars:[21,24,26,29,45,53,54,56,57,58],around:33,arrai:[32,35,36,37,39,41,42],articl:56,arxiv:[19,20,25],as_box:32,as_point:42,as_tensor:[31,35,47,51,52],aspect:47,assert:46,assert_input:26,assert_integ:43,assign:[21,26,28,50],associ:[15,20,42,46,51,52],associt:13,assum:[50,53,58],attach:[6,9,10,11,15,31,32,35,36,37,39,41,42,43,44,46,49],attent:[17,19,48,52,53,54,56,57,58],attribut:[3,15,16,24,26,28,29,50,54,56,57,58],aug_tensor_compat:[23,25],augment:[12,24,25,37,42,44,45,46,48,49],augmented_tensor:[31,32,33,34,38,41,42,43],augmentedtensor:[31,32,33,34,38,41,42,43],augmentend:31,augmet:43,augmt:31,author:1,auto_restor:31,auto_restore_nam:31,automat:[3,15,20,35,36,46,47,50],automaticli:[15,25,44],aux:28,aux_loss:[20,21,25,26,28,29,51,52],aux_loss_stag:[21,26,29],aux_output:[20,25,26,28,29],auxialiri:26,auxilari:[20,21,25,28,51,52],auxiliar:[26,29],auxiliari:[21,25],avaiabl:46,avail:[9,10,26,46,50,57,58],averag:[15,16,56],axi:[15,31,33,34,51,52],b:[13,15,19,31,37,41,42,44,46,47,51,52],back:33,backbon:[18,20,22,25,26,28,46],backbone_nam:[20,25],backbonebas:[19,23],background:[20,25,26,28,29],background_class:[25,28,51,52],backpack:47,backward:6,bad:[33,41],bag:47,bag_box:47,balanc:21,ball:47,bamboo_1:11,bamboo_2:11,banana:47,bandage_1:11,bandage_2:11,base:[0,2,3,4,5,6,7,8,9,10,11,12,13,14,16,17,18,19,21,22,23,24,26,27,28,29,31,32,33,34,35,36,37,38,39,41,42,43,44,46,47,48,49,53,54,56,57,58],base_dataset:1,base_metr:[15,29],base_metrics_callback:15,base_model:28,base_weight:[20,25,28],basebal:47,basedataset:[1,5,6,7,9,10,11,31],basenam:[51,52],basic:[0,18,22,27,30,40,49],bat:47,batch:[1,3,7,13,15,16,20,21,23,24,25,26,28,29,37,42,43,44,46,47,50],batch_cost_matrix:26,batch_fram:43,batch_idx:[15,26,29],batch_list:[3,13,16,20,24,25,28,43,44,51,52,53,54,56,57,58],batch_siz:[1,3,7,13,20,21,24,25,26,28,44,45,51,52,56,58],batchnorm2d:23,batchnorm:[19,23],bb_output:[20,25,28],bbox:13,bbox_3d:13,bbox_emb:25,bbox_proj:13,bear:47,becaus:[26,33,35,36,37,47,51,52,58],becom:35,bed:47,been:[3,32,35,36,53,54],befor:[15,23,25,29,31,47],begin:24,behavior:[19,23],behind:[33,47],being:[15,20,21,26,44,47],below:[3,37,40,50,54,56,57,58],bench:47,benefit:[46,51,52],best:[26,33,41],between:[9,10,12,21,24,26,31,32,35,37,39,40,41,42,46,51,52,56],betweent:56,betwen:[32,42],bev_box:[33,40],beyond:[45,46],bicycl:47,binar:28,binari:[20,21,25,28,37,39],bird:[33,40,47],blank:41,block:[20,28],boat:47,bondixboxes2d:47,book:47,bool:[6,12,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,39,41,42,43,45],both:[16,42,47,57],bottl:47,bottom_offset:31,bound:[20,21,25,26,28,33,40,53,56],bounding_boxes_2d:[32,37,40,46],bounding_boxes_3d:[33,37],boundingboxes2d:[13,15,16,20,21,25,26,28,29,30,33,37,40,46,47,49,50],boundingboxes3d:[13,15,30,37,40,41,49],boungd:33,bout:26,bowl:47,box:[3,13,15,20,21,25,26,27,28,29,32,33,37,38,40,41,43,44,46,47,48,52,53,54,56,57,58],box_refin:20,boxes1:[32,33,40,41],boxes2:[32,33,40,41],boxes2abspo:[32,41],boxes2d:[13,15,16,20,21,25,26,28,29,32,37,38,43,44,46,47,50,53,54,56,57,58],boxes2relpo:[32,41],boxes2xcyc:32,boxes2xyxi:32,boxes2yxyx:32,boxes3d:[13,15,37,40],boxes_3d:[37,40],boxes_3d_hflip:33,boxes_format:[32,37,47,50],boxes_hflip:32,boxes_limit:4,boxes_set:37,bridg:[51,52],bright:12,brightness_r:12,bring:33,broccoli:47,brute:[33,41],bu:47,build:[20,21,24,25,26,29,48,53,54,56,57,58],build_backbon:[20,23,25],build_bbox_emb:25,build_class_emb:25,build_criterion:[21,26,29],build_decod:[20,25],build_decoder_lay:[20,25],build_deforamble_transform:19,build_deformable_detr_r50:20,build_match:[21,26,29],build_model:[21,26,29],build_position_encod:17,build_positional_encod:[20,25],build_transform:[20,23,25],built:[44,45,48],c:[13,31,35,36,37,39,40,42,44,46,47,50,51,52,53],cach:[15,16],cake:47,calc_map:[16,56],calcul:[15,16,21,29,33,39,40,41,56],calib:13,calibr:[13,30,40,49],call:[1,13,15,16,17,19,23,24,26,29,35,36,37,43,47,50,51,52,53],callabl:28,callback:[1,14,18,22,27,49],caluc:41,cam:37,cam_extrins:[13,33,37,40,43],cam_intrins:[13,33,40,43],cam_intris:13,camera:[6,9,10,11,13,30,33,35,40,44,46,48,49],camera_calib:[33,34,43],camera_id:13,camera_label:13,camera_paramet:13,camera_sid:35,cameraextrins:[13,33,34,40,43],cameraintrins:[13,33,34,40,43],camextrins:13,camintrins:13,can:[3,7,13,19,21,24,25,31,35,36,37,39,40,41,42,44,45,46,47,48,50,51,52,53,54,56,57,58],cap:53,car:47,care:[17,19,23,35,36,47],carrot:47,cat:[16,43,46,47,50,54],cat_label:16,categori:[3,16,21,26,39,50,53],caution:46,cave_2:11,cave_4:11,cd:48,cdist:47,cell:[47,51,52],center:[17,25,33,41],center_i:[20,25,28],center_x:[20,25,28],chair:[5,47],chairssdhom:[0,49],chairssdhom_dataset:2,chairssdhomdataset:[0,36,49],chang:[0,21,24,25,28,36,37,44,45,46,47,49,50,51,52,53,54,56,57,58],channel:[37,42],check:[4,13,16,20,26,40,48,53,54,56,57,58],checkout:37,checkpoint:[26,29,53,54,55],child:[1,12,43,54],choic:26,choos:[33,41,54],choosen:41,ckpt:[20,25,28],clasif:[21,26],class_emb:[20,25,28],class_nam:16,classif:[16,20,21,25,26,28,37,50,52],classifi:[21,26,50],clean:[9,10,11],clip:[26,29,44],clock:47,clockwis:41,clone:31,close:53,co:47,coco2017:[50,56,57,58],coco:[0,20,24,25,28,44,45,49,51,52,53,54,56,57,58],coco_dataset:[3,46,50],coco_detection2detr:24,coco_detection_dataset:3,coco_detr:[56,57],coco_load:[45,53,54,56,57,58],coco_panoptic2detr:24,coco_train:45,coco_v:45,cocodataset:[46,58],cocodetect:3,cocodetection2detr:[22,45,53,54,56,57,58],cocodetectiondataset:[3,16,24,44,45,46,47,50],cocopanoptic2detr:[22,49],cocopanopticdataset:[16,24],code:[35,36,40,46,50,53,54,56,57,58],coeffici:[26,29],coffe:[51,52],coher:47,color:[32,35,36,39,41,42,46],color_by_cat:39,colorjitt:12,colors_r:12,com:[15,48],combin:[11,25,46,56],come:15,command:[3,57],common:[46,50,51,52,53,54,56,57,58],common_dim_nam:31,compar:[1,7,46,51,52,56],comparison:[49,51,52],compat:[25,48],complementari:45,complet:[13,39,53,54],complex:[54,57,58],compon:45,compos:[12,44,45],compress:13,comput:[2,5,6,12,15,16,17,19,21,23,26,32,41,44,45,48,49,50,56,57,58],compute_map:16,compute_per_size_ap:16,compute_pq:16,compute_statistical_metr:26,concat:[44,46,47],concaten:[26,29,51,52,56,57,58],concept:[45,50],conclud:56,confer:[9,10],config:[1,3,11,49],configur:[3,20,21,24,26,28,48,50,53,54],configure_optim:[21,26],conflict_handl:[56,57,58],conjunct:54,connect:[24,27,28,45],connector:[14,22,49],consid:15,consist:[35,46],consol:[16,57],constant:[51,52],constructor:25,consult:[57,58],contain:[2,6,7,9,13,15,20,21,23,25,26,28,33,35,36,40,50,51,52,53],content:[31,35,47,50],contrast:12,convent:[30,33],converg:54,convers:13,convert:[13,31,33,41,46,47,51,52],convolut:20,cooordin:34,coordin:[13,20,21,25,26,28,33,34,36,40,41,50],copi:[23,35,53],core:[24,26,47],corner:41,correct:[3,26],correctli:[16,35,36,52],corredl:26,correspond:[2,3,5,26,28,50,51,52],cost:[21,26,56,57,58],cost_box:[21,26],cost_class:[21,26],cost_giou:[21,26],cottontail:3,couch:47,could:[1,13,31,32,41,42,47,50,56,57,58],counter:41,coupl:[28,29],cow:47,cpu:[16,25,28,31,53,54,56,57,58],crash:[51,52],creat:[13,19,20,24,25,26,28,30,32,33,37,38,39,41,42,43,44,45,46,47,50,51,52,53,54,56,58],creation:[35,36],criterion:[18,22,29],crop:[31,40,44,46,47],crope:31,cropped_fram:[40,46],cross:[16,26,29],cross_entropi:15,cross_entropy_valu:15,crowd:[0,44,49],crowd_fram:44,crowd_human_dataset:4,crowdhumandataset:[4,44],csv:50,cuda:[16,19,20,31,33,40,41,48,53,54,56,57,58],cup:[47,51,52],current:[1,11,15,31,32,37,42,46,47],custom:[3,20,25,26,28,29,46,49,52,53,54,56,57,58],customdataset:50,customdatasetsplitmix:50,customrandomcolor:12,cv2:[35,36,53],cvpr2016:6,cvpr2017:2,cyclist:13,d:50,d_ffn:19,d_model:[19,23],data2detr:22,data:[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,22,26,28,29,35,36,37,40,42,44,46,47,48,49,53,54,56,58],data_load:[26,29,45,53,54,56,57,58],data_modul:24,data_point:16,databas:[0,24,49,50,53],dataload:[1,7,15,24,26,29],dataloader_idx:[15,26,29],datamodul:[24,45],datapipelin:31,dataset1:7,dataset2:7,dataset:[0,2,4,5,6,8,9,10,11,15,16,20,24,25,28,31,35,36,37,40,48,49,53,54,56,57,58],dataset_1:50,dataset_2:50,dataset_dir:[1,6,11,13,50],dataset_dir_prepar:4,dataset_n:50,dataset_name_1:50,dataset_name_2:50,dataset_name_n:50,datasets_download_path:1,datataset:50,date:47,dc5:20,ddp:24,dec_lay:20,dec_n_point:[19,20],dec_output:[25,28],decid:[3,25,28,50],declar:[3,24,58],decod:[19,20,21,23,25,28,51],decoder_forward:[19,23],decoder_lay:[19,23],decoder_layer_forward:[19,23],decoder_output:[19,23],decompress:13,decreas:32,deep:[48,50],def:[44,45,50,51,52,53,56,57],defin:[17,19,20,23,25,26,28,29,39,41,46,50,51,52,53,54,56,58],definit:[16,53,56,57,58],deform:[14,19,21,23,29,45,48,49,50,51,54,57,58],deformabe_finetun:53,deformable_detr:[19,20,21,48,51,52,53,56],deformable_detr_r50:20,deformable_detr_r50_finetun:20,deformable_detr_r50_refin:20,deformable_transform:19,deformablecriterion:21,deformabledetr:[20,51,52],deformabledetrhungarianmatch:21,deformabledetrinfer:52,deformabledetrr50:[20,21],deformabledetrr50finetun:[20,56],deformabledetrr50refin:[20,21,51,52],deformabledetrr50refinementfinetun:[20,51,52,53],deformabledetrtrtexport:[51,52],deformabletransform:19,deformabletransformerdecod:19,deformabletransformerdecoderlay:19,deformabletransformerencod:19,deformabletransformerencoderlay:19,deformanbl:49,demand:[56,57,58],demonstr:40,depend:[13,47,50],deploi:[48,53],deploy:[51,52],deriv:10,describ:[25,26,50,56,57,58],descript:[16,42],deseri:[51,52],design:[46,48],desir:[3,12,16,32,33,39,41,42,53,54],destroyallwindow:53,detail:[9,10,13,19,20,40,42,50,56],detect:[0,15,20,21,25,26,28,44,49,50,53,54,56,57,58],detections_set:8,detector:[14,49],detr50:[49,57,58],detr:[14,16,17,19,21,23,26,28,29,45,48,49,50,58],detr_50:[26,29],detr_finetun:[53,54],detr_model:28,detr_modul:[28,58],detr_panopt:[16,28,29,58],detr_r50:[25,29],detr_r50_finetun:25,detr_r50_panoptic_finetun:28,detrcriterion:[21,26,29],detrfinetun:28,detrhungarianmatch:[21,26],detrinfer:51,detrobjectdetectorcallback:26,detrr50:[16,20,25,27,51,53,54,56],detrr50finetun:[25,28,51,54,57,58],detrr50panopticfinetun:[28,58],detrtrtexport:51,develop:[16,50,56,57,58],devic:[16,19,20,25,26,28,31,33,40,41,51,52,53,54,56,57,58],diagram:28,dice:29,dict:[1,2,13,15,16,20,21,25,26,28,29,31,32,33,37,39,41,42,43],dict_input:4,dictionari:[13,16,26,29],dictionnari:[20,25,28],did:[32,42],differ:[7,15,25,26,29,35,36,39,43,44,47,50,51,52,53,54,56,57,58],dilat:[19,20,23,25],dim:[21,25,26,43,46,47],dim_feedforward:[19,20,23,25],dim_valu:31,dimens:[1,7,25,26,29,31,33,34,37,39,42,43,44,47,51,52],dimension:[37,42],dine:47,dir:[4,13],directli:[1,26,58],directori:[1,2,3,4,11,13,50,51,52],dirnam:[51,52],disk:[4,13,44],disp:[6,9,11,35],disp_flip:35,disp_format:35,disp_label:11,disp_occ:[6,9,11],disp_sign:35,disp_tensor:35,dispar:[0,6,30,37,39,46,48,49],displai:[20,25,28,32,35,36,38,39,40,41,42,44,46,47,48],dispnet:6,disregard:[20,25,28],dist:47,distanc:35,document:[40,45,50,57,58],doe:[3,13,15,47,50,58],dog:[47,50,54],don:[1,26,29,46,47,51,52],done:[35,36,45,48,50],donut:47,down:4,download:[0,1,2,5,6,9,10,13,46,49,50,53,56,57,58],download_sampl:1,downward:[13,33],dpi:58,dpm:8,dpm_sequenc:8,drasticli:4,drier:47,drop_label:31,dropout:[19,20,23,25],dtype:[31,51,52],due:[15,53,54],dure:[15,35,36],dx:[33,37],dy:[33,37],dynam:[24,37,42],dz:[33,37],each:[2,3,6,7,9,10,11,13,15,16,19,20,21,24,25,26,28,29,31,32,33,37,39,41,43,44,46,50],eas:49,easi:[21,30,44,46],easier:[37,48],easili:[40,57],eccv2018:5,edit:[15,29],edu:32,effect:[1,4],effici:26,either:[1,15,20,21,24,33,51,52],element:[1,3,9,10,15,16,21,24,25,26,28,32,37,42,43,50],eleph:47,els:[16,53,54,56,57,58],emb:[25,26,29],embed:[17,19,23,53,54],empti:[13,16,42],en:45,enabl:[53,54],enc_lay:20,enc_n_point:[19,20],enc_output:[25,28],enclos:[33,40,41,42],enclosing_boxes_2d:40,enclosing_typ:[33,41],encod:[14,19,23,24,25,28,38,42,47,49,50],encoder_lay:[19,23],end:[6,13,15,23,25,26,28,29,31,56,57],endup:47,engin:[51,52],engine_path:[51,52],enough:[43,51,52],ensur:19,enter:[24,26,29,57],entir:[3,56,57,58],entri:[21,26],entropi:[26,29],enumer:[1,44,56],environ:48,eo:[26,29],eos_coef:[21,26,29],ep:23,epoch:[15,26,29,53,54],equal:[3,12,21,25,42],equival:31,error:[15,21,53,54],estim:16,etc:13,eval:[16,51,52,56,57,58],evalu:[10,16,45],even:[35,36,37,42,44],everi:[15,17,19,23,24,26,29,35,36],everyth:46,exampl:[3,15,21,22,31,35,36,37,39,40,42,44,45,46,47,48,50,51,52,56,57,58],example_label:50,example_label_nam:50,except:[1,16,20,21,24,26,29,39],exclud:43,execut:[3,13,16,50,51,52],exist:[3,35,36,37,39,50],exmpl:45,expe_nam:[26,29,45,53,54,56,57,58],expect:[13,15,28,37,45,47],experi:[26,29],experiment:46,explain:[50,51,52,53,54,56,57,58],explan:[42,43,56],explicit:35,explicitli:57,explor:49,expon:21,export_engin:[51,52],expos:[3,44,45],express:35,ext:6,extend:[46,47,48],extens:[9,45,54],extra:[23,35,36],extract:[13,20,25],extrins:[13,34,40],ey:[33,40],f:[7,29,44,56],face:53,facilit:[46,48,56,57,58],fact:[51,52],factor:21,fake:50,fals:[1,3,6,8,12,16,17,19,20,23,24,25,26,28,31,32,33,37,39,41,42,43,45,47,50,51,52],far:46,fast:54,faster:1,fastest:[33,41],featur:[17,19,23,28,30,33,37,46,49,50],feedfoward:25,ffn:17,figsiz:40,figur:[40,58],file:[1,3,4,11,13,20,24,25,28,35,36,39,48,49,51,52,53,54],filepath:2,fill:24,film:10,filter:[0,20,21,24,25,28,33,47,49],find:[2,33,41,54],finetun:[3,18,22,27,48,49,51,52,56,57,58],finish:[56,57,58],fire:47,first:[7,12,15,16,28,33,40,44,45,46,47,50],fit:[24,45],fit_to_padded_s:[32,42],five:56,fix:[23,24,53,54,57],fix_classes_len:[3,24],flip:[31,32,33,35,36,40,44,46,47],flipped_fram:[40,46,47],flo:36,float16:[51,52],float32:[50,51,52],flow:[0,2,5,6,9,30,35,37,39,44,46,48,49],flow_backward:6,flow_flip:36,flow_forward:36,flow_label:11,flow_occ:[6,10,11],flow_occ_backward:6,flow_tensor:36,flownet2:6,fly:[5,13,32,42],flying_chairs2_dataset:5,flyingchairs2:[0,49],flyingchairs2dataset:5,flyingthings3d:[0,49],flyingthings3d_subset_dataset:6,flyingthings3dsubsetdataset:[0,7,49],fn:16,focal:[21,29,34],focal_alpha:21,folder:[4,6,13,24,50,53,54],follow:[2,5,6,9,10,13,24,25,28,32,42,44,45,46,47,48,50,53,54,56,58],footprint:[51,52],forc:[33,41],fork:47,form:34,format:[3,13,26,32,33,35,41,42,46,50,51,52],former:[17,19,23],forward:[13,15,17,19,20,21,23,25,26,28,29,31,33,53,54,56,57],forward_class_head:[20,25],forward_ffn:19,forward_head:[20,25],forward_out:[20,25,28,51,52],forward_position_head:[20,25],forward_post:23,forward_pr:23,found:[3,13,44,50],fp16:[51,52],fp32:[51,52],fp:16,fpath:4,frame0:3,frame:[1,2,3,5,6,9,10,11,12,13,15,16,20,21,23,24,25,26,28,29,30,32,33,35,36,38,39,40,41,42,43,44,45,47,48,49,51,52,53,54,56,57,58],frame_01:[37,46],frame_1:46,frame_255:[37,46],frame_2:46,frame_3:46,frame_4:46,frame_flip:[35,36],frame_minmax_sym:37,frame_process:53,frame_resnet:[37,46],frame_s:[32,33,41,42,47],framei:32,framer:3,framework:[50,56,57,58],framr:47,frane:[37,42],frcnn:8,frcnn_sequenc:8,freez:28,freeze_detr:28,freiburg:[2,5,6],frisbe:47,from:[1,3,4,6,7,10,12,13,15,16,17,19,20,23,24,25,26,27,28,29,31,32,33,35,36,37,39,40,41,43,44,45,46,47,50,53,54,55,56,57,58],front:[13,37,44],front_fram:13,front_left:[13,40],front_right:[13,40],frozen:[19,23],frozenbatchnorm2d:23,ftmap_mask:17,full:[13,45,47,54],fulli:[1,13],func:31,funetunig:[53,54],future_past:6,fx:34,fy:34,g:41,gamma:21,gamma_r:12,gaussian_std:12,gcc:48,gen_encoder_output_propos:19,gener:[1,2,3,4,5,6,7,8,9,10,11,13,16,17,26,32,46,50],get:[1,3,7,12,13,15,16,19,20,24,25,26,28,29,31,32,33,35,36,37,39,40,41,42,44,46,49,50,53,54],get_dataset_dir:[1,11],get_enclosing_box_2d:[33,40],get_expe_info:[51,52],get_featlabels_el:6,get_features_label:6,get_file_id:2,get_filter_fn:28,get_fram:[2,6,13],get_frame_boxes2d:13,get_frame_boxes3d:13,get_frame_camera_paramet:13,get_label:31,get_loss:26,get_metr:[16,21,26],get_outs_filt:[20,25,28,51,52],get_outs_label:[20,25],get_proposal_pos_emb:19,get_rank:[15,23],get_reference_point:19,get_seg:13,get_sequ:2,get_size_with_aspect_ratio:12,get_slic:31,get_split_fold:50,get_statistical_metr:26,get_valid_ratio:19,get_vertices_3d:[33,40],get_vertices_3d_proj:[33,40],get_view:[3,20,24,25,28,32,33,34,35,36,37,38,39,40,41,42,43,44,46,47,50,51,52,53,54,56,57,58],get_with_format:[32,42],get_world_s:23,getitem:[1,2,3,4,5,6,8,11,13,37,44,46,47,50],getitem_ignore_error:1,giou3d:[33,40],giou3d_with:[33,40],giou:[21,26,29,30,32,33,41,49],giou_tensor:32,giou_with:32,giraff:47,github:34,give:[40,47,50,51,52],given:[1,3,11,12,13,15,16,20,25,26,28,29,32,37,41,42,43,50,54,56,57,58],glass:47,global:[7,17,33],global_color_set:[32,41],glove:47,go:4,goal:[53,54,56,57,58],gonna:47,good:24,gpu:[28,31,40,51,52],grab:[51,52],grad:[26,29],gradient:47,gradient_clip_v:[26,29,56,57],graph:[51,52],great:[44,46],grid:43,grid_siz:43,ground:[13,21,26,47,53,54,56,57,58],group:[2,5,6,26,29],gt:[15,16,21,26,29,40,46],gt_box:[15,16,29,53,54,56,57,58],gt_boxes_2d:[13,26],gt_boxes_3d:[13,40],gt_mask:[15,16,29,58],guid:58,h:[13,19,20,25,28,31,33,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53],h_crop:31,ha:[3,4,35,36,37,50,53,54,56,57,58],hacki:31,hair:47,hand:[50,58],handbag:47,handl:[3,13,39,47,48,50],happen:[21,26],hard:21,hardwar:[13,48],hasattr:45,have:[1,3,7,9,10,13,16,26,28,31,32,34,35,36,41,42,44,46,47,50,51,52,57,58],hb:25,head:[20,21,25,27,33,37,49,52],height:[12,19,20,25,28,32,37,41,42,50],hello:[1,4,8,11,12,49],help:[44,48],helper:[45,51,52],helpful:[20,25,28],here:[1,4,8,9,10,11,12,15,44,45,46,53],hflip:[31,35,36,40,46,47],hidden:25,hidden_dim:[17,20,25],high:[50,56,57,58],hint:[53,54,56],histogram:15,hold:26,home:[1,11,50,53,54],homogen:34,hook:[15,17,19,23,24],horizont:[31,32,33,35,36,40,46],hors:47,hot:[47,50],hour:[13,56,57,58],how:[3,20,21,25,26,28,40,47,49,53,54,56,57,58],howev:[1,20,29,50,53,56,57],html:45,http:[19,20,25,32,34,45,48,58],hue:12,huge:4,human:[0,44,49],humancrowd:44,hungarian:[21,26],hungarian_cost_class:[21,26],hungarian_cost_giou_box:26,hungarian_cost_l1_box:26,hungarina:26,hw:[40,43],hydrant:47,id:[2,3,13,15,25,26,29,38,39,47,50,53,54,56,57,58],id_sf_path:6,idea:47,ideal:[3,34],ident:[51,52],idstring2int:13,idx:[1,2,3,4,5,6,8,11,13,50],ie:25,ignor:[17,19,23,41],ignore_error:1,illustr:47,imag:[3,4,11,13,15,16,17,20,24,25,28,33,34,35,36,37,40,44,47,50,51,52,53,54],image_fold:50,image_path:[25,28,51,52],image_s:50,imagex:13,img:[51,52],img_ann_train_0:58,img_ann_train_1:58,img_ann_train_l:58,img_ann_val_0:58,img_ann_val_1:58,img_ann_val_m:58,img_fold:[3,4,45,50],img_test_0:50,img_test_1:50,img_test_n:50,img_train_0:[50,58],img_train_1:[50,58],img_train_l:[50,58],img_val_0:[50,58],img_val_1:[50,58],img_val_m:[50,58],implement:[1,15,25,27,28,29,43,44,48,51,52,57,58],impli:20,imposs:53,improv:[48,53],imshow:44,includ:[20,25,26,28,45],include_flow:6,include_vertice_behind_image_plan:33,increas:20,inde:42,independ:[13,48],index:[1,3,5,24,31,45,48],index_i:26,index_j:26,indic:[21,26,32],individu:[3,20,25,26,28,50],infer:[15,16,20,21,25,26,28,29,45,49],infernec:15,info:56,inform:[2,5,6,9,10,15,16,20,21,24,25,26,28,34,42,47,50,53,57,58],inherit:57,inhert:[15,17,24,29,31],init:[16,45],init_data_object:16,initi:[3,35,36,39,54,56,57,58],inlin:[53,54,56,57,58],inplac:[31,32],input:[19,21,23,24,25,26,28,39,44,50,51,52,53,56,57,58],input_dim:17,input_mean_std:[20,25,28],input_nam:[51,52],input_shap:[51,52],insid:[15,23],instal:[13,46,49],instanc:[17,19,23,33,37,39,46,47,56],instances_train2017:[24,45],instances_val2017:[3,24,45,50],instancesbasemetricscallback:15,instanti:[20,25,51,52,56,57,58],instead:[1,15,17,19,23,24,47],instruct:[2,5,6],int64:32,intanc:[15,29],integ:44,integr:48,intermedi:[25,51,52],intern:13,internet:50,interpret:[35,50],into_futur:6,intrins:[13,34,40],introduc:45,invalid:1,invari:26,invers:33,invert:35,io:[34,45],iou3d:[33,40],iou3d_with:[33,40],iou:[15,16,29,30,32,33,39,41,47,49],iou_tensor:32,iou_threshold:[16,32],iou_with:[32,39,47],is_avail:[16,53,54,56,57,58],is_dist_avail_and_initi:[15,23],is_empti:16,is_main_process:23,is_tru:16,isopen:53,issu:[3,47],isth:16,isthing_label:16,item:[1,7,13,50,56],iter:[3,7,20,24,35,36,44,52,53,54,56,57,58],its:[3,20,25,28,31,35,36,40,44,46,47,48,50,51,52,56,57,58],itself:39,join:[50,51,52],joiner:[19,23],joint:[11,28],jpeg:[13,37],jpg:[37,50,58],json:[1,3,4,11,13,24,45,49,53,58],just:[50,53,54],k:[26,29],keep:[16,35,42,43,44,46,47,53,54],kei:[13,15,20,25,26,28,29,31,50,51,52],kept:32,keyboard:47,keyword:15,kill:13,kite:47,knife:47,know:[28,45,46,47,53,54,56,57,58],knowledg:58,known:[35,47],ksimek:34,kwarg:[1,2,3,4,5,6,8,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,32,33,34,35,36,37,38,39,41,42,43,50,51,52],l1:[21,26],l:40,label:[3,6,9,10,11,13,16,20,21,24,25,26,28,29,30,31,32,33,35,36,37,39,40,41,42,43,46,49],label_nam:24,labels_nam:[16,38,47,50],labels_set:[32,39,41,42,44],labels_set_1:42,labels_set_2:42,lambda:[7,50],laptop:47,largest:12,last:[7,20,25,33,34,53,54],later:[1,35,36,45],latest:[26,29,45],latter:[17,19,23,46,47,48],layer:[17,19,20,21,23,25,26,28,29,53,54],learn:[25,26,42,46,48,50,53,54,56,57],least:[21,26,28,42],left:[6,7,9,10,11,12,13,35],left_offset:31,left_right:6,len:[13,15,20,21,26,32,46,50,53,54],length:34,less:42,let:[37,46,47,51,52,57,58],level:[19,49,50,56],level_start_index:19,librari:48,lidar_label:13,lig:57,light:[15,47],lightn:[15,24,26,29,45,48,51,52,53,54,56,57,58],lightningdatamodul:[24,45,54],lightningmodul:[15,26,54,56],ligntn:45,like:[42,50,56,57],limit:1,limit_train_batch:54,limit_val_batch:54,line:[3,13,35,36,40,47,56,57,58],link:[2,5,6,9,10,47,53,54],list:[1,3,6,7,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,38,42,43,44,46,48,50],listdir:50,lit_deform:[53,56],lit_detr:[54,56,57,58],lit_model:[51,52,56],lit_panopt:58,litdeformabledetr:[21,52,53,56],litdetr:[21,24,26,29,45,49,51,54,56,58],litpanopticdetr:[16,24,29,49],ll:47,ln:23,load:[2,9,10,13,15,20,24,25,26,28,29,35,45,49,50,53,54,55,56,57,58],load_gt:4,load_json_lin:4,load_train:[51,52,53,54,56,57,58],load_trt_plugins_for_deformable_detr:52,load_weight:[51,52],loader:[1,7,26,29,45,49],local:[33,53],locat:[31,37],log:[15,26,29],log_boxes_2d:15,log_boxes_3d:15,log_every_n_step:15,log_mask:15,logic:[51,52,58],logit:[20,21,25,26,28],longer:[4,13],look:[1,11,31,47],loop:[26,29,53,54],loss:[15,20,21,25,26,28,29],loss_box:26,loss_boxes_weight:[21,26,29],loss_ce_weight:[26,29],loss_dice_weight:29,loss_focal_weight:29,loss_giou_weight:[21,26,29],loss_label:[21,26],loss_label_weight:21,loss_valu:15,lot:4,lsap:26,lt:[40,46],m:[32,33,39],m_input:[51,52],m_output:[15,20,21,25,26,28,29],machin:[46,50],made:[17,20,25,28,44,50],mai:26,main:[4,8,9,10,13,17,25,28],mainli:48,make:[1,7,24,25,37,44,46,47,48,49],manag:[3,13,50],mandatori:[50,54],mani:50,manipul:3,manual:50,map:[13,16,17,19,23,27,35,36,37],map_metrics_callback:[15,29],market_2:11,market_5:11,market_6:11,mask2:39,mask2id:39,mask:[15,16,17,19,20,23,24,25,26,28,29,30,35,36,37,43,44,49,51,52,53,58],mask_float_tensor:39,mask_id:39,mask_tensor:39,maskth:28,match:[21,26,42,46,50,58],matcher:[18,22,29],matcher_fram:26,matplotlib:[35,36,40,46,51,52,53,54,56,57,58],matric:33,matrix:[13,32,34,39],matter:46,max:[1,43],max_epoch:[53,54],max_retry_on_error:1,max_siz:[12,44,45],maxim:25,maximum:12,mayb:31,mean:[20,26,29,37,44,46],mean_std:37,mean_std_norm:[37,46],memor:42,memori:[19,23,31,45,51,52,58],memory_format:31,memory_key_padding_mask:23,memory_mask:23,memory_padding_mask:19,merg:[0,13,44,49],merge_dataset:7,mergeabl:42,mergedataset:[0,49],meter:47,method:[1,4,11,12,13,15,17,20,24,25,26,28,29,32,37,42,43,44,45,46,47,51,52,57],metric:[14,21,26,29,49,56],metrics_callback:15,metricscallback:15,mhattent:23,microwav:47,middl:13,might:[15,29,37,42,44,47,50,51,52],min:26,min_siz:12,mind:[53,54],minim:[0,16,40,49,50],minimum:3,minmax_sym:37,minor:[51,52],misc:23,mistak:54,mlp:[14,25,49],mode:[33,40,50,58],model:[14,15,16,18,21,22,23,24,26,27,29,34,44,47,48,49,50,51,52,53,54,57,58],model_nam:[25,26,29,51,52,56,57,58],model_precis:[51,52],modif:23,modifi:[31,35,36,56,57,58],modul:[3,13,15,17,19,20,21,23,24,25,26,27,28,29,49,51,52,53,54,56,57],monitor:[21,26],more:[2,4,5,6,9,10,13,15,17,19,20,24,26,34,37,40,42,44,45,50,53,54,56,57,58],moreov:[56,57,58],most:[34,44,48,57],mot17:[0,44,49],mot_dataset:44,motorcycl:47,mountain_1:11,mous:47,move:15,mpi:[9,10,11,35],much:[21,26],multi:[0,17,20,25,31,37,48,49],multilabel:39,multipl:[3,7,32,39,41,42,44,48,50,57],multipli:39,multiscal:[19,52],must:[1,3,11,13,15,16,17,20,24,28,31,32,33,35,36,37,39,42,43,47,50,53],my_dataset:50,my_detr_model:58,my_fram:46,my_model:[56,57,58],my_norm:[37,46],my_set:37,n:[13,15,23,24,32,33,37,39,41,42,46,47],n_box:47,n_frame:12,n_head:[19,23],n_level:19,n_point:[19,42],n_tensor:31,naiv:[33,41],name:[1,3,11,13,15,19,20,23,24,25,26,28,29,31,32,33,35,36,37,38,39,40,41,42,43,44,46,47,50,53,56],name_x_axi:15,name_y_axi:15,namespac:[20,24,25,26,29,45,53,54,56,57,58],nan:23,nb:6,ndarrai:13,ne:31,nearli:[51,52],necessari:[1,16,35,50],necessit:[35,36],need:[1,3,13,15,17,19,20,23,24,31,37,42,46,47,48,50,51,52,53,54,58],neg:[21,26,31,33,35],network:[16,26,27,28,56,57,58],neural:27,neuron:25,new_fram:37,next:[1,3,7,24,31,35,36,53,54,56,57,58],ngc:48,nhead:[19,20,23,25],nice:[46,47],nm:32,nn:[17,19,20,23,25,26,28,29],no_augment:24,no_grad:[51,52],no_object:26,no_suffix:[53,54],non:[20,26,42],none:[1,3,4,6,7,8,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,37,38,39,41,42,43,45,46,47,51,52],norm01:[37,46],norm255:[37,46],norm:[23,46],norm_a:37,norm_minmax_sym:37,norm_resnet:[20,25,28,37,45,46,51,52,53],normal:[17,20,25,28,37,50],normalize_befor:[23,25],normalized_fram:37,normnal:37,note:[1,4,13,15,16,19,21,24,26,29,31,37,42,43,44,46,47,53,56],notebook:[51,52],noth:[13,38],notimplementederror:25,now:[1,3,47,50,51,52,54,56],np:[13,37,39,40,41,42,51,52],np_boxes3d:13,np_convert_waymo_to_aloception_coordinate_system:13,num_box:[21,26],num_channel:23,num_class:[20,21,25,26,28,29,51,52,53,54,56,57,58],num_decoder_lay:[19,23,25],num_enc:25,num_encoder_lay:[19,23,25],num_feature_level:[19,20],num_lay:[17,19,23],num_level:19,num_point:19,num_pos_feat:17,num_posit:16,num_process:13,num_queri:[20,21,25,26,28],num_target_box:26,num_work:[1,7,24],number:[1,7,9,10,12,16,20,21,24,25,26,28,29,37,42,51,52,53,54,58],nummer:24,numpi:[35,36,37,40,41,42,51,52],nvidia:48,object:[3,12,14,16,20,21,24,25,26,28,29,30,31,37,39,46,49,50,53,54,56,57,58],object_detector_callback:[15,26,29],objectdetectorcallback:[15,26,29],objectness_recal:26,objectness_true_po:26,objects_class:44,objects_id:44,occlus:[6,9,35,36,39],off:16,offer:50,offici:[23,51,52],offset:[34,35],offset_i:31,offset_x:31,old_path_directori:3,omit:26,on_list:31,on_train_batch_end:[15,26,29],on_train_epoch_end:15,on_validation_batch_end:15,on_validation_end:15,on_validation_epoch_end:[15,26,29],onc:[4,13,15,16,29,40,48,53,54,56,57,58],one:[1,3,4,11,13,15,16,17,19,20,23,24,26,28,29,32,35,36,37,39,40,42,44,45,46,47,50,53,54,57],ones:[35,57],oni:1,onli:[12,19,20,21,25,26,28,29,31,33,35,40,42,47,50,51,52,53,54],onlin:[57,58],onnx:[48,51,52],onnx_graphsurgeon:48,onnx_path:[51,52],onto:40,op:[23,48],open:[10,13,40,53],opencv:46,oper:[1,32,42,47],optic:[0,2,5,6,11,36,46,48,49],optim:[26,29],optimzi:21,optin:20,option:[1,3,13,16,20,21,24,25,26,28,29,32,33,35,36,37,38,39,41,42,43,45,49,50],orang:47,order:[3,15,25,26,28,32,33,41,43,50,51,52,53,54,56,57,58],org:[19,20,25,58],organ:50,orient:[33,40,41],oriented_boxes_2d:41,orientedboxes2d:[30,33,49],origin:[1,3,4,13,25,28,54],os:[50,51,52],other:[9,10,23,25,26,30,32,33,35,36,37,39,41,42,43,46,47,49,50,51,52,57,58],otherwis:[4,12,13,21,32,35,36,37,39,41,42,43],othwerwis:15,othwewis:37,ouput:[26,29],our:[45,47,51,52],out:[21,33,44,54],out_scor:[20,25],outptu:[20,25,26],output:[15,20,21,25,26,27,28,29,31,51,52,58],output_dim:17,outs_label:[20,25],outs_scor:[20,25],oven:47,over:[7,15,16,47],overridden:[17,19,23],overview:46,overwrit:[3,21],own:[20,25,46],p:[12,47],p_bbox:16,p_mask:16,p_t:21,packag:[12,13,45,47,48],pad:[20,25,28,31,32,42,43,44,51,52],pad_box:43,padd:[32,44],padded_boxes2d:32,padded_point:42,padding_mask:19,page:[1,4,8,11,12,50],paht:[50,53],pair:[13,21,26,41],pairwis:32,panopt:[14,15,16,29,39,49],panoptic_train2017:58,panoptic_val2017:58,panopticapmetricscallback:29,panoptichead:[28,58],panopticobjectdetectorcallback:29,paper:[2,5,6,9,10,17,20,25,28],param1:12,param2:12,param:[12,41],paramet:[1,3,5,6,7,9,10,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,37,38,39,41,42,43,46,48,50,56,57,58],params2namespac:[56,57],parent:[26,29,57],parent_pars:[21,24,26,29],park:47,pars:[2,13],parse_arg:[53,54,56,57,58],parser:[21,24,26,29,53,54,56,57,58],part:[2,5,6,42,45,50],partial:39,particular:[44,45],pascal2012:50,pascal:50,pass:[9,10,11,17,19,20,23,25,26,28,31,32,37,42,50],past:23,path:[1,3,4,11,13,25,26,29,35,36,37,39,50,51,52,53,56,57,58],path_to_imag:[20,25],pca:[33,41],pedestrian:13,peopl:3,people_mask:53,per:[16,28],percentag:[12,21,26,31],percentah:31,perceptron:[17,25],perform:[3,17,19,21,23,25,26,32,33,41,47,49,50],person:[3,47,50,52],person_box:47,perspect:34,pet:54,pfm:35,phone:47,pi:[40,41],pickl:13,pinhol:34,pip:48,pipelin:[46,48,49,56,57,58],pixel:[20,25,33,35,39,51,52],pizza:47,pkl:13,pl:[15,45,54],pl_helper:[51,52],pl_modul:[15,26,29],plai:[47,49],plan:33,plant:47,pleas:[2,3,5,6,13,45,48,57,58],plot:44,plt:[51,52,58],plugin:52,png:[35,39],png_negat:35,po:[19,23,47],point2d:42,point:[19,34,37,42,44,47],points2d:[30,37,49],points_2d:[37,42],points_2d_ab:42,points_2d_rel:42,points_2d_xi:42,points_2d_yx:42,points_format:42,popul:[13,33,40],popular:48,pos_emb:[19,23],pose:13,posit:[14,16,19,21,23,25,26,31,32,33,35,41,42,47,49],positin:25,position_embed:[19,23,25],position_encod:17,positionembeddingsin:17,possibl:[6,7,9,10,11,12,15,20,25,26,28,29,32,35,36,37,39,41,42,43,47,50,53,54],post:15,postprocess:[20,25,28],pot:47,potenti:43,pq:[14,49],pq_averag:16,pq_metrics_callback:15,pqmetric:[15,16],pqmetricscallback:15,pqstatcat:16,pre:[19,20,23,25,28,50,54,56,57,58],pre_process_tgt:[19,23],precis:[21,26,29,51,52,56],pred_box:[15,16,20,21,25,26,28,29,53,54,56,57,58],pred_logit:[20,21,25,26,28],pred_mask:[15,16,28,29,58],pred_masks_info:28,predefin:12,predict:[15,16,20,21,24,25,26,27,28,29,49,50,54,56,57,58],predictec:25,preds_box:15,preds_mask:15,prefer:47,prefix:15,preload:13,preloaded_calib:13,preloaded_labels_2d:13,preloaded_labels_3d:13,prepar:[0,1,4,49],prepr:13,preprocess:[20,24,25,28,53],present:[15,50,56],preserv:47,preserve_format:31,pretrain:[20,21,25,28,54,58],prevent:47,previou:[9,10,24,28,31,35,36,45,50,53,56,58],previous:58,princial:34,print:[13,16,35,40,41,44,46,47,53,54,56],print_bindings_info:[51,52],print_error:1,print_result:[16,56],procedur:[26,28],process:[13,15,16,19,21,23,24,26,29,49,50,53,54,56],produc:[23,54],program:47,progress:[15,21,26,29],project:[13,26,28,29,33,34,40,45,51,52,53,54,56,57,58],project_dir:[51,52],project_run_id:[51,52,53,54,56,57,58],promis:48,promiss:45,prompt:[3,50,53],properli:[1,11,15,37],properti:[1,16,43,44,46,47,50],propos:[19,26,44],proprli:43,provi:[32,42],provid:[2,5,6,20,25,32,41,42,44,45,47,50,56,57,58],psoition:17,pth:[20,25,28,55],pts2d:42,purpos:[3,48,50,53,54],purpus:28,push:16,py:[1,13,25,48],pylab:58,pypi:48,pyplot:[51,52],python:[13,48],pytorch:[1,3,15,24,26,28,29,31,37,45,47,48,50,51,52,53,54,56,57,58],pytorch_lightn:[15,24,26,29,45],qualit:[51,52],qualiti:[15,16],queri:[20,25,28],query_emb:[19,23],query_po:[19,23],quick:[3,44,46,50,51,52],quickli:[44,45,48],r11:34,r12:34,r13:34,r21:34,r22:34,r23:34,r31:34,r32:34,r33:34,r50:[16,18,22,26,28,29,51,52,53,54,56,57,58],r:[33,41,48,56],rabbit:[3,50],raccon:50,raccoon:50,racket:47,raft:45,rais:[1,16,20,24,25,26,28,29,39],rand:[39,50],random:[3,41,50,54,56,57,58],random_box:50,random_step:[8,13],randomcrop:[7,12],randomhorizontalflip:[12,44,45],randomli:13,randomresizewithaspectratio:[12,44,45],randomsampl:[1,7],randomselect:[12,44],randomsizecrop:[12,44],rang:[21,44],rank:37,rate:26,ratio:19,rcparam:[40,58],re:[42,46,47],reach:1,read:[3,6,13,20,25,28,35,36,50,53],read_sequ:13,readi:[39,44,48,57,58],readthedoc:45,real:[32,34,42],realisticnois:12,realli:[26,35,36],reason:[26,28,29,50,58],recal:[21,26],receiv:[15,28],recent:50,recip:[17,19,23],recognit:16,recommd:42,recommend:[25,50,56,57,58],recurs:[13,31],recursive_apply_on_labels_:31,reduc:[1,51,52],ref:48,refenr:[19,33],refer:[2,5,6,9,10,19,37,42,45],reference_point:19,refin:[18,52,53,56],reflect:26,refriger:47,regist:[17,19,23],regress:26,regular:47,rel:[3,20,21,25,26,28,32,35,41,42,47,50],rel_area:32,rel_po:[32,41,42],relat:[21,26,47],relationship:47,relative_to_absolut:43,releas:[31,53],relev:3,reli:[20,25,32,41,42,45,46],relu:[19,23],remain:[42,53],remot:47,remov:[1,23,42,47,56,57],remove_pad:[32,42],renam:31,rename_:31,rename_data_to_non:1,render:[3,20,24,25,28,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53,54,56,57,58],repeat:1,replac:[1,4,11,13,20,24,26,29,57],repositori:[23,51,52],repres:[13,20,25,28,35,36,39,42],represent:[28,32,42],requer:[24,28],requir:[3,13,24,25,28,37,46,48,50,52,53],rescal:43,research:48,reset_nam:31,resiz:[12,24,31,39,44,47,51,52],resnet50:[20,25],resnet:[19,20,23,25,28,37,46],resolv:[56,57,58],respect:[16,24,25,50,51,52,54],respons:47,rest:[24,57],restor:45,result:[15,26,51,52,53,54,56,57,58],resum:13,ret:53,ret_iou3d:[33,41],ret_iou:41,ret_union:[32,33,41],retri:1,retriev:[20,25,28],retry_offset:1,return_ann:39,return_bb_output:[20,25],return_cat:39,return_dec_output:[20,25,51],return_enc_output:[20,25],return_interm_lay:[19,20,23,25],return_intermedi:[19,23],return_intermediate_dec:[19,20,23],return_mask:[3,24],revers:35,rgb:[35,36,44,51,52],right:[1,6,7,9,11,13,33,35,41,44,46],right_offset:31,roboflow:50,robust:54,root:[3,13,50,51,52],root_dir:[51,52],rotat:[33,34,41],rotated_gi:41,rotated_giou_with:41,rotated_i:41,rotated_iou_with:41,rouhgli:21,rq:16,rqsrt:23,rtype:[1,13,20,21,33,39,41],run:[13,17,19,20,23,25,26,28,29,45,47,51,52,53,54,56,57,58],run_id:[53,54,55,56,57,58],run_id_dir:[51,52],run_train:[26,29,45,53,54,56,57,58],s:[13,15,34,37,46,47,48,50,51,52,57,58],sa_tensor:[32,43],said:[21,26],same:[7,16,21,24,31,32,33,34,39,41,42,43,44,51,52,53,54,58],same_on_fram:[12,24,45],same_on_sequ:[12,24,45],sampl:[0,1,2,7,12,13,15,16,19,24,26,29,35,36,37,40,45,46,47,49,50],sample_param:12,sampler:[1,7,24,45],sandwich:47,satur:12,save:[1,4,11,26,29,37,50,51,52,53,54,56,57,58],scalar:[15,21],scale:[17,20,44,45,48],scaler:15,scatter:15,scatter_nam:15,scene:[6,12],scheme:50,scienc:[2,5,6],scissor:47,score:[16,20,21,25,32,38],scratch:[50,56,57,58],script:[13,45,57],sdp:8,sdp_sequenc:8,seamless:47,second:[12,15,44,47,50],section:[2,5,6,40,44,50,56,57,58],see:[3,13,20,22,25,26,28,31,40,42,43,47,50,53,54,56,57,58],segment:[13,15,16,20,24,26,27,28,29,37,39,50,58],select:[3,13,20,25,26,28],selected_featur:6,selected_label:6,self:[1,4,12,13,19,20,23,33,45,50,51,52],send:31,sens:44,separ:15,sequenc:[2,5,8,9,10,13,15,24,26,29,40,44],sequence_data:[2,6],sequence_id:13,sequence_s:[7,9,10,13],sequence_skip:[9,10],sequenti:[7,19,23],serial:[51,52],serv:[51,52],set1:42,set2:42,set:[1,3,11,12,13,15,16,24,25,26,28,29,32,33,35,36,37,39,40,41,42,43,44,46,47,48,56],set_dataset_dir:[1,11],set_label:31,set_param:12,setup:[3,24,45,49,53,54],sever:[48,50,56,57,58],sh:48,shaman_2:11,shaman_3:11,shape:[13,16,19,20,21,25,28,33,34,40,41,44,46,47,51,52],sheep:47,shift:[1,31],shift_i:31,shift_x:31,shifted_tensor:31,shot_std:12,should:[3,13,17,19,23,32,34,35,36,42,47,48,50,57],show:[3,15,35,36,39,46,50,51,52,53,54],shown:54,shpae:25,shuffl:[1,7,44],side:[12,35],side_left:13,side_right:13,sigmoid:[20,21,29,52],sigmoid_focal_loss:21,sign:[13,30,47],silent:[17,19,23],similar:17,simpl:[17,33,41,46],simpli:[4,13,15,37,42,44,46],simplic:40,sin:25,sinc:[15,17,19,23,44,47,57],sine:25,singl:[3,15,25,50],sink:47,sintel:[0,35,49],sintel_disparity_dataset:9,sintel_flow_dataset:10,sintel_multi_dataset:11,sintel_sequ:11,sinteldisparitydataset:[0,35,49],sintelflowdataset:[0,49],sintelmultidataset:11,size:[1,7,9,10,12,13,15,16,20,21,24,25,26,28,29,31,32,33,37,39,40,41,42,43,44,46,47,50,51,52,53,56,57],skateboard:47,skew:34,ski:47,skip:[9,10,56,57,58],skipe:[4,13],sleeping_1:11,sleeping_2:11,slice:[31,40],slightli:[33,41,47],slot:[25,26],slot_true_neg:26,slow:4,slowest:[33,41],small:[3,46],smallest:[33,41],smilar:21,snapshot:53,snowboard:47,so:[12,46,47,50],softmax:[20,21,52,53],softmax_threshold:20,solv:26,some:[1,13,15,16,21,26,29,40,45,46,47,50,51,52,53,54,56,57,58],somehow:31,someon:46,someth:24,sort:[32,50],sourc:[10,13],space:[40,42],spatial:[30,31,44,46,48,49],spatial_augmented_tensor:[35,36,37,39,43],spatial_shap:19,spatial_shift:31,spatialaugmentedtensor:[35,36,37,39,42,43],spatialshift:12,specfic:[26,29],special:26,specif:[1,3,7,15,17,24,25,26,29,35,36,45,46,47,54,57],specifi:[51,52],speed:[13,50],split:[1,2,5,8,13,16,40,45,50,51,52],split_fold:[2,5,8,13,50],splitmixin:50,spoon:47,sport:47,sptial:43,sq:16,src:[19,23],src_key_padding_mask:23,src_level_start_index:19,src_mask:23,src_padding_mask:19,src_spatial_shap:19,src_valid_ratio:19,stack:23,stage:[21,24,25,26,29,45,46],standard:17,stanford:32,start:[15,31,37,41,53,54,57],stat:15,state:[32,47],stateful:49,statement:54,statist:[23,26],std:[37,46],step:[4,13,15,21,26,28,29,44,53],step_smooth:15,still:[21,26,46,48],stop:[13,47],storag:[1,53],store:[3,4,13,15,16,21,24,26,29,42,50,56,57,58],str:[1,11,13,15,20,21,24,25,26,28,29,32,33,35,36,37,38,39,41,42,43,45,50],straigthforward:[35,36],stream:[1,7,49],stream_load:[1,3,7,16,35,36,44,50],strength:40,strict:[25,28],strict_load_weight:[20,25,28],stride:20,string:[6,9,10,11,13],structur:[13,45,46,47,50],stuff:16,subclass:[17,19,23],subdirectori:13,subset:[0,49],suddenli:13,suit:45,suitcas:47,sum:19,summari:20,supervis:[21,26],support:[1,25,26,29,40,47],suppos:[20,25,28,42],sure:[44,46],surfboard:47,swap:35,synthet:[5,6],system:[13,33,48],t1:34,t2:34,t3:34,t:[1,6,13,15,26,29,31,33,37,40,42,44,45,46,47,51,52],t_bbox:16,t_co:[1,2,3,4,5,6,7,8,9,10,11,13],t_mask:16,tabl:[15,47],take:[13,17,19,23,26,29,46,47,50,51,52,53,57,58],taken:[3,28],tar:13,target:[12,16,21,26,31,37,42,45],target_fram:37,task:[15,26,28,29,50,53,58],teddi:47,tediou:[35,36],temperatur:17,temple_2:11,temple_3:11,tempor:[15,37,42,43,44],temporal_fram:43,temporarili:1,tenni:47,tensor:[1,13,15,17,19,20,21,23,25,26,28,32,33,34,35,36,37,38,39,40,41,42,44,45,46,48,49,50,51,52],tensorflow:50,tensorrt:[48,49,53],test2017:50,test:[1,8,13,24,44,48,50,51,52],test_experi:45,tfrecord:13,tgt:[19,23],tgt_box:[21,26],tgt_key_padding_mask:[19,23],tgt_mask:23,tgt_path:37,than:[7,23,26,31,37,42,48,58],thefor:[42,46],thei:53,theirs:[25,50],them:[15,17,19,23,24,29,48,50,57],therefor:[15,26,31,35,42,44,45,46,47,50,58],theta:41,thi:[1,2,3,4,5,6,7,9,10,11,12,13,15,16,17,19,21,23,24,25,26,28,29,31,32,33,34,35,36,37,40,41,42,43,44,45,47,48,50,51,52,53,54,56,57,58],thing:[16,37],think:50,third:50,those:[19,24,26,29,33,40],though:47,three:[45,50],threshold:[16,20,21,25,28,32],through:[1,4,26,31,40],throughout:[51,52],thu:[26,42],tie:47,time:[35,44,46,47,51,52,53],timestep:15,titl:[53,54,56,57,58],toaster:47,todo:[17,50],togeth:[7,47],toilet:47,tool:50,toolkit:48,toothbrush:47,top:[12,41,48,50],top_offset:31,torch2trt:[51,52],torch:[1,2,3,4,5,6,7,8,9,10,11,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,38,39,40,41,42,44,46,47,50,51,52,53,54,56,57,58],torch_m_output:[51,52],torch_model:[51,52],torch_pred_box:[51,52],torchvis:[12,20,23],total:[16,21,26,29],tp:16,track:[13,43,44,46,47],track_id:13,traffic:47,train2017:[24,45,50,58],train:[1,2,3,4,5,6,7,8,12,13,14,15,17,18,19,20,22,23,24,25,27,28,31,47,48,49,50,51,52],train_ann:[24,53],train_backbon:[19,20,23,25],train_dataload:[24,45],train_dataset:24,train_fold:[24,53],train_fram:24,train_load:[1,3,7,13,44,45,50],train_on_v:24,train_transform:[24,45],train_val:6,trainer:[15,24,26,29,49,57,58],trainig:45,training_epoch_end:15,training_sequ:8,training_step:[26,29],transform:[0,3,7,13,14,18,20,22,24,25,28,30,33,34,37,45,46,48,49,50,53,56,57],transform_fn:[1,7,44,50],transformer_outptu:[20,25],transformerdecod:[23,25],transformerdecoderlay:[23,25],transformerencod:23,transformerencoderlay:23,transforms1:12,transforms2:12,transit:47,translat:34,transorm:24,transpos:53,treat:[15,26],tree:13,trgat:26,tri:1,trt_export:[51,52],trt_m_output:[51,52],trt_model:[51,52],trt_pred_box:[51,52],trtexecutor:[51,52],truck:47,true_neg:26,truth:[13,21,26,47,53,54,56,57,58],tupl:[13,15,16,17,19,20,21,23,24,26,31,32,33,37,39,41,42],tutori:[3,50,51,52,53,54,56,57,58],tv:47,twice:37,two:[3,6,16,20,21,24,25,26,28,32,34,35,41,44,45,47,48,53,58],two_stag:19,two_stage_num_propos:19,txt:[13,48,50],type:[1,16,20,21,25,28,32,33,37,38,40,41,42,45,46,47,50],umbrella:47,un:26,uncom:[56,57,58],under:[4,56,57,58],union:[15,21,26,29,32,33,37,39,41,42],uniqu:[13,35],unit:35,univers:[2,5,6],unknown:13,unnam:[32,35,36,37,39,41,42,43],unnorm:[20,25,28],unpad:[19,42],unsign:35,unsqueez:47,until:[1,7],up:[26,29,47,50],updat:[16,24,26,44],update_data_object:16,update_loss_map:26,upload:[25,28],upward:13,url:48,us:[0,4,7,11,12,13,15,16,17,19,20,21,24,25,26,28,29,30,31,33,37,40,41,42,45,46,47,48,49,50,51,52,53,54,56,57,58],usag:[0,18,22,27,30,40,49],use_amp:29,useful:[21,26,37,42],user:[1,3,11,24,26,29,50,53],usuabl:23,usual:[1,3,15,35,36,44,45,47,53],util:[1,2,3,4,5,6,7,8,9,10,11,13,24,26,29,50],v2:25,v3:25,val2017:[3,24,45,50],val:[1,2,5,8,13,15,16,24,26,29,45,50],val_ann:[24,53],val_dataload:[15,24,45,53,54,56,57,58],val_dataset:24,val_fold:[24,53],val_fram:[15,24,29],val_nam:15,val_transform:[24,45],valid2017:58,valid:[3,13,15,24,26,29,44,45,50,51,52,53,54],valid_ratio:19,validation_sequ:[8,44],validation_step:[26,29],valu:[1,13,15,20,21,24,25,26,28,29,31,32,34,35,36,37,39,41,42,46,47,50,53,54,56,57],valueerror:[25,28],var_nam:31,variabl:[31,57],variou:17,vase:47,vb_folder:1,ve:46,vector:[34,36],vehicl:13,velodyn:13,veri:[17,44],verif:50,verifi:3,version:[4,17,47,48],vertic:[31,33,40,41],vertices_proj:40,vflip:31,via:[15,57],videocaptur:53,view:[32,33,35,36,38,39,40,41,42,43,44,46],visibl:44,vision:[45,48,49,50,53,54],visual:[30,33,49,51,52],volum:33,vs:21,w:[13,19,20,25,28,31,33,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53],w_crop:31,wa:[3,13,50,51,52,53],wai:[3,31,50,57,58],wait:[51,52],wandb:15,want:[15,20,25,28,29,42,44,51,52,57],warn:[3,46,47,54,58],waymo:[0,37,40,44,49],waymo_data_dir:13,waymo_dataset:[13,40],waymo_fram:44,waymo_prepar:13,waymodataset:[13,37,40,44],wb:25,we:[1,13,17,20,21,25,26,35,37,40,45,46,47,50,51,52,53,54,56,57,58],wear:53,websit:48,weight:[16,20,21,25,26,28,29,53,54,55,56,57,58],weight_path:[51,52],well:[15,16,21,26,29,40,46,50],were:[3,47,50],what:[46,47,51,52,53,54,56,57,58],when:[15,20,21,24,25,26,28,29,31,35,36,39,42],where:[23,26,32,37,39,42,51,52,53],whether:[26,42],which:[3,13,23,25,28,33,35,41,51,52,54],whit:25,whole:[1,46,51,52],width:[12,19,20,25,28,32,37,41,42,50],wil:[32,42],window:[15,35,36],wine:47,wip:15,wish:40,with_box_refin:20,with_pos_emb:[19,23],within:[17,19,23,43,45,46],without:[3,23,26,29,32,35,36,37,39,41,42,43,44,47,50,51,52,53],wll:38,word:[25,42],work:[1,11,15,17,20,33,40,43,50],worker:[1,7,24],workflow:[51,52],world:[34,49],worri:47,wors:[33,41],would:[47,50,53],wrapper:[19,23],write:[1,3,11],wrong:3,x0:[32,34],x1:[32,47],x2:[32,47],x:[13,15,17,20,23,25,28,29,31,32,33,34,35,36,37,38,39,40,41,42,43],x_center:[32,50],xc:[33,37,47],xcyc:[32,37,47,50],xy:42,xyxi:[32,33,47],xz:33,y0:[32,34],y1:[32,47],y2:[32,47],y:[13,15,31,33,40,41,42],y_center:[32,50],yc:[33,37,47],yet:[16,25,26,29,42,47],yield:[7,48],you:[17,20,24,25,28,31,35,36,37,40,42,44,46,47,48,51,52,53],your:[3,20,25,31,37,42,47,48,49,51,52,53,54],your_hom:13,your_path:13,your_project_nam:[51,52],your_run_id:[51,52],yourself:40,yx:42,yxyx:32,z:[13,33,40],zc:[33,37],zebra:47,zero:[35,36,37],zip:50,zone:19},titles:["Alodataset","Base dataset","ChairsSDHom","Coco detection","Crowd human","FlyingChairs2","Flyingthings3D Subset","Merging datasets","MOT17","Sintel Disparity","Sintel Optical Flow","Sintel Multi (Flow + Disparity)","Transformations","Waymo Dataset","Alonet","Callbacks","Metrics","Transformer","Deformable DETR","Architecture","Models","Training","Detr","Architecture","Detr: Data connectors","Models","Training","Detr panoptic","Models","Training","Aloscene","Augmented Tensor","BoundingBoxes2D","BoundingBoxes3D","Camera calibration","Disparity","Flow","Frame","Label","Mask","Visualization","OrientedBoxes2D","Points2D","Spatial Augmented Tensor","Alodataset: Loading your vision datasets","Alonet : Training your models","Aloscene: Computer vision with ease","About augmented tensors","Getting started","Aloception\u2019s documentation","How to setup your data?","DETR","Deformable DETR","Finetuning Deformanble DETR","Finetuning DETR","Exporting DETR / Deformable-DETR to TensorRT","Training Deformable DETR","Training Detr","Training Panoptic Head module"],titleterms:{"1":[50,53,54,56,57,58],"2":[50,53,54,56,57,58],"2d":50,"3":[50,53,56,57,58],"3d":40,"class":3,"default":3,"export":55,about:[47,48],alocept:[48,49],alodataset:[0,44],alonet:[14,45],aloscen:[30,46],also:[15,24],ap:[15,16],api:[2,3,5,6,7,9,10,13,18,22,27,28,33,35,36,39,41,49],architectur:[19,23],argument:57,attach:47,augment:[30,31,43,47],backbon:[19,23],base:[1,15,20,25,50],basedataset:50,basic:[3,7,13,20,25,28,33,35,36,39,41],boundingboxes2d:32,boundingboxes3d:33,box:[16,50],calibr:34,callback:[15,21,26,29],camera:[34,53],chairssdhom:2,chairssdhomdataset:2,chang:3,checkpoint:[51,52],coco:[3,50],cocodetection2detr:24,cocopanoptic2detr:[24,58],comparison:56,comput:46,config:50,connector:24,content:55,convent:35,creat:[35,36],criterion:[21,26],crowd:4,custom:50,data2detr:24,data:[24,45,50],databas:3,dataset:[1,3,7,13,44,45,50],declar:50,deform:[18,20,52,53,55,56],deformanbl:53,detect:3,detector:15,detr50:[53,54],detr:[18,20,22,24,25,27,51,52,53,54,55,56,57],detrr50:28,dimens:46,dispar:[9,11,35],document:49,download:3,eas:46,easi:[35,36],encod:17,exampl:[16,24],expect:16,explor:46,featur:40,file:50,filter:3,finetun:[20,25,28,53,54],flow:[10,11,36],flyingchairs2:5,flyingthings3d:6,flyingthings3dsubsetdataset:6,frame:[37,46,50],from:[51,52],get:48,gete:49,giou:40,goal:50,head:[28,58],hello:46,how:50,human:4,implement:50,infer:[51,52,53,54,56,57,58],instal:48,iou:40,json:50,label:[38,47,50],level:57,litdetr:57,litpanopticdetr:58,load:[44,51,52],loader:44,make:[50,53,54,56,57,58],mask:39,matcher:[21,26],merg:7,mergedataset:7,metric:[15,16],minim:3,mlp:17,model:[20,25,28,45,56],modul:[45,50,58],mot17:8,multi:11,next:50,normal:46,object:[15,35,36],obtain:16,optic:10,option:[48,53],orientedboxes2d:41,other:[40,48],packag:30,panopt:[27,28,58],perform:56,pipelin:45,plai:46,points2d:42,posit:17,pq:[15,16],predict:53,prepar:[13,50],process:[57,58],pth:[51,52],r50:[20,25],refer:30,refin:20,result:16,run_id:[51,52],s:49,sampl:[3,44],see:[15,24],set:50,setup:50,sign:35,sintel:[9,10,11],sinteldisparitydataset:9,sintelflowdataset:10,spatial:43,start:[48,49],stateful:47,stream:44,subset:6,summari:16,tensor:[30,31,43,47],tensorrt:[51,52,55],train:[21,26,29,44,45,53,54,56,57,58],trainer:56,transform:[12,17,19,23,35,36,40,44,47],turori:49,us:[3,35,36,39,44],usag:[7,13,20,25,28,33,41],vision:[44,46],visual:[35,36,40],waymo:13,weight:[51,52],what:50,world:46,your:[44,45,46,50]}})
+Search.setIndex({docnames:["alodataset/alodataset","alodataset/base_dataset","alodataset/chairssdhom_dataset","alodataset/coco_detection_dataset","alodataset/crowd_human_dataset","alodataset/flying_chairs2_dataset","alodataset/flyingthings3D_subset_dataset","alodataset/merge_dataset","alodataset/mot17","alodataset/sintel_disparity_dataset","alodataset/sintel_flow_dataset","alodataset/sintel_multi_dataset","alodataset/transform","alodataset/waymo_dataset","alonet/alonet","alonet/alonet.callbacks","alonet/alonet.metrics","alonet/alonet.transformers","alonet/deformable","alonet/deformable_architecture","alonet/deformable_models","alonet/deformable_training","alonet/detr","alonet/detr_architecture","alonet/detr_connectors","alonet/detr_models","alonet/detr_training","alonet/panoptic","alonet/panoptic_models","alonet/panoptic_training","aloscene/aloscene","aloscene/augmented_tensor","aloscene/bounding_boxes_2d","aloscene/bounding_boxes_3d","aloscene/camera_calib","aloscene/disparity","aloscene/flow","aloscene/frame","aloscene/labels","aloscene/mask","aloscene/notebooks/bounding_boxes_3d","aloscene/oriented_boxes_2d","aloscene/points2d","aloscene/spatial_augmented_tensor","getting_started/alodataset","getting_started/alonet","getting_started/aloscene","getting_started/augmented_tensor","getting_started/getting_started","index","tutorials/data_setup","tutorials/export_tensorrt/Tutorial - Export DETR to TensorRT and inference","tutorials/export_tensorrt/Tutorial - Export Deformable DETR to TensorRT and inference","tutorials/finetuning_deformable_detr","tutorials/finetuning_detr","tutorials/tensort_inference","tutorials/training_deformable_detr","tutorials/training_detr","tutorials/training_panoptic"],envversion:{"sphinx.domains.c":2,"sphinx.domains.changeset":1,"sphinx.domains.citation":1,"sphinx.domains.cpp":4,"sphinx.domains.index":1,"sphinx.domains.javascript":2,"sphinx.domains.math":2,"sphinx.domains.python":3,"sphinx.domains.rst":2,"sphinx.domains.std":2,nbsphinx:3,sphinx:56},filenames:["alodataset/alodataset.rst","alodataset/base_dataset.rst","alodataset/chairssdhom_dataset.rst","alodataset/coco_detection_dataset.rst","alodataset/crowd_human_dataset.rst","alodataset/flying_chairs2_dataset.rst","alodataset/flyingthings3D_subset_dataset.rst","alodataset/merge_dataset.rst","alodataset/mot17.rst","alodataset/sintel_disparity_dataset.rst","alodataset/sintel_flow_dataset.rst","alodataset/sintel_multi_dataset.rst","alodataset/transform.rst","alodataset/waymo_dataset.rst","alonet/alonet.rst","alonet/alonet.callbacks.rst","alonet/alonet.metrics.rst","alonet/alonet.transformers.rst","alonet/deformable.rst","alonet/deformable_architecture.rst","alonet/deformable_models.rst","alonet/deformable_training.rst","alonet/detr.rst","alonet/detr_architecture.rst","alonet/detr_connectors.rst","alonet/detr_models.rst","alonet/detr_training.rst","alonet/panoptic.rst","alonet/panoptic_models.rst","alonet/panoptic_training.rst","aloscene/aloscene.rst","aloscene/augmented_tensor.rst","aloscene/bounding_boxes_2d.rst","aloscene/bounding_boxes_3d.rst","aloscene/camera_calib.rst","aloscene/disparity.rst","aloscene/flow.rst","aloscene/frame.rst","aloscene/labels.rst","aloscene/mask.rst","aloscene/notebooks/bounding_boxes_3d.ipynb","aloscene/oriented_boxes_2d.rst","aloscene/points2d.rst","aloscene/spatial_augmented_tensor.rst","getting_started/alodataset.ipynb","getting_started/alonet.ipynb","getting_started/aloscene.ipynb","getting_started/augmented_tensor.ipynb","getting_started/getting_started.rst","index.rst","tutorials/data_setup.rst","tutorials/export_tensorrt/Tutorial - Export DETR to TensorRT and inference.ipynb","tutorials/export_tensorrt/Tutorial - Export Deformable DETR to TensorRT and inference.ipynb","tutorials/finetuning_deformable_detr.ipynb","tutorials/finetuning_detr.ipynb","tutorials/tensort_inference.rst","tutorials/training_deformable_detr.ipynb","tutorials/training_detr.ipynb","tutorials/training_panoptic.ipynb"],objects:{"alodataset.base_dataset":{BaseDataset:[1,1,1,""],Split:[1,1,1,""],rename_data_to_none:[1,5,1,""],stream_loader:[1,5,1,""],train_loader:[1,5,1,""]},"alodataset.base_dataset.BaseDataset":{download_sample:[1,2,1,""],get:[1,2,1,""],get_dataset_dir:[1,2,1,""],getitem:[1,2,1,""],getitem_ignore_errors:[1,2,1,""],prepare:[1,2,1,""],set_dataset_dir:[1,2,1,""],stream_loader:[1,2,1,""],train_loader:[1,2,1,""],vb_folder:[1,3,1,""]},"alodataset.base_dataset.Split":{TEST:[1,4,1,""],TRAIN:[1,4,1,""],Type:[1,4,1,""],VAL:[1,4,1,""]},"alodataset.chairssdhom_dataset":{ChairsSDHomDataset:[2,1,1,""]},"alodataset.chairssdhom_dataset.ChairsSDHomDataset":{SPLIT_FOLDERS:[2,4,1,""],get_file_ids:[2,2,1,""],get_frames:[2,2,1,""],get_sequences:[2,2,1,""],getitem:[2,2,1,""]},"alodataset.coco_detection_dataset":{CocoBaseDataset:[3,1,1,""]},"alodataset.coco_detection_dataset.CocoBaseDataset":{getitem:[3,2,1,""]},"alodataset.crowd_human_dataset":{CrowdHumanDataset:[4,1,1,""],main:[4,5,1,""]},"alodataset.crowd_human_dataset.CrowdHumanDataset":{getitem:[4,2,1,""],load_gt:[4,2,1,""],load_json_lines:[4,2,1,""],prepare:[4,2,1,""]},"alodataset.flying_chairs2_dataset":{FlyingChairs2Dataset:[5,1,1,""]},"alodataset.flying_chairs2_dataset.FlyingChairs2Dataset":{SPLIT_FOLDERS:[5,4,1,""],getitem:[5,2,1,""]},"alodataset.flyingthings3D_subset_dataset":{FlyingThings3DSubsetDataset:[6,1,1,""]},"alodataset.flyingthings3D_subset_dataset.FlyingThings3DSubsetDataset":{CAMERAS:[6,4,1,""],LABELS:[6,4,1,""],get_featlabels_element:[6,2,1,""],get_features_labels:[6,2,1,""],get_frames:[6,2,1,""],getitem:[6,2,1,""],id_sf_path:[6,2,1,""]},"alodataset.merge_dataset":{MergeDataset:[7,1,1,""]},"alodataset.merge_dataset.MergeDataset":{stream_loader:[7,2,1,""],train_loader:[7,2,1,""]},"alodataset.mot17":{Mot17:[8,1,1,""],main:[8,5,1,""]},"alodataset.mot17.Mot17":{DPM_SEQUENCES:[8,4,1,""],FRCNN_SEQUENCES:[8,4,1,""],SDP_SEQUENCES:[8,4,1,""],SEQUENCES:[8,4,1,""],SPLIT_FOLDERS:[8,4,1,""],getitem:[8,2,1,""]},"alodataset.sintel_disparity_dataset":{SintelDisparityDataset:[9,1,1,""]},"alodataset.sintel_disparity_dataset.SintelDisparityDataset":{CAMERAS:[9,4,1,""],LABELS:[9,4,1,""],PASSES:[9,4,1,""]},"alodataset.sintel_flow_dataset":{SintelFlowDataset:[10,1,1,""]},"alodataset.sintel_flow_dataset.SintelFlowDataset":{CAMERAS:[10,4,1,""],LABELS:[10,4,1,""],PASSES:[10,4,1,""]},"alodataset.sintel_multi_dataset":{SintelMultiDataset:[11,1,1,""]},"alodataset.sintel_multi_dataset.SintelMultiDataset":{CAMERAS:[11,4,1,""],DISP_LABELS:[11,4,1,""],FLOW_LABELS:[11,4,1,""],LABELS:[11,4,1,""],PASSES:[11,4,1,""],SINTEL_SEQUENCES:[11,4,1,""],get_dataset_dir:[11,2,1,""],getitem:[11,2,1,""],set_dataset_dir:[11,2,1,""]},"alodataset.transforms":{AloTransform:[12,1,1,""],ColorJitter:[12,1,1,""],Compose:[12,1,1,""],CustomRandomColoring:[12,1,1,""],RandomCrop:[12,1,1,""],RandomHorizontalFlip:[12,1,1,""],RandomResizeWithAspectRatio:[12,1,1,""],RandomSelect:[12,1,1,""],RandomSizeCrop:[12,1,1,""],RealisticNoise:[12,1,1,""],Resize:[12,1,1,""],SpatialShift:[12,1,1,""]},"alodataset.transforms.AloTransform":{sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.ColorJitter":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""],training:[12,4,1,""]},"alodataset.transforms.Compose":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.CustomRandomColoring":{apply:[12,2,1,""],sample_params:[12,2,1,""]},"alodataset.transforms.RandomCrop":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomHorizontalFlip":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomResizeWithAspectRatio":{apply:[12,2,1,""],get_size_with_aspect_ratio:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomSelect":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RandomSizeCrop":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.RealisticNoise":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.Resize":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.transforms.SpatialShift":{apply:[12,2,1,""],sample_params:[12,2,1,""],set_params:[12,2,1,""]},"alodataset.waymo_dataset":{WaymoDataset:[13,1,1,""],main:[13,5,1,""]},"alodataset.waymo_dataset.WaymoDataset":{CAMERAS:[13,4,1,""],CLASSES:[13,4,1,""],LABELS:[13,4,1,""],SPLIT_FOLDERS:[13,4,1,""],__init__:[13,2,1,""],get_frame_boxes2d:[13,2,1,""],get_frame_boxes3d:[13,2,1,""],get_frame_camera_parameters:[13,2,1,""],get_frames:[13,2,1,""],get_segments:[13,2,1,""],getitem:[13,2,1,""],np_convert_waymo_to_aloception_coordinate_system:[13,2,1,""],prepare:[13,2,1,""],read_sequences:[13,2,1,""]},"alonet.callbacks":{base_metrics_callback:[15,0,0,"-"],map_metrics_callback:[15,0,0,"-"],metrics_callback:[15,0,0,"-"],object_detector_callback:[15,0,0,"-"],pq_metrics_callback:[15,0,0,"-"]},"alonet.callbacks.base_metrics_callback":{InstancesBaseMetricsCallback:[15,1,1,""]},"alonet.callbacks.base_metrics_callback.InstancesBaseMetricsCallback":{add_sample:[15,2,1,""],inference:[15,2,1,""],on_validation_batch_end:[15,2,1,""],on_validation_end:[15,2,1,""]},"alonet.callbacks.map_metrics_callback":{ApMetricsCallback:[15,1,1,""]},"alonet.callbacks.map_metrics_callback.ApMetricsCallback":{on_validation_end:[15,2,1,""]},"alonet.callbacks.metrics_callback":{MetricsCallback:[15,1,1,""],get_rank:[15,5,1,""],is_dist_avail_and_initialized:[15,5,1,""]},"alonet.callbacks.metrics_callback.MetricsCallback":{on_train_batch_end:[15,2,1,""],on_train_epoch_end:[15,2,1,""],on_validation_batch_end:[15,2,1,""],on_validation_epoch_end:[15,2,1,""]},"alonet.callbacks.object_detector_callback":{ObjectDetectorCallback:[15,1,1,""]},"alonet.callbacks.object_detector_callback.ObjectDetectorCallback":{log_boxes_2d:[15,2,1,""],log_boxes_3d:[15,2,1,""],log_masks:[15,2,1,""],on_train_batch_end:[15,2,1,""],on_validation_epoch_end:[15,2,1,""]},"alonet.callbacks.pq_metrics_callback":{PQMetricsCallback:[15,1,1,""]},"alonet.callbacks.pq_metrics_callback.PQMetricsCallback":{on_validation_end:[15,2,1,""]},"alonet.deformable_detr":{backbone:[19,0,0,"-"],callbacks:[21,0,0,"-"],criterion:[21,0,0,"-"],deformable_detr:[20,0,0,"-"],deformable_detr_r50:[20,0,0,"-"],deformable_detr_r50_finetune:[20,0,0,"-"],deformable_detr_r50_refinement:[20,0,0,"-"],deformable_transformer:[19,0,0,"-"],matcher:[21,0,0,"-"],train:[21,0,0,"-"]},"alonet.deformable_detr.backbone":{Backbone:[19,1,1,""],BackboneBase:[19,1,1,""],Joiner:[19,1,1,""]},"alonet.deformable_detr.backbone.Backbone":{training:[19,4,1,""]},"alonet.deformable_detr.backbone.BackboneBase":{training:[19,4,1,""]},"alonet.deformable_detr.backbone.Joiner":{training:[19,4,1,""]},"alonet.deformable_detr.criterion":{DeformableCriterion:[21,1,1,""],sigmoid_focal_loss:[21,5,1,""]},"alonet.deformable_detr.criterion.DeformableCriterion":{get_metrics:[21,2,1,""],loss_labels:[21,2,1,""],training:[21,4,1,""]},"alonet.deformable_detr.deformable_detr":{DeformableDETR:[20,1,1,""],build_deformable_detr_r50:[20,5,1,""]},"alonet.deformable_detr.deformable_detr.DeformableDETR":{INPUT_MEAN_STD:[20,4,1,""],build_backbone:[20,2,1,""],build_decoder:[20,2,1,""],build_decoder_layer:[20,2,1,""],build_positional_encoding:[20,2,1,""],build_transformer:[20,2,1,""],forward:[20,2,1,""],forward_class_heads:[20,2,1,""],forward_heads:[20,2,1,""],forward_position_heads:[20,2,1,""],get_outs_filter:[20,2,1,""],get_outs_labels:[20,2,1,""],inference:[20,2,1,""],training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50":{DeformableDetrR50:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50.DeformableDetrR50":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune":{DeformableDetrR50Finetune:[20,1,1,""],DeformableDetrR50RefinementFinetune:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune.DeformableDetrR50Finetune":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_finetune.DeformableDetrR50RefinementFinetune":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_detr_r50_refinement":{DeformableDetrR50Refinement:[20,1,1,""]},"alonet.deformable_detr.deformable_detr_r50_refinement.DeformableDetrR50Refinement":{training:[20,4,1,""]},"alonet.deformable_detr.deformable_transformer":{DeformableTransformer:[19,1,1,""],DeformableTransformerDecoder:[19,1,1,""],DeformableTransformerDecoderLayer:[19,1,1,""],DeformableTransformerEncoder:[19,1,1,""],DeformableTransformerEncoderLayer:[19,1,1,""],build_deforamble_transformer:[19,5,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformer":{forward:[19,2,1,""],gen_encoder_output_proposals:[19,2,1,""],get_proposal_pos_embed:[19,2,1,""],get_valid_ratio:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerDecoder":{decoder_forward:[19,2,1,""],forward:[19,2,1,""],pre_process_tgt:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerDecoderLayer":{decoder_layer_forward:[19,2,1,""],forward:[19,2,1,""],forward_ffn:[19,2,1,""],pre_process_tgt:[19,2,1,""],training:[19,4,1,""],with_pos_embed:[19,2,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerEncoder":{forward:[19,2,1,""],get_reference_points:[19,2,1,""],training:[19,4,1,""]},"alonet.deformable_detr.deformable_transformer.DeformableTransformerEncoderLayer":{forward:[19,2,1,""],forward_ffn:[19,2,1,""],training:[19,4,1,""],with_pos_embed:[19,2,1,""]},"alonet.deformable_detr.matcher":{DeformableDetrHungarianMatcher:[21,1,1,""]},"alonet.deformable_detr.matcher.DeformableDetrHungarianMatcher":{hungarian_cost_class:[21,2,1,""],training:[21,4,1,""]},"alonet.deformable_detr.train":{LitDeformableDetr:[21,1,1,""]},"alonet.deformable_detr.train.LitDeformableDetr":{add_argparse_args:[21,2,1,""],build_criterion:[21,2,1,""],build_matcher:[21,2,1,""],build_model:[21,2,1,""],configure_optimizers:[21,2,1,""],training:[21,4,1,""]},"alonet.detr":{backbone:[23,0,0,"-"],callbacks:[26,0,0,"-"],criterion:[26,0,0,"-"],detr:[25,0,0,"-"],detr_r50:[25,0,0,"-"],detr_r50_finetune:[25,0,0,"-"],matcher:[26,0,0,"-"],train:[26,0,0,"-"],transformer:[23,0,0,"-"]},"alonet.detr.backbone":{Backbone:[23,1,1,""],BackboneBase:[23,1,1,""],FrozenBatchNorm2d:[23,1,1,""],Joiner:[23,1,1,""],build_backbone:[23,5,1,""],get_rank:[23,5,1,""],get_world_size:[23,5,1,""],is_dist_avail_and_initialized:[23,5,1,""],is_main_process:[23,5,1,""]},"alonet.detr.backbone.Backbone":{training:[23,4,1,""]},"alonet.detr.backbone.BackboneBase":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.backbone.FrozenBatchNorm2d":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.backbone.Joiner":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.callbacks":{DetrObjectDetectorCallback:[26,1,1,""]},"alonet.detr.callbacks.DetrObjectDetectorCallback":{on_train_batch_end:[26,2,1,""],on_validation_epoch_end:[26,2,1,""]},"alonet.detr.criterion":{DetrCriterion:[26,1,1,""]},"alonet.detr.criterion.DetrCriterion":{forward:[26,2,1,""],get_loss:[26,2,1,""],get_metrics:[26,2,1,""],get_statistical_metrics:[26,2,1,""],loss_boxes:[26,2,1,""],loss_labels:[26,2,1,""],training:[26,4,1,""]},"alonet.detr.data_modules":{coco_detection2detr:[24,0,0,"-"],coco_panoptic2detr:[24,0,0,"-"],data2detr:[24,0,0,"-"]},"alonet.detr.data_modules.coco_detection2detr":{CocoDetection2Detr:[24,1,1,""]},"alonet.detr.data_modules.coco_detection2detr.CocoDetection2Detr":{setup:[24,2,1,""]},"alonet.detr.data_modules.coco_panoptic2detr":{CocoPanoptic2Detr:[24,1,1,""]},"alonet.detr.data_modules.coco_panoptic2detr.CocoPanoptic2Detr":{setup:[24,2,1,""]},"alonet.detr.data_modules.data2detr":{Data2Detr:[24,1,1,""]},"alonet.detr.data_modules.data2detr.Data2Detr":{add_argparse_args:[24,2,1,""],setup:[24,2,1,""],train_dataloader:[24,2,1,""],train_transform:[24,2,1,""],val_dataloader:[24,2,1,""],val_transform:[24,2,1,""]},"alonet.detr.detr":{Detr:[25,1,1,""]},"alonet.detr.detr.Detr":{INPUT_MEAN_STD:[25,4,1,""],build_backbone:[25,2,1,""],build_bbox_embed:[25,2,1,""],build_class_embed:[25,2,1,""],build_decoder:[25,2,1,""],build_decoder_layer:[25,2,1,""],build_positional_encoding:[25,2,1,""],build_transformer:[25,2,1,""],forward:[25,2,1,""],forward_class_heads:[25,2,1,""],forward_heads:[25,2,1,""],forward_position_heads:[25,2,1,""],get_outs_filter:[25,2,1,""],get_outs_labels:[25,2,1,""],inference:[25,2,1,""],training:[25,4,1,""]},"alonet.detr.detr_r50":{DetrR50:[25,1,1,""],main:[25,5,1,""]},"alonet.detr.detr_r50.DetrR50":{training:[25,4,1,""]},"alonet.detr.detr_r50_finetune":{DetrR50Finetune:[25,1,1,""]},"alonet.detr.detr_r50_finetune.DetrR50Finetune":{training:[25,4,1,""]},"alonet.detr.matcher":{DetrHungarianMatcher:[26,1,1,""],build_matcher:[26,5,1,""]},"alonet.detr.matcher.DetrHungarianMatcher":{forward:[26,2,1,""],hungarian:[26,2,1,""],hungarian_cost_class:[26,2,1,""],hungarian_cost_giou_boxes:[26,2,1,""],hungarian_cost_l1_boxes:[26,2,1,""],training:[26,4,1,""]},"alonet.detr.train":{LitDetr:[26,1,1,""]},"alonet.detr.train.LitDetr":{add_argparse_args:[26,2,1,""],assert_input:[26,2,1,""],build_criterion:[26,2,1,""],build_matcher:[26,2,1,""],build_model:[26,2,1,""],callbacks:[26,2,1,""],configure_optimizers:[26,2,1,""],forward:[26,2,1,""],inference:[26,2,1,""],run_train:[26,2,1,""],training:[26,4,1,""],training_step:[26,2,1,""],validation_step:[26,2,1,""]},"alonet.detr.transformer":{Transformer:[23,1,1,""],TransformerDecoder:[23,1,1,""],TransformerDecoderLayer:[23,1,1,""],TransformerEncoder:[23,1,1,""],TransformerEncoderLayer:[23,1,1,""],build_transformer:[23,5,1,""]},"alonet.detr.transformer.Transformer":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerDecoder":{decoder_forward:[23,2,1,""],forward:[23,2,1,""],pre_process_tgt:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerDecoderLayer":{decoder_layer_forward:[23,2,1,""],forward:[23,2,1,""],forward_post:[23,2,1,""],forward_pre:[23,2,1,""],pre_process_tgt:[23,2,1,""],training:[23,4,1,""],with_pos_embed:[23,2,1,""]},"alonet.detr.transformer.TransformerEncoder":{forward:[23,2,1,""],training:[23,4,1,""]},"alonet.detr.transformer.TransformerEncoderLayer":{forward:[23,2,1,""],forward_post:[23,2,1,""],forward_pre:[23,2,1,""],training:[23,4,1,""],with_pos_embed:[23,2,1,""]},"alonet.detr_panoptic":{callbacks:[29,0,0,"-"],detr_panoptic:[28,0,0,"-"],detr_r50_panoptic_finetune:[28,0,0,"-"],train:[29,0,0,"-"]},"alonet.detr_panoptic.callbacks":{PanopticApMetricsCallbacks:[29,1,1,""],PanopticObjectDetectorCallback:[29,1,1,""]},"alonet.detr_panoptic.callbacks.PanopticApMetricsCallbacks":{add_sample:[29,2,1,""]},"alonet.detr_panoptic.callbacks.PanopticObjectDetectorCallback":{on_train_batch_end:[29,2,1,""],on_validation_epoch_end:[29,2,1,""]},"alonet.detr_panoptic.detr_panoptic":{PanopticHead:[28,1,1,""],main:[28,5,1,""]},"alonet.detr_panoptic.detr_panoptic.PanopticHead":{INPUT_MEAN_STD:[28,4,1,""],forward:[28,2,1,""],inference:[28,2,1,""],training:[28,4,1,""]},"alonet.detr_panoptic.detr_r50_panoptic_finetune":{DetrR50PanopticFinetune:[28,1,1,""]},"alonet.detr_panoptic.detr_r50_panoptic_finetune.DetrR50PanopticFinetune":{training:[28,4,1,""]},"alonet.detr_panoptic.train":{LitPanopticDetr:[29,1,1,""]},"alonet.detr_panoptic.train.LitPanopticDetr":{add_argparse_args:[29,2,1,""],build_criterion:[29,2,1,""],build_model:[29,2,1,""],callbacks:[29,2,1,""],precision:[29,4,1,""],run_train:[29,2,1,""],training:[29,4,1,""],training_step:[29,2,1,""],use_amp:[29,4,1,""],validation_step:[29,2,1,""]},"alonet.metrics":{compute_map:[16,0,0,"-"],compute_pq:[16,0,0,"-"]},"alonet.metrics.compute_map":{APDataObject:[16,1,1,""],ApMetrics:[16,1,1,""]},"alonet.metrics.compute_map.APDataObject":{add_gt_positives:[16,2,1,""],get_metrics:[16,2,1,""],is_empty:[16,2,1,""],push:[16,2,1,""]},"alonet.metrics.compute_map.ApMetrics":{add_sample:[16,2,1,""],calc_map:[16,2,1,""],init_data_objects:[16,2,1,""]},"alonet.metrics.compute_pq":{PQMetrics:[16,1,1,""],PQStatCat:[16,1,1,""]},"alonet.metrics.compute_pq.PQMetrics":{add_sample:[16,2,1,""],calc_map:[16,2,1,""],pq_average:[16,2,1,""],update_data_objects:[16,2,1,""]},"alonet.transformers":{mlp:[17,0,0,"-"],position_encoding:[17,0,0,"-"]},"alonet.transformers.mlp":{MLP:[17,1,1,""]},"alonet.transformers.mlp.MLP":{forward:[17,2,1,""],training:[17,4,1,""]},"alonet.transformers.position_encoding":{PositionEmbeddingSine:[17,1,1,""],build_position_encoding:[17,5,1,""]},"alonet.transformers.position_encoding.PositionEmbeddingSine":{forward:[17,2,1,""],training:[17,4,1,""]},"aloscene.bounding_boxes_2d":{BoundingBoxes2D:[32,1,1,""]},"aloscene.bounding_boxes_2d.BoundingBoxes2D":{FORMATS:[32,4,1,""],GLOBAL_COLOR_SET:[32,4,1,""],abs_area:[32,2,1,""],abs_pos:[32,2,1,""],append_labels:[32,2,1,""],area:[32,2,1,""],as_boxes:[32,2,1,""],boxes2abspos:[32,2,1,""],boxes2relpos:[32,2,1,""],boxes2xcyc:[32,2,1,""],boxes2xyxy:[32,2,1,""],boxes2yxyx:[32,2,1,""],boxes_hflip:[32,2,1,""],fit_to_padded_size:[32,2,1,""],get_view:[32,2,1,""],get_with_format:[32,2,1,""],giou:[32,2,1,""],giou_with:[32,2,1,""],iou:[32,2,1,""],iou_with:[32,2,1,""],nms:[32,2,1,""],rel_area:[32,2,1,""],rel_pos:[32,2,1,""],remove_padding:[32,2,1,""],xcyc:[32,2,1,""],xyxy:[32,2,1,""],yxyx:[32,2,1,""]},"aloscene.bounding_boxes_3d":{BoundingBoxes3D:[33,1,1,""]},"aloscene.bounding_boxes_3d.BoundingBoxes3D":{bev_boxes:[33,2,1,""],boxes_3d_hflip:[33,2,1,""],get_enclosing_box_2d:[33,2,1,""],get_vertices_3d:[33,2,1,""],get_vertices_3d_proj:[33,2,1,""],get_view:[33,2,1,""],giou3d:[33,2,1,""],giou3d_with:[33,2,1,""],iou3d:[33,2,1,""],iou3d_with:[33,2,1,""]},"aloscene.camera_calib":{CameraExtrinsic:[34,1,1,""],CameraIntrinsic:[34,1,1,""]},"aloscene.camera_calib.CameraExtrinsic":{get_view:[34,2,1,""]},"aloscene.camera_calib.CameraIntrinsic":{get_view:[34,2,1,""]},"aloscene.disparity":{Disparity:[35,1,1,""]},"aloscene.disparity.Disparity":{append_occlusion:[35,2,1,""],signed:[35,2,1,""],unsigned:[35,2,1,""]},"aloscene.flow":{Flow:[36,1,1,""]},"aloscene.flow.Flow":{append_occlusion:[36,2,1,""]},"aloscene.frame":{Frame:[37,1,1,""]},"aloscene.frame.Frame":{append_boxes2d:[37,2,1,""],append_boxes3d:[37,2,1,""],append_disparity:[37,2,1,""],append_flow:[37,2,1,""],append_labels:[37,2,1,""],append_points2d:[37,2,1,""],append_segmentation:[37,2,1,""],mean_std_norm:[37,2,1,""],norm01:[37,2,1,""],norm255:[37,2,1,""],norm_as:[37,2,1,""],norm_minmax_sym:[37,2,1,""],norm_resnet:[37,2,1,""],save:[37,2,1,""]},"aloscene.labels":{Labels:[38,1,1,""]},"aloscene.labels.Labels":{get_view:[38,2,1,""]},"aloscene.mask":{Mask:[39,1,1,""]},"aloscene.mask.Mask":{append_labels:[39,2,1,""],get_view:[39,2,1,""],iou_with:[39,2,1,""],mask2id:[39,2,1,""]},"aloscene.oriented_boxes_2d":{OrientedBoxes2D:[41,1,1,""]},"aloscene.oriented_boxes_2d.OrientedBoxes2D":{GLOBAL_COLOR_SET:[41,4,1,""],abs_pos:[41,2,1,""],append_labels:[41,2,1,""],boxes2abspos:[41,2,1,""],boxes2relpos:[41,2,1,""],corners:[41,2,1,""],get_view:[41,2,1,""],rel_pos:[41,2,1,""],rotated_giou:[41,2,1,""],rotated_giou_with:[41,2,1,""],rotated_iou:[41,2,1,""],rotated_iou_with:[41,2,1,""]},"aloscene.points_2d":{Points2D:[42,1,1,""]},"aloscene.points_2d.Points2D":{FORMATS:[42,4,1,""],abs_pos:[42,2,1,""],append_labels:[42,2,1,""],as_points:[42,2,1,""],fit_to_padded_size:[42,2,1,""],get_view:[42,2,1,""],get_with_format:[42,2,1,""],rel_pos:[42,2,1,""],remove_padding:[42,2,1,""],xy:[42,2,1,""],yx:[42,2,1,""]},"aloscene.tensors":{augmented_tensor:[31,0,0,"-"],spatial_augmented_tensor:[43,0,0,"-"]},"aloscene.tensors.augmented_tensor":{AugmentedTensor:[31,1,1,""]},"aloscene.tensors.augmented_tensor.AugmentedTensor":{COMMON_DIM_NAMES:[31,4,1,""],add_label:[31,2,1,""],add_property:[31,2,1,""],apply_on_label:[31,2,1,""],as_tensor:[31,2,1,""],clone:[31,2,1,""],cpu:[31,2,1,""],crop:[31,2,1,""],cuda:[31,2,1,""],drop_labels:[31,2,1,""],get_labels:[31,2,1,""],get_slices:[31,2,1,""],hflip:[31,2,1,""],pad:[31,2,1,""],recursive_apply_on_labels_:[31,2,1,""],rename:[31,2,1,""],rename_:[31,2,1,""],reset_names:[31,2,1,""],resize:[31,2,1,""],set_labels:[31,2,1,""],spatial_shift:[31,2,1,""],to:[31,2,1,""],vflip:[31,2,1,""]},"aloscene.tensors.spatial_augmented_tensor":{SpatialAugmentedTensor:[43,1,1,""]},"aloscene.tensors.spatial_augmented_tensor.SpatialAugmentedTensor":{H:[43,3,1,""],HW:[43,3,1,""],W:[43,3,1,""],append_cam_extrinsic:[43,2,1,""],append_cam_intrinsic:[43,2,1,""],append_mask:[43,2,1,""],batch:[43,2,1,""],batch_list:[43,2,1,""],get_view:[43,2,1,""],relative_to_absolute:[43,2,1,""],temporal:[43,2,1,""]},alodataset:{base_dataset:[1,0,0,"-"],chairssdhom_dataset:[2,0,0,"-"],coco_detection_dataset:[3,0,0,"-"],crowd_human_dataset:[4,0,0,"-"],flying_chairs2_dataset:[5,0,0,"-"],flyingthings3D_subset_dataset:[6,0,0,"-"],merge_dataset:[7,0,0,"-"],mot17:[8,0,0,"-"],sintel_disparity_dataset:[9,0,0,"-"],sintel_flow_dataset:[10,0,0,"-"],sintel_multi_dataset:[11,0,0,"-"],transforms:[12,0,0,"-"],waymo_dataset:[13,0,0,"-"]},alonet:{transformers:[17,0,0,"-"]},aloscene:{bounding_boxes_2d:[32,0,0,"-"],bounding_boxes_3d:[33,0,0,"-"],camera_calib:[34,0,0,"-"],disparity:[35,0,0,"-"],flow:[36,0,0,"-"],frame:[37,0,0,"-"],labels:[38,0,0,"-"],mask:[39,0,0,"-"],oriented_boxes_2d:[41,0,0,"-"],points_2d:[42,0,0,"-"]}},objnames:{"0":["py","module","Python module"],"1":["py","class","Python class"],"2":["py","method","Python method"],"3":["py","property","Python property"],"4":["py","attribute","Python attribute"],"5":["py","function","Python function"]},objtypes:{"0":"py:module","1":"py:class","2":"py:method","3":"py:property","4":"py:attribute","5":"py:function"},terms:{"0":[3,6,12,16,19,20,21,23,25,26,28,29,31,32,34,36,37,39,40,41,42,44,45,46,47,48,50,51,52,53,54,56,57,58],"0000":47,"00051602e":41,"00070727e":41,"00119355e":32,"00133255e":32,"00159851e":32,"00204472e":32,"00215137e":32,"00252999e":32,"00265454e":41,"00416320e":41,"00508215e":32,"00565671e":41,"00572965e":32,"00630891e":32,"00671225e":41,"00819157e":32,"01":[32,37,41,46,50],"01023613e":41,"01064562e":41,"01078656e":41,"01085460e":41,"01117099e":41,"01284568e":32,"01389831e":41,"01412360e":32,"01414503e":32,"01447680e":41,"01489084e":32,"01573518e":32,"0164":37,"01693818e":32,"01715315e":41,"01773214e":41,"01826448e":41,"02":[8,12,32,41],"02008869e":32,"02071244e":41,"02099545e":41,"02148027e":41,"02181674e":41,"02201126e":32,"02216669e":41,"02224198e":41,"02320372e":41,"02340613e":32,"02524928e":41,"02542744e":41,"02609146e":41,"02637461e":32,"02712669e":32,"02730001e":32,"02754232e":41,"02925814e":41,"02968636e":32,"03":[32,41],"0300":47,"03028475e":32,"03053516e":41,"03128608e":32,"03145841e":32,"03243918e":32,"03257652e":41,"0330":47,"03335943e":32,"03398554e":41,"03478593e":32,"03496304e":41,"03540064e":41,"03600335e":41,"03750293e":41,"03760871e":32,"03872254e":32,"03880605e":32,"04":[8,32,41],"04081079e":32,"04119481e":32,"04159":[19,20],"04167677e":41,"04227657e":41,"04303905e":41,"04376091e":41,"04388196e":32,"04425869e":41,"04437640e":32,"04463429e":41,"04551275e":41,"04587221e":32,"04680798e":41,"04717023e":41,"04761188e":32,"04761873e":41,"04771734e":41,"04772201e":32,"04795949e":32,"04809226e":41,"04822710e":32,"04852029e":41,"04874016e":32,"04947969e":41,"05":[8,44],"05019362e":32,"05144850e":32,"05169697e":32,"05185719e":41,"05273729e":41,"05301257e":41,"05314606e":41,"05625390e":32,"05640656e":32,"05667166e":41,"05756191e":41,"05776466e":41,"05919485e":32,"05920354e":41,"05957368e":41,"05972891e":32,"05994362e":32,"0600":47,"06002222e":32,"06026421e":41,"06148241e":32,"06148329e":32,"06210334e":41,"06325925e":32,"06413557e":32,"06576427e":32,"06577017e":32,"06581083e":41,"06702779e":41,"06757526e":41,"06783829e":32,"06792355e":32,"06797182e":32,"06819102e":32,"06850095e":41,"06857690e":41,"06963077e":41,"0700":47,"07021418e":32,"07183093e":32,"07209305e":41,"07346811e":41,"07361911e":41,"07424044e":32,"07534600e":32,"07556239e":32,"07561498e":41,"07725342e":32,"07753378e":32,"07776551e":32,"0781":40,"07870363e":32,"07880044e":41,"07924467e":41,"08":34,"0800":47,"08027281e":32,"08153991e":32,"08175227e":32,"08308663e":41,"08317980e":41,"08348359e":32,"08448056e":41,"08482371e":32,"08526510e":32,"08542211e":32,"0868":37,"08703782e":32,"08778544e":32,"08817792e":32,"08897503e":32,"08912699e":41,"08938472e":32,"09":8,"09025812e":32,"09036772e":41,"09057900e":41,"09094712e":32,"09124164e":41,"09149733e":41,"09254644e":41,"09388471e":32,"09415695e":32,"09463034e":32,"09494858e":32,"09652805e":41,"09653721e":32,"09710224e":32,"09790113e":32,"09801983e":41,"09888444e":32,"09910303e":41,"09964232e":32,"09974838e":32,"1":[1,7,12,13,19,20,21,23,25,26,28,29,31,32,34,35,36,37,39,40,41,42,43,44,45,46,47,49,51,52],"10":[8,40,44,45,46,47],"100":[15,25,40,46,47],"1000":[44,47,54],"10000":17,"10021417e":41,"10034638e":32,"10047256e":32,"10088848e":32,"101":23,"1024":[19,20],"10292364e":41,"10305588e":32,"10343312e":32,"10596702e":41,"106":47,"10611510e":41,"10633622e":41,"10671862e":41,"107":46,"1080":[56,57],"10899068e":32,"10948697e":41,"11":[8,40,44,48],"110":46,"1100":47,"11124226e":41,"11133025e":41,"11300245e":41,"11347827e":41,"11365090e":32,"114":46,"11412094e":41,"11456517e":41,"11509175e":41,"11572169e":41,"11691907e":32,"11817739e":41,"11863718e":41,"11874017e":41,"11886488e":32,"119":46,"11913570e":32,"12":[16,40,47],"120":[46,58],"1200":[42,47],"12021770e":41,"12169042e":41,"123":47,"12349890e":41,"124":46,"125":[46,47],"12538061e":41,"12544761e":41,"12556038e":32,"126":46,"12659113e":32,"127":46,"12716393e":41,"12764148e":41,"1280":[40,44,51,52],"12865234e":32,"12872":25,"12891664e":32,"12921613e":32,"12997630e":41,"13":[8,34,46],"13072216e":32,"13084059e":32,"13108735e":41,"13168020e":41,"133":[16,46],"13323163e":41,"1333":[44,45],"13334349e":41,"13351250e":41,"13454091e":32,"135":46,"13520974e":32,"13596292e":41,"13632814e":41,"13658118e":41,"138":46,"13817184e":32,"13832005e":41,"13842775e":41,"139":46,"13952673e":41,"14":16,"1400":47,"14032232e":32,"141":46,"14109694e":41,"14128119e":32,"14128824e":32,"142":46,"14213758e":32,"14247333e":32,"143":46,"14312282e":32,"14319587e":41,"14345273e":41,"14361201e":32,"14363851e":32,"14391735e":32,"14424412e":32,"14461673e":41,"145":46,"14539692e":32,"14662452e":32,"147":46,"14770915e":41,"148":46,"14851037e":41,"14901921e":32,"14911471e":41,"14935338e":32,"150":40,"15011014e":32,"15011558e":32,"15016455e":32,"15059445e":32,"15088625e":32,"151":46,"15179000e":41,"152":46,"15205015e":32,"15210213e":32,"153":46,"15333085e":41,"15350205e":41,"154":46,"15448396e":41,"155":46,"15506293e":41,"15524881e":32,"15566470e":41,"15570085e":41,"156":47,"15638290e":32,"15653182e":32,"15730277e":41,"15752975e":41,"15864931e":32,"15872935e":41,"159":47,"15981518e":41,"16":46,"160":[46,47],"1600":47,"16018163e":41,"16088879e":32,"16145634e":41,"16156656e":41,"16201906e":32,"163":46,"16348874e":32,"16391443e":41,"164":46,"16421821e":41,"16431226e":41,"16490777e":41,"16517083e":32,"16562790e":32,"166":[46,47],"16624872e":32,"16665622e":32,"167":46,"16817471e":41,"16826625e":32,"16886530e":41,"16887195e":41,"16902778e":32,"170":46,"17055869e":41,"17066111e":32,"17100105e":32,"17129627e":41,"17260859e":41,"17301763e":41,"17315123e":32,"17322725e":41,"17367997e":32,"17386789e":41,"17392880e":41,"17454702e":32,"17480377e":41,"17488391e":41,"175":47,"176":46,"17600741e":32,"17604837e":32,"17828159e":32,"179":[46,47],"17928680e":32,"17957460e":32,"18":23,"180":47,"18010833e":32,"181":46,"18129853e":41,"18177504e":32,"18192625e":41,"182":46,"18296965e":41,"18372489e":41,"184":46,"185":46,"18546802e":41,"18551544e":32,"18587281e":32,"18590377e":41,"186":46,"18610614e":41,"18648595e":32,"187":46,"18733375e":32,"18753064e":41,"18762107e":41,"18762606e":32,"188":46,"18838328e":32,"18845964e":41,"18892872e":32,"189":46,"18907316e":32,"18926775e":32,"18945779e":41,"18981384e":32,"19":47,"1900":47,"191":46,"19146957e":32,"192":46,"1920":[40,44,51,52,56,57],"19242426e":32,"193":46,"19328787e":32,"194":46,"195":46,"19576752e":32,"19592251e":32,"196":46,"19628180e":41,"19658102e":41,"197":46,"19779656e":32,"19797095e":32,"198":[46,47],"19818748e":41,"19879779e":41,"19900324e":41,"19934985e":41,"2":[1,3,5,7,12,13,15,19,20,21,26,29,32,33,36,37,40,41,42,44,45,46,47,49,51,52],"20":[1,16,40,42],"200":[40,46,54],"2000":47,"2005":25,"20071020e":32,"2010":[19,20],"20113772e":32,"2013":34,"2017":[3,50,54,57,58],"2018":58,"202":46,"20267645e":32,"203":46,"20330610e":41,"20360519e":32,"20390712e":32,"204":46,"20466817e":32,"2048":[23,25],"20483222e":32,"20574823e":41,"20592140e":41,"20625557e":32,"20627365e":41,"207":46,"20788081e":32,"208":46,"20816324e":41,"20873807e":41,"209":46,"20973016e":41,"20979209e":41,"21":16,"210":47,"211":46,"21105030e":32,"212":46,"2120":37,"21200398e":41,"2126":40,"21373030e":32,"214":46,"21516043e":41,"21535104e":32,"21552847e":41,"21621182e":41,"217":47,"21776547e":32,"2178":40,"21874327e":41,"21884820e":41,"219":47,"21935562e":41,"22":37,"220":[46,47],"22074427e":32,"221":46,"22128649e":32,"22136800e":32,"22140182e":41,"22236057e":41,"22349080e":41,"22355425e":32,"224":[20,25,28,37,46],"22413266e":41,"225":[20,25,28,37],"22500":47,"22549850e":41,"22604942e":32,"22653131e":32,"22662459e":41,"227":46,"22738850e":32,"22795155e":32,"22852834e":41,"22862365e":32,"229":[20,25,28,37,46],"22904466e":32,"22959938e":41,"23":16,"2300":47,"23045578e":41,"23059211e":32,"231":47,"23125445e":32,"23190917e":32,"23199876e":32,"23247777e":32,"23266388e":32,"23298118e":41,"23325032e":32,"23332853e":41,"23381958e":41,"235":46,"23599575e":32,"236":46,"23621722e":41,"23647925e":41,"23674810e":41,"238":46,"23802378e":32,"23885994e":32,"239":46,"23904364e":41,"23983146e":41,"23990757e":41,"23998607e":41,"24":42,"240":46,"2400":47,"24025177e":41,"241":46,"24135251e":32,"24157363e":41,"24232008e":41,"24270324e":32,"24286011e":32,"243":46,"24386672e":32,"24405108e":32,"24447882e":41,"24458203e":41,"24510975e":32,"24537933e":41,"24551036e":41,"24568951e":32,"24572259e":32,"246":46,"24603436e":32,"24609747e":32,"24663319e":41,"24813112e":41,"24834611e":32,"24873060e":32,"24927253e":32,"24974845e":41,"24997086e":41,"25":21,"250":[24,28,29],"25138723e":41,"25226439e":41,"25288036e":32,"25394419e":41,"25403250e":32,"25409470e":41,"25432243e":41,"25454067e":32,"25488366e":41,"25493990e":41,"255":[37,44,46,47],"25555607e":41,"256":[19,20,25,37],"257":47,"258":47,"25918035e":32,"2600":47,"26016131e":41,"26040759e":41,"26152547e":32,"26236267e":41,"26250020e":32,"263":47,"265":47,"266":47,"26601184e":32,"26812553e":32,"26835802e":32,"26910887e":32,"26920870e":32,"26956763e":41,"26990011e":41,"270":47,"2700":47,"27099208e":32,"27192841e":41,"27377255e":32,"27495814e":41,"275":47,"27503982e":32,"27529135e":32,"27608733e":32,"27636232e":32,"27690540e":32,"27724104e":41,"27760922e":41,"2779":37,"27835129e":41,"279":47,"27989183e":32,"28":42,"28001460e":41,"28007186e":32,"28085192e":32,"28136816e":41,"28142564e":32,"28164126e":32,"282":47,"28205277e":32,"28238067e":32,"28260103e":32,"28285096e":32,"28407752e":32,"28413887e":32,"28435191e":41,"28471761e":41,"28508218e":32,"28545721e":32,"286":47,"28731766e":32,"28751525e":32,"28767839e":32,"28774609e":41,"28852497e":32,"289":47,"28959047e":41,"28976645e":41,"29":45,"290":47,"29011612e":32,"29149030e":32,"29159375e":41,"29201944e":32,"29264704e":41,"29333737e":32,"294":47,"295":47,"296":47,"29713198e":32,"29738768e":41,"298":47,"29956402e":41,"29976252e":32,"2d":[13,33,34,36,40,41,42,44,46,47,48],"3":[1,13,20,21,25,32,33,34,37,40,41,42,44,46,47,49,51,52],"30":16,"300":[19,20,41,46,47,50],"3000":47,"30000":47,"30150565e":32,"30190983e":41,"302":47,"30243579e":41,"30263620e":41,"30273065e":41,"303":47,"30311304e":41,"30397866e":32,"30646462e":41,"30647157e":41,"30702881e":32,"30704639e":32,"30764343e":41,"30773518e":41,"31":47,"310":47,"3100":47,"31001691e":32,"31076889e":41,"31128677e":41,"31176982e":32,"31183136e":41,"31229004e":41,"3125":47,"315":47,"31666474e":32,"31696089e":32,"31799552e":32,"318":47,"31807037e":32,"31839992e":41,"31857782e":41,"31877959e":41,"31930795e":41,"31939487e":41,"32":47,"3200":47,"32080311e":41,"32133632e":32,"32206036e":32,"32284414e":41,"32289133e":41,"32321250e":41,"32341182e":32,"32366280e":41,"32369469e":41,"324":47,"32434518e":32,"32439932e":32,"32449319e":41,"325":47,"32575747e":41,"32596057e":41,"32635855e":41,"32765525e":41,"32853739e":32,"32861016e":32,"32861072e":41,"32959088e":32,"32966981e":41,"33":[16,47],"33000444e":32,"33112482e":41,"33177747e":32,"33192793e":41,"33229234e":32,"33270789e":32,"33455423e":41,"33465786e":41,"33531934e":41,"33592468e":41,"33595493e":32,"33667201e":32,"33678078e":41,"33735841e":32,"33760713e":41,"33810312e":32,"34":[23,47],"3400":47,"34068612e":41,"34103563e":32,"34146171e":41,"34189645e":41,"34305328e":32,"34473480e":41,"34515366e":41,"34645829e":41,"34748716e":32,"34829280e":32,"34837902e":32,"34851278e":41,"34916476e":41,"34930565e":41,"34949801e":41,"35":47,"3500":47,"35023717e":32,"35038607e":41,"35063551e":32,"35093884e":41,"35117482e":41,"35205932e":41,"35281259e":41,"35338970e":41,"35368218e":41,"35414255e":32,"355":47,"35567359e":32,"35577368e":32,"35584003e":41,"35621356e":41,"35636532e":41,"35703430e":41,"35719585e":41,"35765215e":41,"35784922e":32,"358":16,"35853428e":32,"359":47,"35962038e":41,"36":[16,47],"360":47,"36002306e":32,"36022002e":41,"36165975e":41,"36199434e":41,"362":47,"36231448e":41,"36241428e":32,"36244965e":32,"36276863e":41,"36280255e":41,"36290787e":32,"36317878e":32,"36335809e":32,"36422203e":32,"36431987e":32,"36445537e":41,"36521038e":32,"36553374e":41,"36569415e":41,"36606170e":32,"368":7,"36827042e":41,"36962272e":32,"36981317e":32,"37":47,"3700":47,"37088945e":32,"371":47,"37103539e":41,"37155793e":41,"37174634e":41,"37310897e":32,"37340015e":41,"37402211e":32,"37432704e":41,"37436600e":41,"37532005e":32,"37535755e":32,"37588967e":32,"37594808e":41,"376":47,"37664331e":32,"37664382e":32,"37693244e":32,"377":47,"37711062e":32,"37727856e":41,"37745020e":32,"37782600e":32,"37793197e":32,"37835736e":41,"37844146e":32,"37934865e":32,"38":16,"3800":47,"38015272e":32,"38051907e":32,"38175275e":41,"382":47,"38228521e":41,"38234810e":32,"38256748e":41,"384":[44,47],"38471419e":41,"38481435e":32,"385":47,"38641228e":41,"38698314e":41,"38740534e":32,"388":47,"38854459e":32,"38933458e":41,"38938577e":32,"38944222e":32,"38944719e":32,"38949338e":32,"38963467e":32,"39":[40,44,46,47],"39018344e":41,"39137950e":41,"39148222e":41,"39254133e":32,"393":47,"39320385e":32,"39329815e":41,"39344775e":32,"394":47,"395":47,"39508256e":32,"39590570e":41,"396":47,"39600003e":41,"39606577e":32,"39638784e":41,"39674161e":41,"39763363e":41,"39778635e":41,"39815691e":32,"39864336e":32,"39977294e":32,"3d":[10,13,30,33,34,37,46,48,49],"4":[7,13,19,20,21,25,26,28,29,32,33,34,37,40,41,42,44,46,47,48,50,51,52,56],"40":[16,46],"400":[35,36,39,44,46],"4000":47,"40034328e":41,"40113234e":32,"40114690e":32,"40357930e":32,"40482958e":32,"40582895e":32,"406":[20,25,28,37],"40639729e":32,"40649491e":32,"40667169e":32,"40729489e":32,"40735860e":32,"40752254e":32,"40962720e":41,"41":46,"41002649e":32,"41071030e":32,"4111":37,"41128808e":41,"41241494e":41,"41645203e":32,"41729016e":32,"41786611e":41,"41820729e":41,"41830554e":32,"41831862e":41,"41883656e":32,"41889615e":41,"41894734e":32,"41957316e":32,"42":[37,46],"4200":47,"42163535e":32,"422":46,"42204493e":32,"42271943e":41,"42281033e":32,"42288814e":41,"42297352e":41,"42312343e":41,"42385955e":32,"42528038e":41,"42606505e":32,"42685751e":32,"42730215e":41,"42735322e":41,"42819821e":32,"42853762e":32,"42895935e":41,"42904727e":32,"42925604e":32,"42938804e":32,"42956539e":32,"43":16,"4300":47,"43023087e":41,"43042841e":32,"43075657e":41,"43130401e":32,"43141384e":41,"43194812e":41,"432":16,"43413371e":41,"43423731e":32,"43444735e":41,"43445851e":32,"43492379e":41,"43538160e":41,"43570700e":41,"43597600e":41,"43636328e":41,"43676481e":32,"43878267e":32,"43905462e":32,"43952490e":32,"44":[16,47],"4400":47,"44063067e":32,"44081660e":41,"44172397e":32,"44270017e":41,"44286947e":32,"44385708e":32,"44392263e":32,"44398756e":32,"44399895e":32,"445":47,"44551811e":41,"44630692e":41,"44641618e":41,"44673896e":32,"44688566e":41,"44726311e":32,"44732274e":41,"44734519e":41,"44783429e":41,"44807442e":41,"44810261e":41,"449":16,"44939437e":41,"45":46,"45021435e":41,"45065270e":41,"45066351e":32,"45093133e":32,"45196455e":41,"45234899e":32,"45310316e":32,"45398947e":32,"45454722e":32,"45599786e":32,"456":[20,25,28,37],"45611063e":32,"45633296e":41,"45654602e":41,"45726122e":32,"45766837e":41,"458":42,"45811637e":41,"45848217e":32,"45859593e":32,"45876408e":41,"46":[16,46,47],"46008537e":41,"46064335e":41,"46084398e":32,"46232387e":41,"46315548e":41,"46315707e":41,"46321268e":32,"46385916e":41,"46388149e":32,"46406351e":32,"46417379e":32,"46599072e":32,"46606147e":41,"46674042e":41,"46745001e":41,"46897770e":41,"46924197e":32,"46929039e":41,"47":[16,47],"47057405e":41,"47103447e":32,"47156683e":32,"47200802e":32,"47289682e":32,"47393835e":32,"474":47,"47478223e":32,"47495599e":32,"47538196e":32,"47594311e":41,"476":47,"47603517e":41,"47700368e":41,"47708146e":32,"47869084e":41,"47907059e":41,"47917999e":41,"47944036e":32,"480":[44,45,47],"4800":47,"48028846e":32,"48030262e":41,"48033253e":32,"48081258e":41,"481":16,"48104179e":32,"48147176e":41,"48153055e":41,"48213761e":41,"48244354e":41,"48255898e":32,"485":[20,25,28,37],"48585800e":32,"48598628e":32,"48600382e":32,"48744851e":32,"48818781e":32,"48834808e":41,"48855822e":41,"48886058e":41,"48938736e":32,"48954334e":41,"48989136e":41,"49":[16,42],"4900":47,"49046972e":32,"49128421e":41,"49172560e":41,"49187534e":41,"49219411e":41,"49314570e":41,"49325455e":41,"49378200e":41,"49435524e":32,"49493830e":41,"49555979e":41,"49573932e":32,"49588511e":32,"496":7,"49621867e":41,"49626810e":41,"497":47,"49703236e":32,"49710231e":32,"49713774e":41,"49803308e":32,"49808423e":41,"49885566e":32,"49940534e":41,"49983560e":32,"4k":4,"5":[12,16,21,26,28,29,32,35,37,40,41,42,47,53,54,56],"50":[16,23,40],"500":[44,46,47,53],"5000":47,"50006558e":41,"50085920e":41,"50138889e":32,"50165707e":32,"50186277e":32,"50199273e":41,"50207678e":32,"50293453e":41,"50510245e":32,"50537991e":32,"50632170e":41,"50727936e":41,"50751731e":32,"50835760e":32,"50882574e":41,"50888647e":41,"5095":37,"50996342e":32,"51":[42,46],"5100":47,"51005229e":32,"51024932e":32,"51097186e":32,"51137867e":32,"512":[23,37,42,44,45],"51228718e":32,"51248612e":41,"51259558e":32,"51268173e":32,"51292242e":41,"51396928e":41,"51441696e":32,"51493693e":32,"51499297e":32,"51513787e":32,"51539646e":32,"5155":40,"51613569e":32,"51674214e":41,"51688990e":32,"51800828e":41,"51831804e":41,"51860355e":41,"51887899e":32,"51926796e":32,"51955247e":41,"51955534e":32,"52001675e":32,"52037835e":32,"52071569e":41,"52142056e":32,"52465819e":32,"52506672e":32,"52569220e":41,"52606388e":41,"52658803e":41,"52747529e":41,"52754593e":32,"52757911e":41,"52768676e":41,"52829437e":32,"52851087e":32,"52855277e":41,"52891400e":32,"52902272e":41,"52993649e":41,"53":[16,47],"53208562e":41,"53285631e":32,"53359910e":41,"53381575e":41,"53381805e":41,"53438169e":32,"53471212e":32,"53631860e":41,"537":16,"53742844e":41,"53758957e":32,"53860176e":41,"53878839e":32,"53923576e":32,"53975753e":32,"53979624e":41,"54008178e":41,"54110357e":41,"54207130e":32,"54268708e":32,"54319462e":41,"54334671e":41,"544":[44,45],"54466717e":41,"54512593e":41,"54525691e":32,"54569385e":32,"54574824e":41,"54686235e":41,"54779696e":32,"54782298e":41,"54862369e":41,"54995590e":32,"55":[16,47],"5500":47,"55054480e":41,"55063802e":32,"55228428e":32,"55293852e":41,"55350342e":41,"55354685e":41,"55362918e":41,"55396186e":32,"55398361e":41,"55427503e":41,"55437307e":41,"55583297e":32,"55609667e":32,"55835216e":32,"55939103e":41,"55966798e":41,"55986460e":32,"55989053e":32,"55989290e":32,"56010235e":41,"56051858e":41,"56194948e":32,"56211088e":41,"56224417e":32,"56350143e":32,"56414250e":41,"56469828e":41,"56527501e":41,"56560266e":32,"56565124e":41,"56720742e":32,"56725107e":32,"56746286e":41,"56804765e":41,"56809213e":32,"56877725e":32,"56884720e":41,"56892042e":41,"56978680e":41,"5700":47,"57129571e":41,"57217245e":41,"57303626e":41,"57429174e":41,"57439643e":32,"57449467e":32,"57587518e":32,"576":[44,45],"57713193e":32,"57783664e":41,"57817782e":32,"57948552e":41,"57962473e":32,"5800":47,"58061958e":41,"58137991e":32,"58148635e":32,"58234590e":41,"58270290e":32,"58273479e":32,"58316872e":41,"58318236e":32,"58339510e":41,"58376747e":41,"58519625e":32,"58551430e":32,"58569475e":32,"58687131e":32,"58765499e":41,"58785518e":41,"58853357e":41,"5900":47,"59014283e":41,"59044999e":32,"59054548e":41,"59073593e":32,"59134616e":32,"59178390e":41,"59196241e":41,"59210991e":32,"59247284e":32,"59308577e":41,"59315360e":32,"59333520e":32,"59462401e":41,"59469003e":32,"59475331e":32,"595":16,"59544810e":41,"59565306e":32,"59610660e":32,"59693846e":41,"59758080e":41,"59768601e":32,"59967107e":32,"59970786e":41,"5th":40,"6":[16,19,20,21,23,25,26,29,32,37,40,41,57],"60":16,"600":[35,36,39,44],"6000":47,"60041588e":41,"60048677e":32,"60049439e":41,"60173608e":32,"60185582e":32,"60189655e":32,"60212139e":41,"60247861e":41,"60250229e":41,"60351173e":41,"60356952e":41,"60495137e":32,"60587540e":41,"60610769e":41,"60780852e":32,"60788182e":41,"608":[44,45],"60953736e":32,"61018186e":32,"61021307e":32,"61022231e":32,"61079053e":41,"61103382e":32,"61237738e":41,"61368908e":32,"61420213e":32,"61428914e":32,"61433931e":41,"61495355e":41,"61506634e":41,"61564484e":41,"61587099e":41,"61600747e":32,"61617961e":41,"61637559e":32,"61654997e":32,"61665926e":41,"61676887e":32,"61688289e":32,"61717298e":32,"61718622e":32,"61773856e":41,"61922468e":41,"61952807e":32,"6200":47,"62113942e":32,"62143484e":41,"62263973e":32,"62274267e":41,"62281202e":41,"62404387e":32,"62477445e":32,"62584241e":41,"62588816e":41,"62596573e":32,"62599278e":32,"62633722e":41,"62679464e":41,"62688509e":32,"62728852e":41,"62787151e":41,"63336631e":32,"63396338e":32,"63399274e":41,"63399427e":32,"63423042e":41,"63503936e":41,"63510266e":41,"63545749e":32,"63574387e":32,"63692683e":41,"63707289e":41,"63736254e":41,"63906017e":41,"63913892e":41,"63981616e":32,"64":[17,46],"640":[44,45,46,47],"6400":47,"64021870e":32,"64131004e":41,"64142216e":41,"64166583e":41,"64243515e":32,"64344971e":32,"64359957e":41,"64385619e":41,"64421797e":41,"64427192e":41,"64432783e":32,"64600468e":41,"64662324e":41,"64701496e":41,"64750741e":32,"64756788e":32,"64870111e":41,"64894611e":41,"64911196e":41,"64915846e":41,"64933322e":32,"65":[16,47],"65067771e":32,"65088251e":41,"65098195e":41,"65109356e":32,"65119162e":32,"65147551e":41,"65152192e":41,"65204624e":32,"65284284e":32,"65303336e":32,"65400850e":32,"65421511e":32,"65483130e":41,"65610983e":32,"65652017e":32,"65752488e":32,"65766207e":41,"65893743e":32,"65907715e":32,"66":[16,46],"6600":47,"66010074e":32,"66092573e":32,"66148970e":32,"66270462e":41,"66285481e":41,"66434807e":32,"66452541e":32,"66494927e":32,"66543838e":41,"66659251e":41,"66659612e":41,"66723721e":41,"66735310e":41,"66822109e":32,"66857574e":32,"66905130e":41,"66917036e":32,"66917625e":32,"66960705e":41,"66971043e":41,"67186664e":41,"672":[44,45],"67294100e":32,"67331559e":32,"67376126e":41,"67384841e":32,"67483186e":41,"67483792e":41,"67499068e":41,"67527258e":41,"67567753e":41,"67653709e":41,"67672440e":32,"67693356e":32,"67700563e":41,"67713608e":32,"67761680e":41,"67807741e":32,"67831527e":32,"67877079e":32,"67928400e":41,"67975352e":41,"67979530e":41,"68":16,"68000022e":41,"68046141e":41,"68192688e":32,"68211478e":41,"68220504e":32,"68236094e":41,"68399237e":32,"68479265e":41,"68507392e":41,"68516922e":41,"68583363e":32,"68618004e":32,"68693320e":32,"68705006e":32,"68732776e":41,"68743685e":41,"68751827e":32,"68811191e":41,"68928579e":41,"68971811e":41,"68983008e":32,"69041584e":32,"69090609e":41,"69146903e":32,"69166225e":32,"69224194e":41,"69284451e":41,"69303778e":32,"69435785e":32,"69470848e":32,"69479280e":41,"69482472e":32,"69515134e":41,"69743604e":32,"69853080e":32,"69859848e":32,"69879433e":41,"69946622e":32,"69998771e":32,"7":[13,16,32,33,40,41],"70":16,"700":[44,53],"70002547e":41,"70008731e":41,"70072899e":32,"70087830e":41,"70103464e":41,"70163010e":41,"70273302e":41,"70313620e":41,"70339913e":32,"70343670e":32,"704":[44,45],"70430650e":41,"70582914e":41,"70594696e":32,"70613403e":32,"70685630e":32,"70712863e":32,"70718343e":41,"70738131e":41,"70792886e":32,"70804075e":32,"70921545e":41,"70954888e":32,"71":47,"7100":47,"71011013e":32,"71052133e":41,"71176431e":32,"71190578e":32,"71259938e":41,"71267741e":41,"71332358e":32,"71382250e":41,"71434822e":41,"71457265e":41,"71499615e":41,"71508068e":32,"71521833e":32,"71522226e":32,"71602391e":32,"71614282e":41,"71659425e":41,"71664230e":32,"71677147e":41,"71780270e":32,"71791272e":41,"71857940e":32,"72":47,"7200":47,"72007854e":41,"72028956e":41,"72103431e":32,"72150734e":32,"72187218e":41,"72226359e":41,"72271830e":41,"72366775e":32,"72665983e":32,"72679423e":41,"72939716e":41,"73100681e":41,"73111726e":32,"73210450e":41,"73226457e":41,"73304404e":41,"73455073e":32,"73455615e":41,"73510326e":32,"73568654e":41,"736":[44,45],"73635007e":32,"73657264e":41,"73661135e":41,"73737889e":32,"73823287e":32,"73981643e":41,"73997584e":32,"74":45,"74112186e":32,"74129347e":41,"74143943e":32,"74146505e":41,"74192287e":41,"74220980e":32,"74268780e":32,"74578149e":32,"74845295e":32,"74953936e":41,"74958449e":41,"74985858e":41,"75":16,"7500":47,"75011383e":41,"75019079e":41,"75021987e":32,"75101960e":41,"75140731e":41,"75259072e":32,"75654494e":32,"75667244e":41,"75735485e":32,"75753328e":41,"75873084e":41,"75963803e":32,"76":47,"7600":47,"76006581e":41,"76019975e":41,"76141550e":32,"76275759e":32,"76323899e":41,"76334942e":32,"76360560e":32,"76368554e":32,"76438300e":41,"76451505e":32,"76487271e":41,"76610047e":41,"768":[44,45],"76964335e":32,"77":[46,47],"77243902e":32,"77297886e":32,"77308245e":32,"77310918e":32,"77407023e":41,"77438827e":32,"77441592e":41,"77486195e":41,"77529370e":32,"77586006e":41,"77598847e":32,"77616067e":32,"77629993e":41,"77646448e":32,"77698357e":32,"77751604e":41,"77756810e":32,"77773580e":41,"77797875e":32,"77801385e":41,"77826693e":41,"77856308e":32,"779":16,"77914277e":41,"77931165e":32,"77946945e":32,"78":[16,47],"78037319e":41,"78091111e":41,"78129718e":41,"78176969e":32,"78197124e":32,"78372089e":41,"78424306e":41,"78438102e":32,"78455093e":41,"78460247e":32,"78463067e":41,"78496098e":41,"78535344e":32,"78593137e":41,"78637232e":41,"78722001e":32,"78777501e":32,"78862323e":41,"78869568e":41,"789":16,"79035263e":41,"79110819e":32,"79143252e":41,"79162661e":41,"79185960e":32,"79207557e":32,"79241608e":41,"79270662e":41,"79300919e":41,"79319073e":32,"79384543e":32,"79448105e":41,"79474811e":32,"79491331e":32,"79491716e":41,"795":16,"79509705e":32,"79603274e":32,"79725511e":32,"79792720e":41,"79839197e":41,"79853955e":41,"79890251e":32,"79909151e":41,"79979004e":41,"8":[3,12,16,19,20,23,25,32,33,40,41],"80":[16,40],"800":[44,45],"80122702e":32,"80164268e":41,"80249115e":41,"80379387e":41,"80387101e":32,"80421045e":32,"80481041e":32,"80484864e":41,"80548756e":41,"80576126e":41,"80758245e":32,"80799150e":41,"80824122e":41,"80945642e":41,"81":47,"8100":47,"81045856e":32,"81066590e":41,"81130519e":41,"81157385e":32,"81346552e":41,"81514809e":32,"81639375e":41,"81666755e":41,"81685497e":32,"81706826e":32,"81749518e":41,"81779440e":32,"81817431e":41,"82":47,"82050048e":32,"82143918e":41,"82177723e":41,"82201724e":32,"82227314e":32,"82237027e":41,"82255091e":41,"82357466e":41,"82402189e":41,"82422198e":32,"82470218e":41,"82508046e":41,"82511740e":32,"82536143e":32,"82588178e":32,"82621410e":41,"82679031e":32,"82708323e":32,"82773104e":32,"82937714e":32,"83030218e":32,"83051399e":32,"83091248e":41,"83101718e":41,"83109898e":41,"83324946e":32,"83329206e":41,"83478347e":32,"83557829e":32,"83647122e":32,"83708934e":41,"83763381e":32,"83780808e":32,"83815209e":41,"83872906e":41,"83890262e":32,"83936217e":32,"83961364e":32,"83987536e":41,"83995094e":32,"84":[46,47],"84010504e":41,"84090080e":41,"84101993e":32,"84146473e":41,"84158592e":32,"84164432e":32,"84251230e":32,"84268810e":41,"84279112e":32,"84296254e":41,"84301525e":32,"84331772e":32,"84339841e":32,"84388865e":41,"84463618e":41,"84465717e":32,"84538982e":41,"84682345e":41,"84749912e":41,"84768479e":32,"84845043e":41,"84866116e":41,"84896261e":41,"84996651e":32,"85":[16,47],"85021536e":32,"85091537e":32,"85120679e":41,"85211241e":41,"85238648e":32,"85366584e":41,"85393316e":41,"85620672e":41,"85671024e":41,"85685524e":32,"85841072e":41,"85921243e":41,"85940756e":32,"85944980e":32,"8600":47,"86003087e":32,"86123463e":32,"86166188e":32,"86326450e":41,"86393952e":41,"86460032e":41,"86501833e":41,"86584065e":41,"86661993e":32,"86730818e":32,"86814944e":41,"86865287e":41,"86865847e":41,"86893799e":41,"86910936e":41,"86956994e":41,"87143027e":41,"87310275e":41,"87415176e":41,"87435697e":32,"87451562e":41,"87564552e":32,"87583674e":41,"87755884e":32,"87812190e":32,"87899595e":32,"87901050e":32,"8800":47,"88070478e":41,"88312613e":41,"88357981e":41,"88372827e":41,"88466606e":32,"88564202e":32,"88586194e":32,"88630857e":41,"88660435e":32,"88764308e":32,"88766773e":32,"88837141e":32,"88881648e":41,"88891904e":32,"88915652e":41,"88978771e":41,"88997795e":41,"89":16,"89131362e":41,"89141990e":41,"89174282e":32,"89301891e":32,"89312950e":32,"89339351e":32,"89341429e":41,"89365069e":41,"89374835e":41,"89449456e":41,"89524910e":41,"89543529e":32,"89661277e":32,"89689550e":41,"89754352e":32,"89778210e":32,"89874894e":41,"89885559e":41,"89964217e":32,"9":[16,32,40,41,44],"90":16,"90004727e":41,"90063780e":32,"90076284e":32,"90100879e":32,"90206548e":32,"90246467e":41,"90291605e":32,"90313894e":41,"90345274e":32,"90623152e":41,"90666433e":41,"90748601e":32,"90754411e":32,"90760191e":32,"90879256e":32,"90948464e":32,"91":[20,21,25,26,51,52,54],"91010232e":41,"91144374e":41,"91153068e":41,"91191438e":32,"91256492e":41,"91299428e":41,"91303686e":32,"91348126e":41,"91404197e":32,"91447936e":32,"91559028e":32,"91583537e":32,"91605333e":32,"91632384e":41,"91752845e":32,"91784979e":41,"91825279e":32,"91838632e":41,"91866292e":32,"91948573e":32,"91977332e":41,"92020874e":41,"92035442e":32,"92058403e":41,"92086297e":32,"92103441e":41,"92270963e":41,"92275251e":32,"92363776e":32,"92396779e":32,"92420729e":32,"92519624e":41,"92566638e":32,"92764504e":41,"92827347e":41,"92906166e":32,"92927763e":41,"92941282e":41,"93008534e":32,"93014832e":41,"93024218e":32,"93101057e":32,"93132891e":32,"93189364e":32,"9325":37,"93250958e":32,"93299667e":32,"93357992e":32,"93365737e":32,"93404331e":41,"93574127e":32,"93591252e":32,"93610902e":32,"93714209e":32,"93715457e":32,"93814590e":41,"93820383e":41,"93923941e":41,"93950878e":41,"93970080e":41,"9400":47,"94003097e":32,"94014943e":41,"94024553e":41,"94055706e":32,"94087964e":32,"94134256e":41,"94207973e":41,"94254100e":41,"94374577e":32,"94398543e":32,"94429515e":32,"94497217e":32,"94516350e":41,"94535338e":41,"94554753e":41,"94617276e":32,"94652585e":41,"94653947e":41,"94669003e":32,"94687653e":32,"94712473e":32,"95":16,"9500":47,"95017101e":32,"95049659e":32,"95088060e":32,"95102486e":41,"95142190e":41,"95150689e":41,"95174015e":41,"95208665e":32,"95322363e":32,"95329289e":32,"95393683e":41,"95397235e":41,"95431976e":41,"95460682e":32,"95590970e":32,"95687222e":41,"95763808e":41,"95774269e":41,"95853576e":41,"95874170e":32,"95883835e":41,"95958340e":41,"95960441e":32,"96186448e":41,"96191997e":32,"96198589e":32,"96248508e":41,"96349881e":41,"96446913e":32,"96451715e":32,"96482586e":32,"96495936e":41,"96506289e":32,"96526082e":41,"96544465e":32,"96587537e":32,"96620758e":32,"96653278e":32,"96657595e":41,"96715933e":41,"96767380e":32,"96790655e":41,"9700":47,"97039598e":41,"97093255e":32,"97093540e":32,"97154195e":32,"97223866e":41,"97272138e":32,"97327524e":41,"97422870e":32,"97425823e":41,"97456217e":41,"97482060e":41,"97487153e":32,"97532977e":32,"97618669e":32,"97622434e":41,"97637064e":32,"97664087e":32,"97701028e":32,"97864437e":41,"97909964e":41,"97917295e":32,"97964669e":41,"98":[16,47],"9800":47,"98065324e":41,"98117998e":41,"98118201e":32,"98195025e":32,"98352541e":41,"98457295e":41,"98457887e":32,"98486173e":32,"98630104e":41,"98648200e":32,"98677499e":41,"98685712e":41,"98695788e":32,"98771755e":41,"98802021e":41,"98811528e":32,"99":46,"9900":47,"99009071e":32,"99061946e":41,"99098426e":41,"99201204e":32,"99214877e":41,"99223449e":32,"99253306e":41,"99328799e":32,"99377370e":41,"99413904e":41,"99460100e":32,"99507426e":32,"99540086e":41,"99682481e":32,"99706296e":32,"99775088e":41,"99808706e":41,"99813722e":32,"99829637e":32,"99835651e":32,"99921852e":32,"boolean":28,"break":44,"case":[3,13,26,34,37,39,44,51,52],"class":[0,1,2,4,5,6,7,8,9,10,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,34,35,36,37,38,39,40,41,42,43,45,46,49,50,51,52,53,54,56,57,58],"default":[0,1,5,7,13,15,16,20,21,24,25,26,28,29,31,32,33,35,39,40,41,42,43,44,46,49,50,54,56,57,58],"do":[1,7,24,26,31,43,47],"enum":1,"export":[48,49,51,52,53],"final":[9,10,11,15,16,25,33,42,43,54,57,58],"float":[16,20,21,25,26,28,29,31,32,39,44],"function":[3,7,13,15,17,19,20,21,23,24,25,26,28,29,31,33,35,36,39,50,52,53,54,56,57,58],"import":[3,7,13,16,20,24,25,28,35,36,39,40,41,42,44,45,46,47,50,51,52,53,54,56,57,58],"int":[1,3,5,7,9,10,12,13,15,16,20,21,24,25,26,28,29,31,33,43,50],"long":47,"new":[1,3,4,11,12,13,16,21,26,28,29,31,32,35,36,37,41,42,43,46,47,58],"return":[1,2,3,5,6,7,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,32,33,35,37,38,39,41,42,43,44,45,46,50,51,52,53,57],"short":10,"static":[6,12,13,19,21,24,26,29,31,32,33,41,43,53,57],"super":[45,50],"switch":35,"true":[1,3,6,12,16,20,21,23,24,25,26,28,29,31,33,34,35,36,37,40,41,42,44,45,46,47,48,50,53,54,56,57,58],"try":[1,15,40,44,47,54],"while":[15,17,19,23,26,32,41,42,44,47,51,52],A:[1,7,13,15,19,21,23,24,26,31,35,36,39,43,45,46,47,50,51,52,54],As:[47,50,51,52],Be:47,But:[1,35,36,47],By:[3,16,20,21,28,33,40,43,44,46,47,50,54,58],For:[2,3,5,6,13,15,19,20,24,25,26,29,34,35,36,37,40,44,46,47,48,50,51,52,53,54,56,57,58],If:[1,3,4,13,15,16,20,21,24,25,28,31,32,33,35,36,37,38,39,40,41,42,43,47,50,53],In:[3,13,25,26,35,36,42,44,45,46,47,50,51,52,53,54,56,57,58],Is:[28,47],It:[3,7,20,25,28,39,42,44,46,50,51,52,56,57,58],Its:54,NOT:26,No:[12,46],Not:1,OR:15,On:[50,58],One:[6,9,10,11,13,15,26,29,37,42,47,48],Or:[36,42,48],TO:[51,52],That:[20,21,26],The:[1,2,3,4,5,7,9,10,12,13,15,17,20,21,25,26,27,28,29,32,33,34,35,36,37,39,42,44,45,46,47,48,51,52,54,57,58],Then:[23,33,58],There:[50,57],These:[20,25,28,57],Theses:15,To:[1,2,3,4,5,6,9,10,11,15,20,25,28,32,37,45,47,56],Will:37,With:[33,35,36,45,50,53,58],_:[16,51,52],__:[13,51,52],__calib:13,__call__:[51,52],__camera_label:13,__getitem__:[1,31],__image0:13,__image1:13,__image2:13,__image3:13,__image4:13,__init__:[13,45,50,51,52],__lidar_label:13,__pose:13,__test:13,__train:13,__valid:13,__velodyn:13,_annot:[3,53],_h:12,_id:6,_load_ann:50,_load_imag:50,_match_image_ann:50,_prepar:13,_r:12,_size:12,_w:12,ab:[19,20,25],abl:[26,28,50,53,58],about:[2,5,6,24,26,42,45,49,50,53,54,56,57,58],abov:[20,25,26,28,40,42,44,47,51,52],abs_area:32,abs_box:41,abs_po:[32,41,42,47],absolut:[31,32,33,35,37,41,42,47,50],accel:13,acceler:13,access:[3,13,15,53],accord:[27,50,53,54,56],accordingli:[44,46],account:[26,29],accross:26,accumul:[26,29],accumulate_grad_batch:[26,29,56,57],achiev:[56,57,58],across:[15,21,26,43,46],activ:[19,20,21,23,25,28,52],activation_fn:[20,21,52,53],ad:[21,23,32,35,36,37,39,41,42,43,46],adamw:[21,26],adapt:[29,45],add:[13,15,16,21,26,29,37,43,44,47,56,57,58],add_argparse_arg:[21,24,26,29,53,54,56,57,58],add_gt_posit:16,add_label:31,add_properti:31,add_sampl:[15,16,29,56],addit:[13,24,25,26,29,36,46,50],addition:15,adit:[20,25],adjust:24,advantag:[40,54],advis:3,affin:23,after:[3,15,35,36,50,51,52],afterward:[17,19,23],aggreg:15,air:[51,52],airplan:47,albedo:[10,11],aldreadi:13,algorithm:[33,41],align:[33,41],all:[1,2,4,7,12,13,15,16,17,19,20,21,23,24,25,26,28,31,33,43,44,45,46,47,50,53,54,57,58],all_gt:8,alley_1:11,alley_2:11,allow:[3,25,27,28,39,50,53,56,57,58],alocept:[1,3,4,11,12,13,23,44,45,50,51,52,53,54,56,57,58],aloception_config:50,alodataset:[1,2,3,4,5,6,7,8,9,10,11,12,13,16,31,35,36,37,40,45,46,47,48,49,50],alodataset_config:[1,3,4,11,13],alonet:[15,16,17,19,20,21,23,24,25,26,28,29,48,49,51,52,53,54,56,57,58],along:[23,33],aloscen:[1,3,5,6,9,10,11,12,13,16,20,21,24,25,26,28,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,47,48,49,50,51,52,53,54],alotransform:12,alpha:21,alreadi:[4,13,20,25,28],also:[3,17,22,33,35,36,40,41,42,46,47,50,53,54,57,58],although:[17,19,23],alwai:[12,13],ambush_2:11,ambush_4:11,ambush_5:11,ambush_6:11,ambush_7:11,among:[21,26],amount:3,an:[1,3,9,15,16,20,25,26,28,35,36,37,39,40,42,44,45,47,50,51,52,53,54,57,58],angl:41,ani:[1,3,11,15,23,25,28,43,44,46,47,51,52],anim:[10,54],ann_fil:[3,4,45,50],ann_fold:50,ann_img_train_0:50,ann_img_train_1:50,ann_img_train_l:50,ann_img_val_0:50,ann_img_val_1:50,ann_img_val_m:50,ann_train:50,ann_valid:50,annot:[3,9,24,39,45,50,58],anoth:[3,39,40,41,46,47],ap50:16,ap:[14,29,49,56],apdataobject:16,api:[0,30,44,47,51,52],apli:12,apmask:16,apmetr:[15,16,56],apmetricscallback:[15,29],append:[16,24,25,26,29,37],append_boxes2d:[20,25,28,37,47,50],append_boxes3d:37,append_cam_extrins:43,append_cam_intrins:43,append_dispar:37,append_flow:37,append_label:[32,37,39,41,42,50],append_mask:43,append_occlus:[35,36],append_points2d:37,append_segment:[28,37],appl:47,appli:[7,12,16,20,21,25,26,31,35,36,43,46,50],applic:[3,50,53,54],apply_on_label:31,approach:56,approxim:[33,41],ar:[3,4,9,10,13,15,20,21,23,25,26,28,29,31,32,35,36,37,39,41,42,43,44,45,46,47,48,50,53,57,58],architectur:[14,18,22,24,25,28,45,48,49,53,54,56,57,58],area:[16,32,41,42,43,44,47],arg:[9,10,11,12,15,19,20,21,24,25,26,28,29,31,32,33,34,35,36,37,38,39,41,42,43,51,52,53,54,56,57,58],argpars:[45,53,54,56,57,58],argument:[20,21,24,25,26,29,49,56,58],argument_group:[26,29],argumentpars:[21,24,26,29,45,53,54,56,57,58],around:33,arrai:[32,35,36,37,39,41,42],articl:56,arxiv:[19,20,25],as_box:32,as_point:42,as_tensor:[31,35,47,51,52],aspect:47,assert:46,assert_input:26,assert_integ:43,assign:[21,26,28,50],associ:[15,20,42,46,51,52],associt:13,assum:[50,53,58],attach:[6,9,10,11,15,31,32,35,36,37,39,41,42,43,44,46,49],attent:[17,19,48,52,53,54,56,57,58],attribut:[3,15,16,24,26,28,29,50,54,56,57,58],aug_tensor_compat:[23,25],augment:[12,24,25,37,42,44,45,46,48,49],augmented_tensor:[31,32,33,34,38,41,42,43],augmentedtensor:[31,32,33,34,38,41,42,43],augmentend:31,augmet:43,augmt:31,author:1,auto_restor:31,auto_restore_nam:31,automat:[3,15,20,35,36,46,47,50],automaticli:[15,25,44],aux:28,aux_loss:[20,21,25,26,28,29,51,52],aux_loss_stag:[21,26,29],aux_output:[20,25,26,28,29],auxialiri:26,auxilari:[20,21,25,28,51,52],auxiliar:[26,29],auxiliari:[21,25],avaiabl:46,avail:[9,10,26,46,50,57,58],averag:[15,16,56],axi:[15,31,33,34,51,52],b:[13,15,19,31,37,41,42,44,46,47,51,52],back:33,backbon:[18,20,22,25,26,28,46],backbone_nam:[20,25],backbonebas:[19,23],background:[20,25,26,28,29],background_class:[25,28,51,52],backpack:47,backward:6,bad:[33,41],bag:47,bag_box:47,balanc:21,ball:47,bamboo_1:11,bamboo_2:11,banana:47,bandage_1:11,bandage_2:11,base:[0,2,3,4,5,6,7,8,9,10,11,12,13,14,16,17,18,19,21,22,23,24,26,27,28,29,31,32,33,34,35,36,37,38,39,41,42,43,44,46,47,48,49,53,54,56,57,58],base_dataset:1,base_metr:[15,29],base_metrics_callback:15,base_model:28,base_weight:[20,25,28],basebal:47,basedataset:[1,5,6,7,9,10,11,31],basenam:[51,52],basic:[0,18,22,27,30,40,49],bat:47,batch:[1,3,7,13,15,16,20,21,23,24,25,26,28,29,37,42,43,44,46,47,50],batch_cost_matrix:26,batch_fram:43,batch_idx:[15,26,29],batch_list:[3,13,16,20,24,25,28,43,44,51,52,53,54,56,57,58],batch_siz:[1,3,7,13,20,21,24,25,26,28,44,45,51,52,56,58],batchnorm2d:23,batchnorm:[19,23],bb_output:[20,25,28],bbox:13,bbox_3d:13,bbox_emb:25,bbox_proj:13,bear:47,becaus:[26,33,35,36,37,47,51,52,58],becom:35,bed:47,been:[3,32,35,36,53,54],befor:[15,23,25,29,31,47],begin:24,behavior:[19,23],behind:[33,47],being:[15,20,21,26,44,47],below:[3,37,40,50,54,56,57,58],bench:47,benefit:[46,51,52],best:[26,33,41],between:[9,10,12,21,24,26,31,32,35,37,39,40,41,42,46,51,52,56],betweent:56,betwen:[32,42],bev_box:[33,40],beyond:[45,46],bicycl:47,binar:28,binari:[20,21,25,28,37,39],bird:[33,40,47],blank:41,block:[20,28],boat:47,bondixboxes2d:47,book:47,bool:[6,12,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,39,41,42,43,45],both:[16,42,47,57],bottl:47,bottom_offset:31,bound:[20,21,25,26,28,33,40,53,56],bounding_boxes_2d:[32,37,40,46],bounding_boxes_3d:[33,37],boundingboxes2d:[13,15,16,20,21,25,26,28,29,30,33,37,40,46,47,49,50],boundingboxes3d:[13,15,30,37,40,41,49],boungd:33,bout:26,bowl:47,box:[3,13,15,20,21,25,26,27,28,29,32,33,37,38,40,41,43,44,46,47,48,52,53,54,56,57,58],box_refin:20,boxes1:[32,33,40,41],boxes2:[32,33,40,41],boxes2abspo:[32,41],boxes2d:[13,15,16,20,21,25,26,28,29,32,37,38,43,44,46,47,50,53,54,56,57,58],boxes2relpo:[32,41],boxes2xcyc:32,boxes2xyxi:32,boxes2yxyx:32,boxes3d:[13,15,37,40],boxes_3d:[37,40],boxes_3d_hflip:33,boxes_format:[32,37,47,50],boxes_hflip:32,boxes_limit:4,boxes_set:37,bridg:[51,52],bright:12,brightness_r:12,bring:33,broccoli:47,brute:[33,41],bu:47,build:[20,21,24,25,26,29,48,53,54,56,57,58],build_backbon:[20,23,25],build_bbox_emb:25,build_class_emb:25,build_criterion:[21,26,29],build_decod:[20,25],build_decoder_lay:[20,25],build_deforamble_transform:19,build_deformable_detr_r50:20,build_match:[21,26,29],build_model:[21,26,29],build_position_encod:17,build_positional_encod:[20,25],build_transform:[20,23,25],built:[44,45,48],c:[13,31,35,36,37,39,40,42,44,46,47,50,51,52,53],cach:[15,16],cake:47,calc_map:[16,56],calcul:[15,16,21,29,33,39,40,41,56],calib:13,calibr:[13,30,40,49],call:[1,13,15,16,17,19,23,24,26,29,35,36,37,43,47,50,51,52,53],callabl:28,callback:[1,14,18,22,27,49],caluc:41,cam:37,cam_extrins:[13,33,37,40,43],cam_intrins:[13,33,40,43],cam_intris:13,camera:[6,9,10,11,13,30,33,35,40,44,46,48,49],camera_calib:[33,34,43],camera_id:13,camera_label:13,camera_paramet:13,camera_sid:35,cameraextrins:[13,33,34,40,43],cameraintrins:[13,33,34,40,43],camextrins:13,camintrins:13,can:[3,7,13,19,21,24,25,31,35,36,37,39,40,41,42,44,45,46,47,48,50,51,52,53,54,56,57,58],cap:53,car:47,care:[17,19,23,35,36,47],carrot:47,cat:[16,43,46,47,50,54],cat_label:16,categori:[3,16,21,26,39,50,53],caution:46,cave_2:11,cave_4:11,cd:48,cdist:47,cell:[47,51,52],center:[17,25,33,41],center_i:[20,25,28],center_x:[20,25,28],chair:[5,47],chairssdhom:[0,49],chairssdhom_dataset:2,chairssdhomdataset:[0,36,49],chang:[0,21,24,25,28,36,37,44,45,46,47,49,50,51,52,53,54,56,57,58],channel:[37,42],check:[4,13,16,20,26,40,48,53,54,56,57,58],checkout:37,checkpoint:[26,29,53,54,55],child:[1,12,43,54],choic:26,choos:[33,41,54],choosen:41,ckpt:[20,25,28],clasif:[21,26],class_emb:[20,25,28],class_nam:16,classif:[16,20,21,25,26,28,37,50,52],classifi:[21,26,50],clean:[9,10,11],clip:[26,29,44],clock:47,clockwis:41,clone:31,close:53,co:47,coco2017:[50,56,57,58],coco:[0,20,24,25,28,44,45,49,51,52,53,54,56,57,58],coco_dataset:[3,46,50],coco_detection2detr:24,coco_detection_dataset:3,coco_detr:[56,57],coco_load:[45,53,54,56,57,58],coco_panoptic2detr:24,coco_train:45,coco_v:45,cocodataset:[46,58],cocodetect:3,cocodetection2detr:[22,45,53,54,56,57,58],cocodetectiondataset:[3,16,24,44,45,46,47,50],cocopanoptic2detr:[22,49],cocopanopticdataset:[16,24],code:[35,36,40,46,50,53,54,56,57,58],coeffici:[26,29],coffe:[51,52],coher:47,color:[32,35,36,39,41,42,46],color_by_cat:39,colorjitt:12,colors_r:12,com:[15,48],combin:[11,25,46,56],come:15,command:[3,57],common:[46,50,51,52,53,54,56,57,58],common_dim_nam:31,compar:[1,7,46,51,52,56],comparison:[49,51,52],compat:[25,48],complementari:45,complet:[13,39,53,54],complex:[54,57,58],compon:45,compos:[12,44,45],compress:13,comput:[2,5,6,12,15,16,17,19,21,23,26,32,41,44,45,48,49,50,56,57,58],compute_map:16,compute_per_size_ap:16,compute_pq:16,compute_statistical_metr:26,concat:[44,46,47],concaten:[26,29,51,52,56,57,58],concept:[45,50],conclud:56,confer:[9,10],config:[1,3,11,49],configur:[3,20,21,24,26,28,48,50,53,54],configure_optim:[21,26],conflict_handl:[56,57,58],conjunct:54,connect:[24,27,28,45],connector:[14,22,49],consid:15,consist:[35,46],consol:[16,57],constant:[51,52],constructor:25,consult:[57,58],contain:[2,6,7,9,13,15,20,21,23,25,26,28,33,35,36,40,50,51,52,53],content:[31,35,47,50],contrast:12,convent:[30,33],converg:54,convers:13,convert:[13,31,33,41,46,47,51,52],convolut:20,cooordin:34,coordin:[13,20,21,25,26,28,33,34,36,40,41,50],copi:[23,35,53],core:[24,26,47],corner:41,correct:[3,26],correctli:[16,35,36,52],corredl:26,correspond:[2,3,5,26,28,50,51,52],cost:[21,26,56,57,58],cost_box:[21,26],cost_class:[21,26],cost_giou:[21,26],cottontail:3,couch:47,could:[1,13,31,32,41,42,47,50,56,57,58],counter:41,coupl:[28,29],cow:47,cpu:[16,25,28,31,53,54,56,57,58],crash:[51,52],creat:[13,19,20,24,25,26,28,30,32,33,37,38,39,41,42,43,44,45,46,47,50,51,52,53,54,56,58],creation:[35,36],criterion:[18,22,29],crop:[31,40,44,46,47],crope:31,cropped_fram:[40,46],cross:[16,26,29],cross_entropi:15,cross_entropy_valu:15,crowd:[0,44,49],crowd_fram:44,crowd_human_dataset:4,crowdhumandataset:[4,44],csv:50,cuda:[16,19,20,31,33,40,41,48,53,54,56,57,58],cup:[47,51,52],current:[1,11,15,31,32,37,42,46,47],custom:[3,20,25,26,28,29,46,49,52,53,54,56,57,58],customdataset:50,customdatasetsplitmix:50,customrandomcolor:12,cv2:[35,36,53],cvpr2016:6,cvpr2017:2,cyclist:13,d:50,d_ffn:19,d_model:[19,23],data2detr:22,data:[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,22,26,28,29,35,36,37,40,42,44,46,47,48,49,53,54,56,58],data_load:[26,29,45,53,54,56,57,58],data_modul:24,data_point:16,databas:[0,24,49,50,53],dataload:[1,7,15,24,26,29],dataloader_idx:[15,26,29],datamodul:[24,45],datapipelin:31,dataset1:7,dataset2:7,dataset:[0,2,4,5,6,8,9,10,11,15,16,20,24,25,28,31,35,36,37,40,48,49,53,54,56,57,58],dataset_1:50,dataset_2:50,dataset_dir:[1,6,11,13,50],dataset_dir_prepar:4,dataset_n:50,dataset_name_1:50,dataset_name_2:50,dataset_name_n:50,datasets_download_path:1,datataset:50,date:47,dc5:20,ddp:24,dec_lay:20,dec_n_point:[19,20],dec_output:[25,28],decid:[3,25,28,50],declar:[3,24,58],decod:[19,20,21,23,25,28,51],decoder_forward:[19,23],decoder_lay:[19,23],decoder_layer_forward:[19,23],decoder_output:[19,23],decompress:13,decreas:32,deep:[48,50],def:[44,45,50,51,52,53,56,57],defin:[17,19,20,23,25,26,28,29,39,41,46,50,51,52,53,54,56,58],definit:[16,53,56,57,58],deform:[14,19,21,23,29,45,48,49,50,51,54,57,58],deformabe_finetun:53,deformable_detr:[19,20,21,48,51,52,53,56],deformable_detr_r50:20,deformable_detr_r50_finetun:20,deformable_detr_r50_refin:20,deformable_transform:19,deformablecriterion:21,deformabledetr:[20,51,52],deformabledetrhungarianmatch:21,deformabledetrinfer:52,deformabledetrr50:[20,21],deformabledetrr50finetun:[20,56],deformabledetrr50refin:[20,21,51,52],deformabledetrr50refinementfinetun:[20,51,52,53],deformabledetrtrtexport:[51,52],deformabletransform:19,deformabletransformerdecod:19,deformabletransformerdecoderlay:19,deformabletransformerencod:19,deformabletransformerencoderlay:19,deformanbl:49,demand:[56,57,58],demonstr:40,depend:[13,47,50],deploi:[48,53],deploy:[51,52],deriv:10,describ:[25,26,50,56,57,58],descript:[16,42],deseri:[51,52],design:[46,48],desir:[3,12,16,32,33,39,41,42,53,54],destroyallwindow:53,detail:[9,10,13,19,20,40,42,50,56],detect:[0,15,20,21,25,26,28,44,49,50,53,54,56,57,58],detections_set:8,detector:[14,49],detr50:[49,57,58],detr:[14,16,17,19,21,23,26,28,29,45,48,49,50,58],detr_50:[26,29],detr_finetun:[53,54],detr_model:28,detr_modul:[28,58],detr_panopt:[16,28,29,58],detr_r50:[25,29],detr_r50_finetun:25,detr_r50_panoptic_finetun:28,detrcriterion:[21,26,29],detrfinetun:28,detrhungarianmatch:[21,26],detrinfer:51,detrobjectdetectorcallback:26,detrr50:[16,20,25,27,51,53,54,56],detrr50finetun:[25,28,51,54,57,58],detrr50panopticfinetun:[28,58],detrtrtexport:51,develop:[16,50,56,57,58],devic:[16,19,20,25,26,28,31,33,40,41,51,52,53,54,56,57,58],diagram:28,dice:29,dict:[1,2,13,15,16,20,21,25,26,28,29,31,32,33,37,39,41,42,43],dict_input:4,dictionari:[13,16,26,29],dictionnari:[20,25,28],did:[32,42],differ:[7,15,25,26,29,35,36,39,43,44,47,50,51,52,53,54,56,57,58],dilat:[19,20,23,25],dim:[21,25,26,43,46,47],dim_feedforward:[19,20,23,25],dim_valu:31,dimens:[1,7,25,26,29,31,33,34,37,39,42,43,44,47,51,52],dimension:[37,42],dine:47,dir:[4,13],directli:[1,26,58],directori:[1,2,3,4,11,13,50,51,52],dirnam:[51,52],disk:[4,13,44],disp:[6,9,11,35],disp_flip:35,disp_format:35,disp_label:11,disp_occ:[6,9,11],disp_sign:35,disp_tensor:35,dispar:[0,6,30,37,39,46,48,49],displai:[20,25,28,32,35,36,38,39,40,41,42,44,46,47,48],dispnet:6,disregard:[20,25,28],dist:47,distanc:35,document:[40,45,50,57,58],doe:[3,13,15,47,50,58],dog:[47,50,54],don:[1,26,29,46,47,51,52],done:[35,36,45,48,50],donut:47,down:4,download:[0,1,2,5,6,9,10,13,46,49,50,53,56,57,58],download_sampl:1,downward:[13,33],dpi:58,dpm:8,dpm_sequenc:8,drasticli:4,drier:47,drop_label:31,dropout:[19,20,23,25],dtype:[31,51,52],due:[15,53,54],dure:[15,35,36],dx:[33,37],dy:[33,37],dynam:[24,37,42],dz:[33,37],each:[2,3,6,7,9,10,11,13,15,16,19,20,21,24,25,26,28,29,31,32,33,37,39,41,43,44,46,50],eas:49,easi:[21,30,44,46],easier:[37,48],easili:[40,57],eccv2018:5,edit:[15,29],edu:32,effect:[1,4],effici:26,either:[1,15,20,21,24,33,51,52],element:[1,3,9,10,15,16,21,24,25,26,28,32,37,42,43,50],eleph:47,els:[16,53,54,56,57,58],emb:[25,26,29],embed:[17,19,23,53,54],empti:[13,16,42],en:45,enabl:[53,54],enc_lay:20,enc_n_point:[19,20],enc_output:[25,28],enclos:[33,40,41,42],enclosing_boxes_2d:40,enclosing_typ:[33,41],encod:[14,19,23,24,25,28,38,42,47,49,50],encoder_lay:[19,23],end:[6,13,15,23,25,26,28,29,31,56,57],endup:47,engin:[51,52],engine_path:[51,52],enough:[43,51,52],ensur:19,enter:[24,26,29,57],entir:[3,56,57,58],entri:[21,26],entropi:[26,29],enumer:[1,44,56],environ:48,eo:[26,29],eos_coef:[21,26,29],ep:23,epoch:[15,26,29,53,54],equal:[3,12,21,25,42],equival:31,error:[15,21,53,54],estim:16,etc:13,eval:[16,51,52,56,57,58],evalu:[10,16,45],even:[35,36,37,42,44],everi:[15,17,19,23,24,26,29,35,36],everyth:46,exampl:[3,15,21,22,31,35,36,37,39,40,42,44,45,46,47,48,50,51,52,56,57,58],example_label:50,example_label_nam:50,except:[1,16,20,21,24,26,29,39],exclud:43,execut:[3,13,16,50,51,52],exist:[3,35,36,37,39,50],exmpl:45,expe_nam:[26,29,45,53,54,56,57,58],expect:[13,15,28,37,45,47],experi:[26,29],experiment:46,explain:[50,51,52,53,54,56,57,58],explan:[42,43,56],explicit:35,explicitli:57,explor:49,expon:21,export_engin:[51,52],expos:[3,44,45],express:35,ext:6,extend:[46,47,48],extens:[9,45,54],extra:[23,35,36],extract:[13,20,25],extrins:[13,34,40],ey:[33,40],f:[7,29,44,56],face:53,facilit:[46,48,56,57,58],fact:[51,52],factor:21,fake:50,fals:[1,3,6,8,12,16,17,19,20,23,24,25,26,28,31,32,33,37,39,41,42,43,45,47,50,51,52],far:46,fast:54,faster:1,fastest:[33,41],featur:[17,19,23,28,30,33,37,46,49,50],feedfoward:25,ffn:17,figsiz:40,figur:[40,58],file:[1,3,4,11,13,20,24,25,28,35,36,39,48,49,51,52,53,54],filepath:2,fill:24,film:10,filter:[0,20,21,24,25,28,33,47,49],find:[2,33,41,54],finetun:[3,18,22,27,48,49,51,52,56,57,58],finish:[56,57,58],fire:47,first:[7,12,15,16,28,33,40,44,45,46,47,50],fit:[24,45],fit_to_padded_s:[32,42],five:56,fix:[23,24,53,54,57],fix_classes_len:[3,24],flip:[31,32,33,35,36,40,44,46,47],flipped_fram:[40,46,47],flo:36,float16:[51,52],float32:[50,51,52],flow:[0,2,5,6,9,30,35,37,39,44,46,48,49],flow_backward:6,flow_flip:36,flow_forward:36,flow_label:11,flow_occ:[6,10,11],flow_occ_backward:6,flow_tensor:36,flownet2:6,fly:[5,13,32,42],flying_chairs2_dataset:5,flyingchairs2:[0,49],flyingchairs2dataset:5,flyingthings3d:[0,49],flyingthings3d_subset_dataset:6,flyingthings3dsubsetdataset:[0,7,49],fn:16,focal:[21,29,34],focal_alpha:21,folder:[4,6,13,24,50,53,54],follow:[2,5,6,9,10,13,24,25,28,32,42,44,45,46,47,48,50,53,54,56,58],footprint:[51,52],forc:[33,41],fork:47,form:34,format:[3,13,26,32,33,35,41,42,46,50,51,52],former:[17,19,23],forward:[13,15,17,19,20,21,23,25,26,28,29,31,33,53,54,56,57],forward_class_head:[20,25],forward_ffn:19,forward_head:[20,25],forward_out:[20,25,28,51,52],forward_position_head:[20,25],forward_post:23,forward_pr:23,found:[3,13,44,50],fp16:[51,52],fp32:[51,52],fp:16,fpath:4,frame0:3,frame:[1,2,3,5,6,9,10,11,12,13,15,16,20,21,23,24,25,26,28,29,30,32,33,35,36,38,39,40,41,42,43,44,45,47,48,49,51,52,53,54,56,57,58],frame_01:[37,46],frame_1:46,frame_255:[37,46],frame_2:46,frame_3:46,frame_4:46,frame_flip:[35,36],frame_minmax_sym:37,frame_process:53,frame_resnet:[37,46],frame_s:[32,33,41,42,47],framei:32,framer:3,framework:[50,56,57,58],framr:47,frane:[37,42],frcnn:8,frcnn_sequenc:8,freez:28,freeze_detr:28,freiburg:[2,5,6],frisbe:47,from:[1,3,4,6,7,10,12,13,15,16,17,19,20,23,24,25,26,27,28,29,31,32,33,35,36,37,39,40,41,43,44,45,46,47,50,53,54,55,56,57,58],front:[13,37,44],front_fram:13,front_left:[13,40],front_right:[13,40],frozen:[19,23],frozenbatchnorm2d:23,ftmap_mask:17,full:[13,45,47,54],fulli:[1,13],func:31,funetunig:[53,54],future_past:6,fx:34,fy:34,g:41,gamma:21,gamma_r:12,gaussian_std:12,gcc:48,gen_encoder_output_propos:19,gener:[1,2,3,4,5,6,7,8,9,10,11,13,16,17,26,32,46,50],get:[1,3,7,12,13,15,16,19,20,24,25,26,28,29,31,32,33,35,36,37,39,40,41,42,44,46,49,50,53,54],get_dataset_dir:[1,11],get_enclosing_box_2d:[33,40],get_expe_info:[51,52],get_featlabels_el:6,get_features_label:6,get_file_id:2,get_filter_fn:28,get_fram:[2,6,13],get_frame_boxes2d:13,get_frame_boxes3d:13,get_frame_camera_paramet:13,get_label:31,get_loss:26,get_metr:[16,21,26],get_outs_filt:[20,25,28,51,52],get_outs_label:[20,25],get_proposal_pos_emb:19,get_rank:[15,23],get_reference_point:19,get_seg:13,get_sequ:2,get_size_with_aspect_ratio:12,get_slic:31,get_split_fold:50,get_statistical_metr:26,get_valid_ratio:19,get_vertices_3d:[33,40],get_vertices_3d_proj:[33,40],get_view:[3,20,24,25,28,32,33,34,35,36,37,38,39,40,41,42,43,44,46,47,50,51,52,53,54,56,57,58],get_with_format:[32,42],get_world_s:23,getitem:[1,2,3,4,5,6,8,11,13,37,44,46,47,50],getitem_ignore_error:1,giou3d:[33,40],giou3d_with:[33,40],giou:[21,26,29,30,32,33,41,49],giou_tensor:32,giou_with:32,giraff:47,github:34,give:[40,47,50,51,52],given:[1,3,11,12,13,15,16,20,25,26,28,29,32,37,41,42,43,50,54,56,57,58],glass:47,global:[7,17,33],global_color_set:[32,41],glove:47,go:4,goal:[53,54,56,57,58],gonna:47,good:24,gpu:[28,31,40,51,52],grab:[51,52],grad:[26,29],gradient:47,gradient_clip_v:[26,29,56,57],graph:[51,52],great:[44,46],grid:43,grid_siz:43,ground:[13,21,26,47,53,54,56,57,58],group:[2,5,6,26,29],gt:[15,16,21,26,29,40,46],gt_box:[15,16,29,53,54,56,57,58],gt_boxes_2d:[13,26],gt_boxes_3d:[13,40],gt_mask:[15,16,29,58],guid:58,h:[13,19,20,25,28,31,33,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53],h_crop:31,ha:[3,4,35,36,37,50,53,54,56,57,58],hacki:31,hair:47,hand:[50,58],handbag:47,handl:[3,13,39,47,48,50],happen:[21,26],hard:21,hardwar:[13,48],hasattr:45,have:[1,3,7,9,10,13,16,26,28,31,32,34,35,36,41,42,44,46,47,50,51,52,57,58],hb:25,head:[20,21,25,27,33,37,49,52],height:[12,19,20,25,28,32,37,41,42,50],hello:[1,4,8,11,12,49],help:[44,48],helper:[45,51,52],helpful:[20,25,28],here:[1,4,8,9,10,11,12,15,44,45,46,53],hflip:[31,35,36,40,46,47],hidden:25,hidden_dim:[17,20,25],high:[50,56,57,58],hint:[53,54,56],histogram:15,hold:26,home:[1,11,50,53,54],homogen:34,hook:[15,17,19,23,24],horizont:[31,32,33,35,36,40,46],hors:47,hot:[47,50],hour:[13,56,57,58],how:[3,20,21,25,26,28,40,47,49,53,54,56,57,58],howev:[1,20,29,50,53,56,57],html:45,http:[19,20,25,32,34,45,48,58],hue:12,huge:4,human:[0,44,49],humancrowd:44,hungarian:[21,26],hungarian_cost_class:[21,26],hungarian_cost_giou_box:26,hungarian_cost_l1_box:26,hungarina:26,hw:[40,43],hydrant:47,id:[2,3,13,15,25,26,29,38,39,47,50,53,54,56,57,58],id_sf_path:6,idea:47,ideal:[3,34],ident:[51,52],idstring2int:13,idx:[1,2,3,4,5,6,8,11,13,50],ie:25,ignor:[17,19,23,41],ignore_error:1,illustr:47,imag:[3,4,11,13,15,16,17,20,24,25,28,33,34,35,36,37,40,44,47,50,51,52,53,54],image_fold:50,image_path:[25,28,51,52],image_s:50,imagex:13,img:[51,52],img_ann_train_0:58,img_ann_train_1:58,img_ann_train_l:58,img_ann_val_0:58,img_ann_val_1:58,img_ann_val_m:58,img_fold:[3,4,45,50],img_test_0:50,img_test_1:50,img_test_n:50,img_train_0:[50,58],img_train_1:[50,58],img_train_l:[50,58],img_val_0:[50,58],img_val_1:[50,58],img_val_m:[50,58],implement:[1,15,25,27,28,29,43,44,48,51,52,57,58],impli:20,imposs:53,improv:[48,53],imshow:44,includ:[20,25,26,28,45],include_flow:6,include_vertice_behind_image_plan:33,increas:20,inde:42,independ:[13,48],index:[1,3,5,24,31,45,48],index_i:26,index_j:26,indic:[21,26,32],individu:[3,20,25,26,28,50],infer:[15,16,20,21,25,26,28,29,45,49],infernec:15,info:56,inform:[2,5,6,9,10,15,16,20,21,24,25,26,28,34,42,47,50,53,57,58],inherit:57,inhert:[15,17,24,29,31],init:[16,45],init_data_object:16,initi:[3,35,36,39,54,56,57,58],inlin:[53,54,56,57,58],inplac:[31,32],input:[19,21,23,24,25,26,28,39,44,50,51,52,53,56,57,58],input_dim:17,input_mean_std:[20,25,28],input_nam:[51,52],input_shap:[51,52],insid:[15,23],instal:[13,46,49],instanc:[17,19,23,33,37,39,46,47,56],instances_train2017:[24,45],instances_val2017:[3,24,45,50],instancesbasemetricscallback:15,instanti:[20,25,51,52,56,57,58],instead:[1,15,17,19,23,24,47],instruct:[2,5,6],int64:32,intanc:[15,29],integ:44,integr:48,intermedi:[25,51,52],intern:13,internet:50,interpret:[35,50],into_futur:6,intrins:[13,34,40],introduc:45,invalid:1,invari:26,invers:33,invert:35,io:[34,45],iou3d:[33,40],iou3d_with:[33,40],iou:[15,16,29,30,32,33,39,41,47,49],iou_tensor:32,iou_threshold:[16,32],iou_with:[32,39,47],is_avail:[16,53,54,56,57,58],is_dist_avail_and_initi:[15,23],is_empti:16,is_main_process:23,is_tru:16,isopen:53,issu:[3,47],isth:16,isthing_label:16,item:[1,7,13,50,56],iter:[3,7,20,24,35,36,44,52,53,54,56,57,58],its:[3,20,25,28,31,35,36,40,44,46,47,48,50,51,52,56,57,58],itself:39,join:[50,51,52],joiner:[19,23],joint:[11,28],jpeg:[13,37],jpg:[37,50,58],json:[1,3,4,11,13,24,45,49,53,58],just:[50,53,54],k:[26,29],keep:[16,35,42,43,44,46,47,53,54],kei:[13,15,20,25,26,28,29,31,50,51,52],kept:32,keyboard:47,keyword:15,kill:13,kite:47,knife:47,know:[28,45,46,47,53,54,56,57,58],knowledg:58,known:[35,47],ksimek:34,kwarg:[1,2,3,4,5,6,8,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,32,33,34,35,36,37,38,39,41,42,43,50,51,52],l1:[21,26],l:40,label:[3,6,9,10,11,13,16,20,21,24,25,26,28,29,30,31,32,33,35,36,37,39,40,41,42,43,46,49],label_nam:24,labels_nam:[16,38,47,50],labels_set:[32,39,41,42,44],labels_set_1:42,labels_set_2:42,lambda:[7,50],laptop:47,largest:12,last:[7,20,25,33,34,53,54],later:[1,35,36,45],latest:[26,29,45],latter:[17,19,23,46,47,48],layer:[17,19,20,21,23,25,26,28,29,53,54],learn:[25,26,42,46,48,50,53,54,56,57],least:[21,26,28,42],left:[6,7,9,10,11,12,13,35],left_offset:31,left_right:6,len:[13,15,20,21,26,32,46,50,53,54],length:34,less:42,let:[37,46,47,51,52,57,58],level:[19,49,50,56],level_start_index:19,librari:48,lidar_label:13,lig:57,light:[15,47],lightn:[15,24,26,29,45,48,51,52,53,54,56,57,58],lightningdatamodul:[24,45,54],lightningmodul:[15,26,54,56],ligntn:45,like:[42,50,56,57],limit:1,limit_train_batch:54,limit_val_batch:54,line:[3,13,35,36,40,47,56,57,58],link:[2,5,6,9,10,47,53,54],list:[1,3,6,7,9,10,11,12,13,15,16,19,20,21,23,24,25,26,28,29,31,38,42,43,44,46,48,50],listdir:50,lit_deform:[53,56],lit_detr:[54,56,57,58],lit_model:[51,52,56],lit_panopt:58,litdeformabledetr:[21,52,53,56],litdetr:[21,24,26,29,45,49,51,54,56,58],litpanopticdetr:[16,24,29,49],ll:47,ln:23,load:[2,9,10,13,15,20,24,25,26,28,29,35,45,49,50,53,54,55,56,57,58],load_gt:4,load_json_lin:4,load_train:[51,52,53,54,56,57,58],load_trt_plugins_for_deformable_detr:52,load_weight:[51,52],loader:[1,7,26,29,45,49],local:[33,53],locat:[31,37],log:[15,26,29],log_boxes_2d:15,log_boxes_3d:15,log_every_n_step:15,log_mask:15,logic:[51,52,58],logit:[20,21,25,26,28],longer:[4,13],look:[1,11,31,47],loop:[26,29,53,54],loss:[15,20,21,25,26,28,29],loss_box:26,loss_boxes_weight:[21,26,29],loss_ce_weight:[26,29],loss_dice_weight:29,loss_focal_weight:29,loss_giou_weight:[21,26,29],loss_label:[21,26],loss_label_weight:21,loss_valu:15,lot:4,lsap:26,lt:[40,46],m:[32,33,39],m_input:[51,52],m_output:[15,20,21,25,26,28,29],machin:[46,50],made:[17,20,25,28,44,50],mai:26,main:[4,8,9,10,13,17,25,28],mainli:48,make:[1,7,24,25,37,44,46,47,48,49],manag:[3,13,50],mandatori:[50,54],mani:50,manipul:3,manual:50,map:[13,16,17,19,23,27,35,36,37],map_metrics_callback:[15,29],market_2:11,market_5:11,market_6:11,mask2:39,mask2id:39,mask:[15,16,17,19,20,23,24,25,26,28,29,30,35,36,37,43,44,49,51,52,53,58],mask_float_tensor:39,mask_id:39,mask_tensor:39,maskth:28,match:[21,26,42,46,50,58],matcher:[18,22,29],matcher_fram:26,matplotlib:[35,36,40,46,51,52,53,54,56,57,58],matric:33,matrix:[13,32,34,39],matter:46,max:[1,43],max_epoch:[53,54],max_retry_on_error:1,max_siz:[12,44,45],maxim:25,maximum:12,mayb:31,mean:[20,26,29,37,44,46],mean_std:37,mean_std_norm:[37,46],memor:42,memori:[19,23,31,45,51,52,58],memory_format:31,memory_key_padding_mask:23,memory_mask:23,memory_padding_mask:19,merg:[0,13,44,49],merge_dataset:7,mergeabl:42,mergedataset:[0,49],meter:47,method:[1,4,11,12,13,15,17,20,24,25,26,28,29,32,37,42,43,44,45,46,47,51,52,57],metric:[14,21,26,29,49,56],metrics_callback:15,metricscallback:15,mhattent:23,microwav:47,middl:13,might:[15,29,37,42,44,47,50,51,52],min:26,min_siz:12,mind:[53,54],minim:[0,16,40,49,50],minimum:3,minmax_sym:37,minor:[51,52],misc:23,mistak:54,mlp:[14,25,49],mode:[33,40,50,58],model:[14,15,16,18,21,22,23,24,26,27,29,34,44,47,48,49,50,51,52,53,54,57,58],model_nam:[25,26,29,51,52,56,57,58],model_precis:[51,52],modif:23,modifi:[31,35,36,56,57,58],modul:[3,13,15,17,19,20,21,23,24,25,26,27,28,29,49,51,52,53,54,56,57],monitor:[21,26],more:[2,4,5,6,9,10,13,15,17,19,20,24,26,34,37,40,42,44,45,50,53,54,56,57,58],moreov:[56,57,58],most:[34,44,48,57],mot17:[0,44,49],mot_dataset:44,motorcycl:47,mountain_1:11,mous:47,move:15,mpi:[9,10,11,35],much:[21,26],multi:[0,17,20,25,31,37,48,49],multilabel:39,multipl:[3,7,32,39,41,42,44,48,50,57],multipli:39,multiscal:[19,52],must:[1,3,11,13,15,16,17,20,24,28,31,32,33,35,36,37,39,42,43,47,50,53],my_dataset:50,my_detr_model:58,my_fram:46,my_model:[56,57,58],my_norm:[37,46],my_set:37,n:[13,15,23,24,32,33,37,39,41,42,46,47],n_box:47,n_frame:12,n_head:[19,23],n_level:19,n_point:[19,42],n_tensor:31,naiv:[33,41],name:[1,3,11,13,15,19,20,23,24,25,26,28,29,31,32,33,35,36,37,38,39,40,41,42,43,44,46,47,50,53,56],name_x_axi:15,name_y_axi:15,namespac:[20,24,25,26,29,45,53,54,56,57,58],nan:23,nb:6,ndarrai:13,ne:31,nearli:[51,52],necessari:[1,16,35,50],necessit:[35,36],need:[1,3,13,15,17,19,20,23,24,31,37,42,46,47,48,50,51,52,53,54,58],neg:[21,26,31,33,35],network:[16,26,27,28,56,57,58],neural:27,neuron:25,new_fram:37,next:[1,3,7,24,31,35,36,53,54,56,57,58],ngc:48,nhead:[19,20,23,25],nice:[46,47],nm:32,nn:[17,19,20,23,25,26,28,29],no_augment:24,no_grad:[51,52],no_object:26,no_suffix:[53,54],non:[20,26,42],none:[1,3,4,6,7,8,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,37,38,39,41,42,43,45,46,47,51,52],norm01:[37,46],norm255:[37,46],norm:[23,46],norm_a:37,norm_minmax_sym:37,norm_resnet:[20,25,28,37,45,46,51,52,53],normal:[17,20,25,28,37,50],normalize_befor:[23,25],normalized_fram:37,normnal:37,note:[1,4,13,15,16,19,21,24,26,29,31,37,42,43,44,46,47,53,56],notebook:[51,52],noth:[13,38],notimplementederror:25,now:[1,3,47,50,51,52,54,56],np:[13,37,39,40,41,42,51,52],np_boxes3d:13,np_convert_waymo_to_aloception_coordinate_system:13,num_box:[21,26],num_channel:23,num_class:[20,21,25,26,28,29,51,52,53,54,56,57,58],num_decoder_lay:[19,23,25],num_enc:25,num_encoder_lay:[19,23,25],num_feature_level:[19,20],num_lay:[17,19,23],num_level:19,num_point:19,num_pos_feat:17,num_posit:16,num_process:13,num_queri:[20,21,25,26,28],num_target_box:26,num_work:[1,7,24],number:[1,7,9,10,12,16,20,21,24,25,26,28,29,37,42,51,52,53,54,58],nummer:24,numpi:[35,36,37,40,41,42,51,52],nvidia:48,object:[3,12,14,16,20,21,24,25,26,28,29,30,31,37,39,46,49,50,53,54,56,57,58],object_detector_callback:[15,26,29],objectdetectorcallback:[15,26,29],objectness_recal:26,objectness_true_po:26,objects_class:44,objects_id:44,occlus:[6,9,35,36,39],off:16,offer:50,offici:[23,51,52],offset:[34,35],offset_i:31,offset_x:31,old_path_directori:3,omit:26,on_list:31,on_train_batch_end:[15,26,29],on_train_epoch_end:15,on_validation_batch_end:15,on_validation_end:15,on_validation_epoch_end:[15,26,29],onc:[4,13,15,16,29,40,48,53,54,56,57,58],one:[1,3,4,11,13,15,16,17,19,20,23,24,26,28,29,32,35,36,37,39,40,42,44,45,46,47,50,53,54,57],ones:[35,57],oni:1,onli:[12,19,20,21,25,26,28,29,31,33,35,40,42,47,50,51,52,53,54],onlin:[57,58],onnx:[48,51,52],onnx_graphsurgeon:48,onnx_path:[51,52],onto:40,op:[23,48],open:[10,13,40,53],opencv:46,oper:[1,32,42,47],optic:[0,2,5,6,11,36,46,48,49],optim:[26,29],optimzi:21,optin:20,option:[1,3,13,16,20,21,24,25,26,28,29,32,33,35,36,37,38,39,41,42,43,45,49,50],orang:47,order:[3,15,25,26,28,32,33,41,43,50,51,52,53,54,56,57,58],org:[19,20,25,58],organ:50,orient:[33,40,41],oriented_boxes_2d:41,orientedboxes2d:[30,33,49],origin:[1,3,4,13,25,28,54],os:[50,51,52],other:[9,10,23,25,26,30,32,33,35,36,37,39,41,42,43,46,47,49,50,51,52,57,58],otherwis:[4,12,13,21,32,35,36,37,39,41,42,43],othwerwis:15,othwewis:37,ouput:[26,29],our:[45,47,51,52],out:[21,33,44,54],out_scor:[20,25],outptu:[20,25,26],output:[15,20,21,25,26,27,28,29,31,51,52,58],output_dim:17,outs_label:[20,25],outs_scor:[20,25],oven:47,over:[7,15,16,47],overridden:[17,19,23],overview:46,overwrit:[3,21],own:[20,25,46],p:[12,47],p_bbox:16,p_mask:16,p_t:21,packag:[12,13,45,47,48],pad:[20,25,28,31,32,42,43,44,51,52],pad_box:43,padd:[32,44],padded_boxes2d:32,padded_point:42,padding_mask:19,page:[1,4,8,11,12,50],paht:[50,53],pair:[13,21,26,41],pairwis:32,panopt:[14,15,16,29,39,49],panoptic_train2017:58,panoptic_val2017:58,panopticapmetricscallback:29,panoptichead:[28,58],panopticobjectdetectorcallback:29,paper:[2,5,6,9,10,17,20,25,28],param1:12,param2:12,param:[12,41],paramet:[1,3,5,6,7,9,10,11,12,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,37,38,39,41,42,43,46,48,50,56,57,58],params2namespac:[56,57],parent:[26,29,57],parent_pars:[21,24,26,29],park:47,pars:[2,13],parse_arg:[53,54,56,57,58],parser:[21,24,26,29,53,54,56,57,58],part:[2,5,6,42,45,50],partial:39,particular:[44,45],pascal2012:50,pascal:50,pass:[9,10,11,17,19,20,23,25,26,28,31,32,37,42,50],past:23,path:[1,3,4,11,13,25,26,29,35,36,37,39,50,51,52,53,56,57,58],path_to_imag:[20,25],pca:[33,41],pedestrian:13,peopl:3,people_mask:53,per:[16,28],percentag:[12,21,26,31],percentah:31,perceptron:[17,25],perform:[3,17,19,21,23,25,26,32,33,41,47,49,50],person:[3,47,50,52],person_box:47,perspect:34,pet:54,pfm:35,phone:47,pi:[40,41],pickl:13,pinhol:34,pip:48,pipelin:[46,48,49,56,57,58],pixel:[20,25,33,35,39,51,52],pizza:47,pkl:13,pl:[15,45,54],pl_helper:[51,52],pl_modul:[15,26,29],plai:[47,49],plan:33,plant:47,pleas:[2,3,5,6,13,45,48,57,58],plot:44,plt:[51,52,58],plugin:52,png:[35,39],png_negat:35,po:[19,23,47],point2d:42,point:[19,34,37,42,44,47],points2d:[30,37,49],points_2d:[37,42],points_2d_ab:42,points_2d_rel:42,points_2d_xi:42,points_2d_yx:42,points_format:42,popul:[13,33,40],popular:48,pos_emb:[19,23],pose:13,posit:[14,16,19,21,23,25,26,31,32,33,35,41,42,47,49],positin:25,position_embed:[19,23,25],position_encod:17,positionembeddingsin:17,possibl:[6,7,9,10,11,12,15,20,25,26,28,29,32,35,36,37,39,41,42,43,47,50,53,54],post:15,postprocess:[20,25,28],pot:47,potenti:43,pq:[14,49],pq_averag:16,pq_metrics_callback:15,pqmetric:[15,16],pqmetricscallback:15,pqstatcat:16,pre:[19,20,23,25,28,50,54,56,57,58],pre_process_tgt:[19,23],precis:[21,26,29,51,52,56],pred_box:[15,16,20,21,25,26,28,29,53,54,56,57,58],pred_logit:[20,21,25,26,28],pred_mask:[15,16,28,29,58],pred_masks_info:28,predefin:12,predict:[15,16,20,21,24,25,26,27,28,29,49,50,54,56,57,58],predictec:25,preds_box:15,preds_mask:15,prefer:47,prefix:15,preload:13,preloaded_calib:13,preloaded_labels_2d:13,preloaded_labels_3d:13,prepar:[0,1,4,49],prepr:13,preprocess:[20,24,25,28,53],present:[15,50,56],preserv:47,preserve_format:31,pretrain:[20,21,25,28,54,58],prevent:47,previou:[9,10,24,28,31,35,36,45,50,53,56,58],previous:58,princial:34,print:[13,16,35,40,41,44,46,47,53,54,56],print_bindings_info:[51,52],print_error:1,print_result:[16,56],procedur:[26,28],process:[13,15,16,19,21,23,24,26,29,49,50,53,54,56],produc:[23,54],program:47,progress:[15,21,26,29],project:[13,26,28,29,33,34,40,45,51,52,53,54,56,57,58],project_dir:[51,52],project_run_id:[51,52,53,54,56,57,58],promis:48,promiss:45,prompt:[3,50,53],properli:[1,11,15,37],properti:[1,16,43,44,46,47,50],propos:[19,26,44],proprli:43,provi:[32,42],provid:[2,5,6,20,25,32,41,42,44,45,47,50,56,57,58],psoition:17,pth:[20,25,28,55],pts2d:42,purpos:[3,48,50,53,54],purpus:28,push:16,py:[1,13,25,48],pylab:58,pypi:48,pyplot:[51,52],python:[13,48],pytorch:[1,3,15,24,26,28,29,31,37,45,47,48,50,51,52,53,54,56,57,58],pytorch_lightn:[15,24,26,29,45],qualit:[51,52],qualiti:[15,16],queri:[20,25,28],query_emb:[19,23],query_po:[19,23],quick:[3,44,46,50,51,52],quickli:[44,45,48],r11:34,r12:34,r13:34,r21:34,r22:34,r23:34,r31:34,r32:34,r33:34,r50:[16,18,22,26,28,29,51,52,53,54,56,57,58],r:[33,41,48,56],rabbit:[3,50],raccon:50,raccoon:50,racket:47,raft:45,rais:[1,16,20,24,25,26,28,29,39],rand:[39,50],random:[3,41,50,54,56,57,58],random_box:50,random_step:[8,13],randomcrop:[7,12],randomhorizontalflip:[12,44,45],randomli:13,randomresizewithaspectratio:[12,44,45],randomsampl:[1,7],randomselect:[12,44],randomsizecrop:[12,44],rang:[21,44],rank:37,rate:26,ratio:19,rcparam:[40,58],re:[42,46,47],reach:1,read:[3,6,13,20,25,28,35,36,50,53],read_sequ:13,readi:[39,44,48,57,58],readthedoc:45,real:[32,34,42],realisticnois:12,realli:[26,35,36],reason:[26,28,29,50,58],recal:[21,26],receiv:[15,28],recent:50,recip:[17,19,23],recognit:16,recommd:42,recommend:[25,50,56,57,58],recurs:[13,31],recursive_apply_on_labels_:31,reduc:[1,51,52],ref:48,refenr:[19,33],refer:[2,5,6,9,10,19,37,42,45],reference_point:19,refin:[18,52,53,56],reflect:26,refriger:47,regist:[17,19,23],regress:26,regular:47,rel:[3,20,21,25,26,28,32,35,41,42,47,50],rel_area:32,rel_po:[32,41,42],relat:[21,26,47],relationship:47,relative_to_absolut:43,releas:[31,53],relev:3,reli:[20,25,32,41,42,45,46],relu:[19,23],remain:[42,53],remot:47,remov:[1,23,42,47,56,57],remove_pad:[32,42],renam:31,rename_:31,rename_data_to_non:1,render:[3,20,24,25,28,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53,54,56,57,58],repeat:1,replac:[1,4,11,13,20,24,26,29,57],repositori:[23,51,52],repres:[13,20,25,28,35,36,39,42],represent:[28,32,42],requer:[24,28],requir:[3,13,24,25,28,37,46,48,50,52,53],rescal:43,research:48,reset_nam:31,resiz:[12,24,31,39,44,47,51,52],resnet50:[20,25],resnet:[19,20,23,25,28,37,46],resolv:[56,57,58],respect:[16,24,25,50,51,52,54],respons:47,rest:[24,57],restor:45,result:[15,26,51,52,53,54,56,57,58],resum:13,ret:53,ret_iou3d:[33,41],ret_iou:41,ret_union:[32,33,41],retri:1,retriev:[20,25,28],retry_offset:1,return_ann:39,return_bb_output:[20,25],return_cat:39,return_dec_output:[20,25,51],return_enc_output:[20,25],return_interm_lay:[19,20,23,25],return_intermedi:[19,23],return_intermediate_dec:[19,20,23],return_mask:[3,24],revers:35,rgb:[35,36,44,51,52],right:[1,6,7,9,11,13,33,35,41,44,46],right_offset:31,roboflow:50,robust:54,root:[3,13,50,51,52],root_dir:[51,52],rotat:[33,34,41],rotated_gi:41,rotated_giou_with:41,rotated_i:41,rotated_iou_with:41,rouhgli:21,rq:16,rqsrt:23,rtype:[1,13,20,21,33,39,41],run:[13,17,19,20,23,25,26,28,29,45,47,51,52,53,54,56,57,58],run_id:[53,54,55,56,57,58],run_id_dir:[51,52],run_train:[26,29,45,53,54,56,57,58],s:[13,15,34,37,46,47,48,50,51,52,57,58],sa_tensor:[32,43],said:[21,26],same:[7,16,21,24,31,32,33,34,39,41,42,43,44,51,52,53,54,58],same_on_fram:[12,24,45],same_on_sequ:[12,24,45],sampl:[0,1,2,7,12,13,15,16,19,24,26,29,35,36,37,40,45,46,47,49,50],sample_param:12,sampler:[1,7,24,45],sandwich:47,satur:12,save:[1,4,11,26,29,37,50,51,52,53,54,56,57,58],scalar:[15,21],scale:[17,20,44,45,48],scaler:15,scatter:15,scatter_nam:15,scene:[6,12],scheme:50,scienc:[2,5,6],scissor:47,score:[16,20,21,25,32,38],scratch:[50,56,57,58],script:[13,45,57],sdp:8,sdp_sequenc:8,seamless:47,second:[12,15,44,47,50],section:[2,5,6,40,44,50,56,57,58],see:[3,13,20,22,25,26,28,31,40,42,43,47,50,53,54,56,57,58],segment:[13,15,16,20,24,26,27,28,29,37,39,50,58],select:[3,13,20,25,26,28],selected_featur:6,selected_label:6,self:[1,4,12,13,19,20,23,33,45,50,51,52],send:31,sens:44,separ:15,sequenc:[2,5,8,9,10,13,15,24,26,29,40,44],sequence_data:[2,6],sequence_id:13,sequence_s:[7,9,10,13],sequence_skip:[9,10],sequenti:[7,19,23],serial:[51,52],serv:[51,52],set1:42,set2:42,set:[1,3,11,12,13,15,16,24,25,26,28,29,32,33,35,36,37,39,40,41,42,43,44,46,47,48,56],set_dataset_dir:[1,11],set_label:31,set_param:12,setup:[3,24,45,49,53,54],sever:[48,50,56,57,58],sh:48,shaman_2:11,shaman_3:11,shape:[13,16,19,20,21,25,28,33,34,40,41,44,46,47,51,52],sheep:47,shift:[1,31],shift_i:31,shift_x:31,shifted_tensor:31,shot_std:12,should:[3,13,17,19,23,32,34,35,36,42,47,48,50,57],show:[3,15,35,36,39,46,50,51,52,53,54],shown:54,shpae:25,shuffl:[1,7,44],side:[12,35],side_left:13,side_right:13,sigmoid:[20,21,29,52],sigmoid_focal_loss:21,sign:[13,30,47],silent:[17,19,23],similar:17,simpl:[17,33,41,46],simpli:[4,13,15,37,42,44,46],simplic:40,sin:25,sinc:[15,17,19,23,44,47,57],sine:25,singl:[3,15,25,50],sink:47,sintel:[0,35,49],sintel_disparity_dataset:9,sintel_flow_dataset:10,sintel_multi_dataset:11,sintel_sequ:11,sinteldisparitydataset:[0,35,49],sintelflowdataset:[0,49],sintelmultidataset:11,size:[1,7,9,10,12,13,15,16,20,21,24,25,26,28,29,31,32,33,37,39,40,41,42,43,44,46,47,50,51,52,53,56,57],skateboard:47,skew:34,ski:47,skip:[9,10,56,57,58],skipe:[4,13],sleeping_1:11,sleeping_2:11,slice:[31,40],slightli:[33,41,47],slot:[25,26],slot_true_neg:26,slow:4,slowest:[33,41],small:[3,46],smallest:[33,41],smilar:21,snapshot:53,snowboard:47,so:[12,46,47,50],softmax:[20,21,52,53],softmax_threshold:20,solv:26,some:[1,13,15,16,21,26,29,40,45,46,47,50,51,52,53,54,56,57,58],somehow:31,someon:46,someth:24,sort:[32,50],sourc:[10,13],space:[40,42],spatial:[30,31,44,46,48,49],spatial_augmented_tensor:[35,36,37,39,43],spatial_shap:19,spatial_shift:31,spatialaugmentedtensor:[35,36,37,39,42,43],spatialshift:12,specfic:[26,29],special:26,specif:[1,3,7,15,17,24,25,26,29,35,36,45,46,47,54,57],specifi:[51,52],speed:[13,50],split:[1,2,5,8,13,16,40,45,50,51,52],split_fold:[2,5,8,13,50],splitmixin:50,spoon:47,sport:47,sptial:43,sq:16,src:[19,23],src_key_padding_mask:23,src_level_start_index:19,src_mask:23,src_padding_mask:19,src_spatial_shap:19,src_valid_ratio:19,stack:23,stage:[21,24,25,26,29,45,46],standard:17,stanford:32,start:[15,31,37,41,53,54,57],stat:15,state:[32,47],stateful:49,statement:54,statist:[23,26],std:[37,46],step:[4,13,15,21,26,28,29,44,53],step_smooth:15,still:[21,26,46,48],stop:[13,47],storag:[1,53],store:[3,4,13,15,16,21,24,26,29,42,50,56,57,58],str:[1,11,13,15,20,21,24,25,26,28,29,32,33,35,36,37,38,39,41,42,43,45,50],straigthforward:[35,36],stream:[1,7,49],stream_load:[1,3,7,16,35,36,44,50],strength:40,strict:[25,28],strict_load_weight:[20,25,28],stride:20,string:[6,9,10,11,13],structur:[13,45,46,47,50],stuff:16,subclass:[17,19,23],subdirectori:13,subset:[0,49],suddenli:13,suit:45,suitcas:47,sum:19,summari:20,supervis:[21,26],support:[1,25,26,29,40,47],suppos:[20,25,28,42],sure:[44,46],surfboard:47,swap:35,synthet:[5,6],system:[13,33,48],t1:34,t2:34,t3:34,t:[1,6,13,15,26,29,31,33,37,40,42,44,45,46,47,51,52],t_bbox:16,t_co:[1,2,3,4,5,6,7,8,9,10,11,13],t_mask:16,tabl:[15,47],take:[13,17,19,23,26,29,46,47,50,51,52,53,57,58],taken:[3,28],tar:13,target:[12,16,21,26,31,37,42,45],target_fram:37,task:[15,26,28,29,50,53,58],teddi:47,tediou:[35,36],temperatur:17,temple_2:11,temple_3:11,tempor:[15,37,42,43,44],temporal_fram:43,temporarili:1,tenni:47,tensor:[1,13,15,17,19,20,21,23,25,26,28,32,33,34,35,36,37,38,39,40,41,42,44,45,46,48,49,50,51,52],tensorflow:50,tensorrt:[48,49,53],test2017:50,test:[1,8,13,24,44,48,50,51,52],test_experi:45,tfrecord:13,tgt:[19,23],tgt_box:[21,26],tgt_key_padding_mask:[19,23],tgt_mask:23,tgt_path:37,than:[7,23,26,31,37,42,48,58],thefor:[42,46],thei:53,theirs:[25,50],them:[15,17,19,23,24,29,48,50,57],therefor:[15,26,31,35,42,44,45,46,47,50,58],theta:41,thi:[1,2,3,4,5,6,7,9,10,11,12,13,15,16,17,19,21,23,24,25,26,28,29,31,32,33,34,35,36,37,40,41,42,43,44,45,47,48,50,51,52,53,54,56,57,58],thing:[16,37],think:50,third:50,those:[19,24,26,29,33,40],though:47,three:[45,50],threshold:[16,20,21,25,28,32],through:[1,4,26,31,40],throughout:[51,52],thu:[26,42],tie:47,time:[35,44,46,47,51,52,53],timestep:15,titl:[53,54,56,57,58],toaster:47,todo:[17,50],togeth:[7,47],toilet:47,tool:50,toolkit:48,toothbrush:47,top:[12,41,48,50],top_offset:31,torch2trt:[51,52],torch:[1,2,3,4,5,6,7,8,9,10,11,13,15,16,17,19,20,21,23,24,25,26,28,29,31,32,33,35,36,38,39,40,41,42,44,46,47,50,51,52,53,54,56,57,58],torch_m_output:[51,52],torch_model:[51,52],torch_pred_box:[51,52],torchvis:[12,20,23],total:[16,21,26,29],tp:16,track:[13,43,44,46,47],track_id:13,traffic:47,train2017:[24,45,50,58],train:[1,2,3,4,5,6,7,8,12,13,14,15,17,18,19,20,22,23,24,25,27,28,31,47,48,49,50,51,52],train_ann:[24,53],train_backbon:[19,20,23,25],train_dataload:[24,45],train_dataset:24,train_fold:[24,53],train_fram:24,train_load:[1,3,7,13,44,45,50],train_on_v:24,train_transform:[24,45],train_val:6,trainer:[15,24,26,29,49,57,58],trainig:45,training_epoch_end:15,training_sequ:8,training_step:[26,29],transform:[0,3,7,13,14,18,20,22,24,25,28,30,33,34,37,45,46,48,49,50,53,56,57],transform_fn:[1,7,44,50],transformer_outptu:[20,25],transformerdecod:[23,25],transformerdecoderlay:[23,25],transformerencod:23,transformerencoderlay:23,transforms1:12,transforms2:12,transit:47,translat:34,transorm:24,transpos:53,treat:[15,26],tree:13,trgat:26,tri:1,trt_export:[51,52],trt_m_output:[51,52],trt_model:[51,52],trt_pred_box:[51,52],trtexecutor:[51,52],truck:47,true_neg:26,truth:[13,21,26,47,53,54,56,57,58],tupl:[13,15,16,17,19,20,21,23,24,26,31,32,33,37,39,41,42],tutori:[3,50,51,52,53,54,56,57,58],tv:47,twice:37,two:[3,6,16,20,21,24,25,26,28,32,34,35,41,44,45,47,48,53,58],two_stag:19,two_stage_num_propos:19,txt:[13,48,50],type:[1,16,20,21,25,28,32,33,37,38,40,41,42,45,46,47,50],umbrella:47,un:26,uncom:[56,57,58],under:[4,56,57,58],union:[15,21,26,29,32,33,37,39,41,42],uniqu:[13,35],unit:35,univers:[2,5,6],unknown:13,unnam:[32,35,36,37,39,41,42,43],unnorm:[20,25,28],unpad:[19,42],unsign:35,unsqueez:47,until:[1,7],up:[26,29,47,50],updat:[16,24,26,44],update_data_object:16,update_loss_map:26,upload:[25,28],upward:13,url:48,us:[0,4,7,11,12,13,15,16,17,19,20,21,24,25,26,28,29,30,31,33,37,40,41,42,45,46,47,48,49,50,51,52,53,54,56,57,58],usag:[0,18,22,27,30,40,49],use_amp:29,useful:[21,26,37,42],user:[1,3,11,24,26,29,50,53],usuabl:23,usual:[1,3,15,35,36,44,45,47,53],util:[1,2,3,4,5,6,7,8,9,10,11,13,24,26,29,50],v2:25,v3:25,val2017:[3,24,45,50],val:[1,2,5,8,13,15,16,24,26,29,45,50],val_ann:[24,53],val_dataload:[15,24,45,53,54,56,57,58],val_dataset:24,val_fold:[24,53],val_fram:[15,24,29],val_nam:15,val_transform:[24,45],valid2017:58,valid:[3,13,15,24,26,29,44,45,50,51,52,53,54],valid_ratio:19,validation_sequ:[8,44],validation_step:[26,29],valu:[1,13,15,20,21,24,25,26,28,29,31,32,34,35,36,37,39,41,42,46,47,50,53,54,56,57],valueerror:[25,28],var_nam:31,variabl:[31,57],variou:17,vase:47,vb_folder:1,ve:46,vector:[34,36],vehicl:13,velodyn:13,veri:[17,44],verif:50,verifi:3,version:[4,17,47,48],vertic:[31,33,40,41],vertices_proj:40,vflip:31,via:[15,57],videocaptur:53,view:[32,33,35,36,38,39,40,41,42,43,44,46],visibl:44,vision:[45,48,49,50,53,54],visual:[30,33,49,51,52],volum:33,vs:21,w:[13,19,20,25,28,31,33,35,36,37,39,40,41,42,43,44,46,47,50,51,52,53],w_crop:31,wa:[3,13,50,51,52,53],wai:[3,31,50,57,58],wait:[51,52],wandb:15,want:[15,20,25,28,29,42,44,51,52,57],warn:[3,46,47,54,58],waymo:[0,37,40,44,49],waymo_data_dir:13,waymo_dataset:[13,40],waymo_fram:44,waymo_prepar:13,waymodataset:[13,37,40,44],wb:25,we:[1,13,17,20,21,25,26,35,37,40,45,46,47,50,51,52,53,54,56,57,58],wear:53,websit:48,weight:[16,20,21,25,26,28,29,53,54,55,56,57,58],weight_path:[51,52],well:[15,16,21,26,29,40,46,50],were:[3,47,50],what:[46,47,51,52,53,54,56,57,58],when:[15,20,21,24,25,26,28,29,31,35,36,39,42],where:[23,26,32,37,39,42,51,52,53],whether:[26,42],which:[3,13,23,25,28,33,35,41,51,52,54],whit:25,whole:[1,46,51,52],width:[12,19,20,25,28,32,37,41,42,50],wil:[32,42],window:[15,35,36],wine:47,wip:15,wish:40,with_box_refin:20,with_pos_emb:[19,23],within:[17,19,23,43,45,46],without:[3,23,26,29,32,35,36,37,39,41,42,43,44,47,50,51,52,53],wll:38,word:[25,42],work:[1,11,15,17,20,33,40,43,50],worker:[1,7,24],workflow:[51,52],world:[34,49],worri:47,wors:[33,41],would:[47,50,53],wrapper:[19,23],write:[1,3,11],wrong:3,x0:[32,34],x1:[32,47],x2:[32,47],x:[13,15,17,20,23,25,28,29,31,32,33,34,35,36,37,38,39,40,41,42,43],x_center:[32,50],xc:[33,37,47],xcyc:[32,37,47,50],xy:42,xyxi:[32,33,47],xz:33,y0:[32,34],y1:[32,47],y2:[32,47],y:[13,15,31,33,40,41,42],y_center:[32,50],yc:[33,37,47],yet:[16,25,26,29,42,47],yield:[7,48],you:[17,20,24,25,28,31,35,36,37,40,42,44,46,47,48,51,52,53],your:[3,20,25,31,37,42,47,48,49,51,52,53,54],your_hom:13,your_path:13,your_project_nam:[51,52],your_run_id:[51,52],yourself:40,yx:42,yxyx:32,z:[13,33,40],zc:[33,37],zebra:47,zero:[35,36,37],zip:50,zone:19},titles:["Alodataset","Base dataset","ChairsSDHom","Coco detection","Crowd human","FlyingChairs2","Flyingthings3D Subset","Merging datasets","MOT17","Sintel Disparity","Sintel Optical Flow","Sintel Multi (Flow + Disparity)","Transformations","Waymo Dataset","Alonet","Callbacks","Metrics","Transformer","Deformable DETR","Architecture","Models","Training","Detr","Architecture","Detr: Data connectors","Models","Training","Detr panoptic","Models","Training","Aloscene","Augmented Tensor","BoundingBoxes2D","BoundingBoxes3D","Camera calibration","Disparity","Flow","Frame","Label","Mask","Visualization","OrientedBoxes2D","Points2D","Spatial Augmented Tensor","Alodataset: Loading your vision datasets","Alonet : Training your models","Aloscene: Computer vision with ease","About augmented tensors","Getting started","Aloception\u2019s documentation","How to setup your data?","DETR","Deformable DETR","Finetuning Deformanble DETR","Finetuning DETR","Exporting DETR / Deformable-DETR to TensorRT","Training Deformable DETR","Training Detr","Training Panoptic Head module"],titleterms:{"1":[50,53,54,56,57,58],"2":[50,53,54,56,57,58],"2d":50,"3":[50,53,56,57,58],"3d":40,"class":3,"default":3,"export":55,about:[47,48],alocept:[48,49],alodataset:[0,44],alonet:[14,45],aloscen:[30,46],also:[15,24],ap:[15,16],api:[2,3,5,6,7,9,10,13,18,22,27,28,33,35,36,39,41,49],architectur:[19,23],argument:57,attach:47,augment:[30,31,43,47],backbon:[19,23],base:[1,15,20,25,50],basedataset:50,basic:[3,7,13,20,25,28,33,35,36,39,41],boundingboxes2d:32,boundingboxes3d:33,box:[16,50],calibr:34,callback:[15,21,26,29],camera:[34,53],chairssdhom:2,chairssdhomdataset:2,chang:3,checkpoint:[51,52],coco:[3,50],cocodetection2detr:24,cocopanoptic2detr:[24,58],comparison:56,comput:46,config:50,connector:24,content:55,convent:35,creat:[35,36],criterion:[21,26],crowd:4,custom:50,data2detr:24,data:[24,45,50],databas:3,dataset:[1,3,7,13,44,45,50],declar:50,deform:[18,20,52,53,55,56],deformanbl:53,detect:3,detector:15,detr50:[53,54],detr:[18,20,22,24,25,27,51,52,53,54,55,56,57],detrr50:28,dimens:46,dispar:[9,11,35],document:49,download:3,eas:46,easi:[35,36],encod:17,exampl:[16,24],expect:16,explor:46,featur:40,file:50,filter:3,finetun:[20,25,28,53,54],flow:[10,11,36],flyingchairs2:5,flyingthings3d:6,flyingthings3dsubsetdataset:6,frame:[37,46,50],from:[51,52],get:48,gete:49,giou:40,goal:50,head:[28,58],hello:46,how:50,human:4,implement:50,infer:[51,52,53,54,56,57,58],instal:48,iou:40,json:50,label:[38,47,50],level:57,litdetr:57,litpanopticdetr:58,load:[44,51,52],loader:44,make:[50,53,54,56,57,58],mask:39,matcher:[21,26],merg:7,mergedataset:7,metric:[15,16],minim:3,mlp:17,model:[20,25,28,45,56],modul:[45,50,58],mot17:8,multi:11,next:50,normal:46,object:[15,35,36],obtain:16,optic:10,option:[48,53],orientedboxes2d:41,other:[40,48],packag:30,panopt:[27,28,58],perform:56,pipelin:45,plai:46,points2d:42,posit:17,pq:[15,16],predict:53,prepar:[13,50],process:[57,58],pth:[51,52],r50:[20,25],refer:30,refin:20,result:16,run_id:[51,52],s:49,sampl:[3,44],see:[15,24],set:50,setup:50,sign:35,sintel:[9,10,11],sinteldisparitydataset:9,sintelflowdataset:10,spatial:43,start:[48,49],stateful:47,stream:44,subset:6,summari:16,tensor:[30,31,43,47],tensorrt:[51,52,55],train:[21,26,29,44,45,53,54,56,57,58],trainer:56,transform:[12,17,19,23,35,36,40,44,47],turori:49,us:[3,35,36,39,44],usag:[7,13,20,25,28,33,41],vision:[44,46],visual:[35,36,40],waymo:13,weight:[51,52],what:50,world:46,your:[44,45,46,50]}})
